Reduce categorical values

This function lets the user reduce categorical values in a vector. It is tidyverse friendly for use on pipelines

categ_reducer(
  df,
  ...,
  nmin = 0,
  pmin = 0,
  pcummax = 100,
  top = NA,
  other_label = "other"
)

Arguments

df
Categorical Vector

...
Variables. Which variable do you wish to reduce?

nmin
Integer. Number of minimum times a value is repeated

pmin
Numerical. Porcentage of minimum times a value is repeated

pcummax
Numerical. Top cumulative porcentage of most repeated values

top
Integer. Keep the n most frequently repeated values

other_label
Character. With which text do you wish to replace the filtered values with?

See also

Examples

data(dft) # Titanic dataset
categ_reducer(dft, Embarked, top = 2) %>% freqs(Embarked)

#> # A tibble: 3 x 5
#>   Embarked     n     p order  pcum
#>   <chr>    <int> <dbl> <int> <dbl>
#> 1 S          644 72.3      1  72.3
#> 2 C          168 18.9      2  91.1
#> 3 other       79  8.87     3 100. 

categ_reducer(dft, Ticket, nmin = 7, other_label = "Other Ticket") %>% freqs(Ticket)

#> # A tibble: 4 x 5
#>   Ticket           n     p order  pcum
#>   <chr>        <int> <dbl> <int> <dbl>
#> 1 Other Ticket   870 97.6      1  97.6
#> 2 1601             7  0.79     2  98.4
#> 3 347082           7  0.79     3  99.2
#> 4 CA. 2343         7  0.79     4 100.