Confidence Intervals on Dataframe

Calculate confidence intervals for a continuous numerical column on a dataframe, given a confidence level. You may also group results using another variable. Tidyverse friendly.

ci_var(df, var, group_var = NULL, conf = 0.95)

Arguments

df

Dataframe

var

Variable name. Must be a numerical column.

group_var

Variable name. Group results by another variable.

conf

Numeric. Confidence level (0-1).

Value

data.frame mean, standard deviation, counter, upper and lower CIs.

See also

Examples

data(dft) # Titanic dataset
ci_var(dft, Fare)
#> # A tibble: 1 × 5
#>   Fare_mean Fare_sd     n lower_ci upper_ci
#>       <dbl>   <dbl> <int>    <dbl>    <dbl>
#> 1      32.2    49.7   891     28.9     35.5
ci_var(dft, Fare, Pclass)
#> # A tibble: 3 × 6
#>   Pclass Fare_mean Fare_sd     n lower_ci upper_ci
#>   <fct>      <dbl>   <dbl> <int>    <dbl>    <dbl>
#> 1 1           84.2    78.4   216     73.6     94.7
#> 2 2           20.7    13.4   184     18.7     22.6
#> 3 3           13.7    11.8   491     12.6     14.7
ci_var(dft, Fare, Pclass, conf = 0.99)
#> # A tibble: 3 × 6
#>   Pclass Fare_mean Fare_sd     n lower_ci upper_ci
#>   <fct>      <dbl>   <dbl> <int>    <dbl>    <dbl>
#> 1 1           84.2    78.4   216     70.3     98.0
#> 2 2           20.7    13.4   184     18.1     23.2
#> 3 3           13.7    11.8   491     12.3     15.1