Clean title names of a data.frame/tibble object

Resulting names are unique and consist only of the _ character, numbers, and ASCII letters. Capitalization preferences can be specified using the lower parameter. Inspired by janitor::clean_names.

cleanNames(df, num = "x", ...)

Arguments

df

data.frame/tibble

num

Add character before only-numeric names

...

Additional parameters passed to cleanText

Value

Character vector with transformed strings.

See also

Examples

df <- dft[1:5, 1:6] # Dummy data
colnames(df) <- c("ID.", "34", "x_2", "Num 123", "Nòn-äscì", "  white   Spaces  ")
print(df)
#>   ID.    34 x_2 Num 123 Nòn-äscì   white   Spaces  
#> 1   1 FALSE   3    male       22                  1
#> 2   2  TRUE   1  female       38                  1
#> 3   3  TRUE   3  female       26                  0
#> 4   4  TRUE   1  female       35                  1
#> 5   5 FALSE   3    male       35                  0
cleanNames(df)
#>   id   x34 x2 num_123 nonasci white_spaces
#> 1  1 FALSE  3    male      22            1
#> 2  2  TRUE  1  female      38            1
#> 3  3  TRUE  3  female      26            0
#> 4  4  TRUE  1  female      35            1
#> 5  5 FALSE  3    male      35            0
cleanNames(df, lower = FALSE)
#>   ID   x34 x2 Num_123 Nonasci white_Spaces
#> 1  1 FALSE  3    male      22            1
#> 2  2  TRUE  1  female      38            1
#> 3  3  TRUE  3  female      26            0
#> 4  4  TRUE  1  female      35            1
#> 5  5 FALSE  3    male      35            0