One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-12-20       2020          12        20         51           Sun
#> 2  2020-07-08       2020           7         8         28           Wed
#> 3  2021-01-05       2021           1         5          1           Tue
#> 4  2021-04-26       2021           4        26         17           Mon
#> 5  2020-10-20       2020          10        20         42           Tue
#> 6  2021-04-19       2021           4        19         16           Mon
#> 7  2021-02-08       2021           2         8          6           Mon
#> 8  2021-02-23       2021           2        23          8           Tue
#> 9  2021-06-24       2021           6        24         25           Thu
#> 10 2020-11-24       2020          11        24         47           Tue
#>    dates_weekend dates_year_day               times times_year times_month
#> 1           TRUE            354 2021-04-16 03:34:21       2021           4
#> 2          FALSE            189 2021-03-13 09:26:42       2021           3
#> 3          FALSE              4 2021-05-11 23:31:48       2021           5
#> 4          FALSE            115 2021-05-05 17:39:00       2021           5
#> 5          FALSE            293 2021-03-31 10:22:53       2021           3
#> 6          FALSE            108 2021-05-30 04:09:52       2021           5
#> 7          FALSE             38 2021-03-04 08:23:07       2021           3
#> 8          FALSE             53 2021-04-22 13:57:40       2021           4
#> 9          FALSE            174 2021-06-05 21:25:37       2021           6
#> 10         FALSE            328 2021-06-25 11:29:22       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         16         16           Fri         FALSE            105          3
#> 2         13         11           Sat          TRUE             71          9
#> 3         11         19           Tue         FALSE            130         23
#> 4          5         18           Wed         FALSE            124         17
#> 5         31         13           Wed         FALSE             89         10
#> 6         30         22           Sun          TRUE            149          4
#> 7          4          9           Thu         FALSE             62          8
#> 8         22         16           Thu         FALSE            111         13
#> 9          5         23           Sat          TRUE            155         21
#> 10        25         26           Fri         FALSE            175         11
#>    times_minute times_minutes times_second
#> 1            34           214           21
#> 2            26           566           42
#> 3            31          1411           48
#> 4            39          1059            0
#> 5            22           622           53
#> 6             9           249           52
#> 7            23           503            7
#> 8            57           837           40
#> 9            25          1285           37
#> 10           29           689           22

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 5  2020     6  29   26     Mon   FALSE      180
#> 26 2020     7  20   29     Mon   FALSE      201
#> 44 2020     8   7   32     Fri   FALSE      219
#> 52 2020     8  15   33     Sat    TRUE      227
#> 54 2020     8  17   33     Mon   FALSE      229
#> 87 2020     9  19   38     Sat    TRUE      262
#>                            holidayname      holidaytype holiday_national
#> 5  Feast of Saint Peter and Saint Paul National holiday             TRUE
#> 26                    Independence Day National holiday             TRUE
#> 44                Battle of Boyacá Day National holiday             TRUE
#> 52                  Assumption of Mary National holiday             TRUE
#> 54      Day off for Assumption of Mary National holiday             TRUE
#> 87                     Valentine's Day       Observance            FALSE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 5               FALSE        FALSE           FALSE          FALSE
#> 26              FALSE        FALSE           FALSE          FALSE
#> 44              FALSE        FALSE           FALSE          FALSE
#> 52              FALSE        FALSE           FALSE          FALSE
#> 54              FALSE        FALSE           FALSE          FALSE
#> 87               TRUE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 5           FALSE
#> 26          FALSE
#> 44          FALSE
#> 52          FALSE
#> 54          FALSE
#> 87          FALSE

# }