Facebook Ads API

This returns all available FB ads for any account, campaign, or ad set id. For more information on Ad' API, go to the original documentaion

fb_ads(
  token,
  which,
  start = Sys.Date() - 31,
  end = Sys.Date(),
  fields = NA,
  api_version = "v8.0",
  process = TRUE
)

Arguments

token
Character. This must be a valid access token with sufficient privileges. Visit the Facebook API Graph Explorer to acquire one.

which
Character vector. This is the accounts, campaigns, adsets, or ads IDs to be queried. Remember: if report_level = "account", you must start the ID with act_.

start
Character. The first full day to report, in the format "YYYY-MM-DD"

end
Character. The last full day to report, in the format "YYYY-MM-DD"

fields
Character, json format. Leave NA for default fields.

api_version
Character. Facebook API version

process
Boolean. Process GET results to a more friendly format?

Details

This function was based on FBinsightsR.

See also

Examples

if (FALSE) {
token <- YOURTOKEN
which <- act_ADACCOUNT

# Query all ads for "which" with results in the last 10 days
ads <- fb_accounts(YOURTOKEN, which, start = Sys.Date() - 10) 
}