Facebook API Report Status Check

This returns all available FB insights per day including any given breakdown to the specified report level, and place into a data frame. For more information on Ad Insights' API, go to the original documentaion.

Usage

fb_report_check(
  token,
  report_run_id,
  api_version = NULL,
  live = FALSE,
  sleep = 10,
  quiet = FALSE
)

Arguments

token

Character. Valid access token with sufficient privileges. Visit the Facebook API Graph Explorer to acquire one.

report_run_id

Integer. Report ID to check status.

api_version

Character. Facebook API version.

live

Boolean. Run until status report is finished?

sleep

Boolean. If live=TRUE, then how many seconds should we wait until next check?

quiet

Boolean. Quiet messages?

Value

List with API status results.

See also

Examples

if (FALSE) { # \dontrun{
token <- "YOURTOKEN"
report_run_id <- "123456789"
fb_report_check(token, report_run_id, live = TRUE, quiet = FALSE)
} # }