Facebook's Long-Life User API Token

Using a 1-hour generic user token you can generate a 60 day token. You will need to have an App ID and App secret, and a valid token. Generate a new valid User Token with the API Graph.

Usage

fb_token(app_id, app_secret, token, api_version = NULL)

Arguments

app_id, app_secret

Character. Application ID and Secret.

token

Character. User token, created with API Graph or with this same fb_token()'s token.

api_version

Character. Facebook API version.

Value

Character. String with token requested. If successful, it'll contain an attribute called "expiration" with date and time of expiration.

Details

More info: Long-Lived Access Tokens

See also