Google Drive Files (API v4)

Authenticate and find Google Drive files and IDs by name.

Usage

filesGD(title, server = FALSE, json = NULL, api_key = NULL, email = NULL)

Arguments

title

Character. Title of Google Drive file. Uses regular expressions so you may fetch with patterns instead of names.

server

Boolean. Force interacting auth process?

json

Character. JSON filename with service auth

email, api_key

Character. If you have multiple pre-authorized accounts in your machine, you may non-interactively select which one you wish to use by email and/or api_key.

Value

Vector with found file names based on title on Google Drive.

See also