Google Trends: Related and Time Plots

This function creates a plot with Google Trend's related topics and queries, and let the user compare different keywords.

This function creates a plot with google trend's data on timelines and let the user compare different keywords.

Usage

gtrends_related(gtrend, top = NA, title = NA, note = NA, exclude = NULL)

gtrends_time(gtrend, title = NA)

Arguments

gtrend

List. Result from gtrendsR::gtrends(keyword, geo, time)

top

Integer. Filter top n results only.

title

Character. Custom title for the plot.

note

Character. Add a note to the plot if needed.

exclude

Character vector. Which observations do you wish to exclude?

Value

plot for Google Trend's results input gtrend.

Plot for Google Trend's results input gtrend.

See also