Automated H2O's AutoML

This function lets the user create a robust and fast model, using H2O's AutoML function. The result is a list with the best model, its parameters, datasets, performance metrics, variables importances, and plots. If the input is categorical, classification models will be trained and if is a continuous variable, regression models will be trained.

h2o_automl(
  df,
  y = "tag",
  ignore = c(),
  train_test = NA,
  split = 0.7,
  weight = NULL,
  target = "auto",
  balance = FALSE,
  impute = FALSE,
  no_outliers = TRUE,
  center = FALSE,
  scale = FALSE,
  seed = 0,
  nfolds = 5,
  thresh = 5,
  max_models = 3,
  max_time = 10 * 60,
  start_clean = TRUE,
  exclude_algos = c("StackedEnsemble", "DeepLearning"),
  plots = TRUE,
  alarm = TRUE,
  quiet = FALSE,
  save = FALSE,
  subdir = NA,
  project = "ML Project"
)

Arguments

df
Dataframe. Dataframe containing all your data, including the independent variable labeled as 'tag'. If you want to define which variable should be used instead, use the y parameter.

y
Variable or Character. Name of the independent variable.

ignore
Character vector. Force columns for the model to ignore

train_test
Character. If needed, df's column name with 'test' and 'train' values to split

split
Numeric. Value between 0 and 1 to split as train/test datasets. Value is for training set. Set value to 1 to train will all available data and test with same data (cross-validation will still be used when training)

weight
Column with observation weights. Giving some observation a weight of zero is equivalent to excluding it from the dataset; giving an observation a relative weight of 2 is equivalent to repeating that row twice. Negative weights are not allowed.

target
Value. Which is your target positive value? If set to 'auto', the target with largest mean(score) will be selected. Change the value to overwrite. Only used when binary categorical model.

balance
Boolean. Auto-balance train dataset with under-sampling?

impute
Boolean. Fill NA values with MICE?

no_outliers
Boolean/Numeric. Remove y's outliers in the dataset? Will remove those values that are farther than n standard deviations from the independent variable's mean (Z-score). Set to `TRUE` for default (3) or numeric to set a different multiplier.

center, scale
Boolean. Using the base function scale, do you wish to center and/or scale all numerical values?

seed
Integer. Set a seed for reproducibility. AutoML can only guarantee reproducibility if max_models is used because max_time is resource limited.

nfolds
Integer. Number of folds for k-fold cross-validation of the models. If set to 0, the test data will be used as validation, and cross-validation amd Stacked Ensembles disableded

thresh
Integer. Threshold for selecting binary or regression models: this number is the threshold of unique values we should have in 'tag' (more than: regression; less than: classification)

max_models, max_time
Numeric. Max number of models and seconds you wish for the function to iterate. Note that max_models guarantees reproducibility and max_time not (because it depends entirely on your machine's computational characteristics)

start_clean
Boolean. Erase everything in the current h2o instance before we start to train models?

exclude_algos
Vector of character strings. Algorithms to skip during the model-building phase. Set NULL to use all

plots
Boolean. Create plots objects?

alarm
Boolean. Ping an alarm when ready! Needs beepr installed

quiet
Boolean. Quiet messages, warnings, recommendations?

save
Boolean. Do you wish to save/export results into your working directory?

subdir
Character. In which directory do you wish to save the results? Working directory as default.

project
Character. Your project's name

List of algorithms

"DRF" (Distributed Random Forest, including Random Forest (RF) and Extremely-Randomized Trees (XRT)), "GLM" (Generalized Linear Model), "XGBoost" (eXtreme Grading Boosting), "GBM" (Gradient Boosting Machine), "DeepLearning" (Fully-connected multi-layer artificial neural network) and "StackedEnsemble". Read more here.

See also

Examples

# \donttest{
data(dft) # Titanic dataset
dft <- subset(dft, select = -c(Ticket, PassengerId, Cabin))

# Classification: Binomial - 2 Classes
r <- h2o_automl(dft, y = Survived, max_models = 1, impute = FALSE, target = "TRUE")

#> 2020-08-13 00:45:28 | Started process...

#> Variable to predict: Survived

#> - NOTE: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - NOTE: There are 4 non-numerical features. Consider using ohse() prior for One Hot Smart Encoding your categorical variables.

#> Model type: Classifier

#>     tag   n     p order   pcum
#> 1 FALSE 549 61.62     1  61.62
#> 2  TRUE 342 38.38     2 100.00

#> >>> Splitting datasets

#> train_size  test_size 
#>        623        268 

#> Algorithms excluded: 'StackedEnsemble', 'DeepLearning'

#> >>> Iterating until 1 models or 600 seconds...

#> Succesfully generated 1 models:

#>                           model_id       auc   logloss
#> 1 XGBoost_1_AutoML_20200813_004540 0.8946749 0.3931082
#> 
#> [1 row x 3 columns] 

#> Check results in H2O Flow's nice interface: http://localhost:54321/flow/index.html

#> Model selected: XGBoost_1_AutoML_20200813_004540

#> NOTE: The following variables were NOT important: SibSp, Pclass.2, Embarked.C

#> >>> Running predictions for Survived...

#> Target value: TRUE

#> >>> Generating plots...

#> Process duration: 21.6s

#>       AUC     ACC     PRC     TPR      TNR
#> 1 0.89467 0.16418 0.18325 0.33981 0.054545

print(r)

#> Leader Model: XGBoost_1_AutoML_20200813_004540
#> Independent Variable: Survived
#> Type: Classifier
#> Algorithm: XGBOOST
#> Trained: 1 models
#> Split: 70% training data (of 891 observations)
#> Seed: 0
#> 
#> Metrics: 
#>    AUC = 0.89467
#>    ACC = 0.16418
#>    PRC = 0.18325
#>    TPR = 0.33981
#>    TNR = 0.054545
#> 
#> Most important variables:
#>    Sex.female (29.7%)
#>    Fare (25.9%)
#>    Age (20.1%)
#>    Pclass.3 (8.1%)
#>    Sex.male (4.3%)

# Let's see all the stuff we have inside:
lapply(r, names)

#> $model
#> NULL
#> 
#> $y
#> NULL
#> 
#> $scores_test
#> [1] "tag"    "scores"
#> 
#> $metrics
#> [1] "dictionary"       "confusion_matrix" "gain_lift"        "metrics"         
#> [5] "plots"            "max_metrics"     
#> 
#> $importance
#> [1] "variable"            "relative_importance" "scaled_importance"  
#> [4] "importance"         
#> 
#> $datasets
#> [1] "global" "test"  
#> 
#> $scoring_history
#> [1] "timestamp"                     "duration"                     
#> [3] "number_of_trees"               "training_rmse"                
#> [5] "training_logloss"              "training_auc"                 
#> [7] "training_pr_auc"               "training_lift"                
#> [9] "training_classification_error"
#> 
#> $parameters
#>  [1] "model_id"                          "training_frame"                   
#>  [3] "nfolds"                            "keep_cross_validation_models"     
#>  [5] "keep_cross_validation_predictions" "fold_assignment"                  
#>  [7] "stopping_tolerance"                "seed"                             
#>  [9] "distribution"                      "ntrees"                           
#> [11] "max_depth"                         "min_rows"                         
#> [13] "min_child_weight"                  "sample_rate"                      
#> [15] "subsample"                         "col_sample_rate"                  
#> [17] "colsample_bylevel"                 "col_sample_rate_per_tree"         
#> [19] "colsample_bytree"                  "score_tree_interval"              
#> [21] "x"                                 "y"                                
#> 
#> $categoricals
#> [1] "Pclass"     "Sex"        "Embarked"   "train_test" "predict"   
#> 
#> $type
#> NULL
#> 
#> $split
#> NULL
#> 
#> $threshold
#> NULL
#> 
#> $model_name
#> NULL
#> 
#> $algorithm
#> NULL
#> 
#> $leaderboard
#> [1] "model_id"             "auc"                  "logloss"             
#> [4] "aucpr"                "mean_per_class_error" "rmse"                
#> [7] "mse"                 
#> 
#> $project
#> NULL
#> 
#> $seed
#> NULL
#> 
#> $h2o
#> NULL
#> 
#> $plots
#> [1] "dashboard"   "importance"  "ROC"         "conf_matrix" "response"   
#> [6] "gains"      
#> 

# Classification: Multi-Categorical - 3 Classes
r <- h2o_automl(dft, Pclass, ignore = c("Fare", "Cabin"), max_time = 30, plots = FALSE)

#> 2020-08-13 00:45:50 | Started process...

#> Variable to predict: Pclass

#> - NOTE: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - NOTE: There are 4 non-numerical features. Consider using ohse() prior for One Hot Smart Encoding your categorical variables.

#> Model type: Classifier

#>   tag   n     p order   pcum
#> 1 n_3 491 55.11     1  55.11
#> 2 n_1 216 24.24     2  79.35
#> 3 n_2 184 20.65     3 100.00

#> >>> Splitting datasets

#> train_size  test_size 
#>        623        268 

#> Ignored variables for training models: 'Fare', 'Cabin'

#> Algorithms excluded: 'StackedEnsemble', 'DeepLearning'

#> >>> Iterating until 3 models or 30 seconds...

#> Succesfully generated 3 models:

#>                           model_id mean_per_class_error   logloss
#> 1 XGBoost_2_AutoML_20200813_004551            0.4413151 0.7750237
#> 2 XGBoost_3_AutoML_20200813_004551            0.4434945 0.8010866
#> 3 XGBoost_1_AutoML_20200813_004551            0.4458870 0.7725101
#> 
#> [3 rows x 3 columns] 

#> Check results in H2O Flow's nice interface: http://localhost:54321/flow/index.html

#> Model selected: XGBoost_2_AutoML_20200813_004551

#> >>> Running predictions for Pclass...

#> Process duration: 11.8s

#>       AUC     ACC
#> 1 0.78358 0.66791

print(r)

#> Leader Model: XGBoost_2_AutoML_20200813_004551
#> Independent Variable: Pclass
#> Type: Classifier
#> Algorithm: XGBOOST
#> Trained: 3 models
#> Split: 70% training data (of 891 observations)
#> Seed: 0
#> 
#> Metrics: 
#>    AUC = 0.78358
#>    ACC = 0.66791
#> 
#> Most important variables:
#>    Age (51.7%)
#>    Survived.FALSE (12.3%)
#>    Embarked.C (7.6%)
#>    Sex.female (5.6%)
#>    Survived.TRUE (5.3%)

# Regression: Continuous Values
r <- h2o_automl(dft, y = "Fare", ignore = c("Pclass"), exclude_algos = NULL)

#> 2020-08-13 00:46:02 | Started process...

#> Variable to predict: Fare

#> - NOTE: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - NOTE: 2.2% (20) of Fare values are considered outliers (Z-Score: >3sd). Removing them from the dataset for better results. To keep them, set the 'no_outliers' parameter.

#> - NOTE: There are 4 non-numerical features. Consider using ohse() prior for One Hot Smart Encoding your categorical variables.

#> Model type: Regression

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   0.000   7.896  13.863  26.530  30.000 164.867 

#> >>> Splitting datasets

#> train_size  test_size 
#>        609        262 

#> Ignored variables for training models: 'Pclass'

#> >>> Iterating until 3 models or 600 seconds...

#> Succesfully generated 4 models:

#>                                           model_id mean_residual_deviance
#> 1 StackedEnsemble_AllModels_AutoML_20200813_004603               412.4477
#> 2                 XGBoost_2_AutoML_20200813_004603               445.0766
#> 3                 XGBoost_1_AutoML_20200813_004603               461.0355
#> 4                 XGBoost_3_AutoML_20200813_004603               519.5304
#>       rmse
#> 1 20.30881
#> 2 21.09684
#> 3 21.47174
#> 4 22.79321
#> 
#> [4 rows x 3 columns] 

#> Check results in H2O Flow's nice interface: http://localhost:54321/flow/index.html

#> Model selected: StackedEnsemble_AllModels_AutoML_20200813_004603

#> NOTE: No importance features for Stacked Ensemble Models

#> >>> Running predictions for Fare...

#> >>> Generating plots...

#> Process duration: 12.5s

#>       rmse      mae       mape      mse    rsq   rsqa
#> 1 20.30881 14.24359 0.07303959 412.4477 0.3169 0.3143

print(r)

#> Leader Model: StackedEnsemble_AllModels_AutoML_20200813_004603
#> Independent Variable: Fare
#> Type: Regression
#> Algorithm: STACKEDENSEMBLE
#> Trained: 4 models
#> Split: 70% training data (of 871 observations)
#> Seed: 0
#> 
#> Metrics: 
#>    rmse = 20.309
#>    mae = 14.244
#>    mape = 0.07304
#>    mse = 412.45
#>    rsq = 0.3169
#>    rsqa = 0.3143
#> 
#> 

# WITH PRE-DEFINED TRAIN/TEST DATAFRAMES
splits <- msplit(dft, size = 0.8)

#> train_size  test_size 
#>        712        179 

splits$train$split <- "train"
splits$test$split <- "test"
df <- rbind(splits$train, splits$test)
r <- h2o_automl(df, "Survived", max_models = 1, train_test = "split")

#> 2020-08-13 00:46:14 | Started process...

#> Variable to predict: Survived

#> - NOTE: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - NOTE: There are 5 non-numerical features. Consider using ohse() prior for One Hot Smart Encoding your categorical variables.

#> Model type: Classifier

#>     tag   n     p order   pcum
#> 1 FALSE 549 61.62     1  61.62
#> 2  TRUE 342 38.38     2 100.00
#> 
#>  test train 
#>   179   712 

#> Ignored variables for training models: 'split'

#> Algorithms excluded: 'StackedEnsemble', 'DeepLearning'

#> >>> Iterating until 1 models or 600 seconds...

#> Succesfully generated 1 models:

#>                           model_id       auc   logloss
#> 1 XGBoost_1_AutoML_20200813_004615 0.8838028 0.4011806
#> 
#> [1 row x 3 columns] 

#> Check results in H2O Flow's nice interface: http://localhost:54321/flow/index.html

#> Model selected: XGBoost_1_AutoML_20200813_004615

#> NOTE: The following variables were NOT important: Embarked.S, Embarked.Q

#> >>> Running predictions for Survived...

#> Target value: FALSE

#> >>> Generating plots...

#> Process duration: 8.63s

#>      AUC     ACC     PRC     TPR     TNR
#> 1 0.8838 0.86034 0.89655 0.73239 0.94444

print(r)

#> Leader Model: XGBoost_1_AutoML_20200813_004615
#> Independent Variable: Survived
#> Type: Classifier
#> Algorithm: XGBOOST
#> Trained: 1 models
#> Split: 80% training data (of 891 observations)
#> Seed: 0
#> 
#> Metrics: 
#>    AUC = 0.8838
#>    ACC = 0.86034
#>    PRC = 0.89655
#>    TPR = 0.73239
#>    TNR = 0.94444
#> 
#> Most important variables:
#>    Sex.female (26.5%)
#>    Fare (23.7%)
#>    Age (18.2%)
#>    Pclass.3 (7.8%)
#>    Sex.male (7.3%)

# }