Iterate Seeds on AutoML

This functions lets the user iterate and search for best seed. Note that if the results change a lot, you are having a high variance in your data.

Usage

iter_seeds(df, y, tries = 10, ...)

Arguments

df

Dataframe. Dataframe containing all your data, including the dependent variable labeled as 'tag'. If you want to define which variable should be used instead, use the y parameter.

y

Variable or Character. Name of the dependent variable or response.

tries

Integer. Number of iterations

...

Additional arguments passed to h2o_automl

Value

data.frame with performance results by seed tried on every row.

See also