Get My Personal LinkedIn Data

This function brings a list with your personal LinkedIn data

Usage

li_profile(token = NA)

Arguments

token

Object. OAuth Authentication: li_auth()'s output.

Value

List. Results of your own profile data given the token.

See also