List files in a directory

This function lets the user list all files on a given directory. It also lets filter files which contains a string.

listfiles(folder = getwd(), recursive = TRUE, regex = NA, images = FALSE)

Arguments

folder
Character. Directory which contains files

recursive
Boolean. Should the listing recurse into directories?

regex
Character. String to use for filtering files

images
Boolean. Bring only image files?

See also

Examples

# All files in current directory (without recursive files)
df <- listfiles(recursive = TRUE)
head(df, 3)

#>            filename size isdir mode               mtime               ctime
#> 1     autoline.html    7 FALSE  644 2020-06-29 21:48:13 2020-06-29 21:48:13
#> 2 balance_data.html    9 FALSE  644 2020-06-29 21:48:13 2020-06-29 21:48:13
#> 3    bindfiles.html    6 FALSE  644 2020-06-29 21:48:13 2020-06-29 21:48:13
#>                 atime uid gid  uname grname
#> 1 2020-06-29 21:48:21 501  20 runner  staff
#> 2 2020-06-29 21:48:21 501  20 runner  staff
#> 3 2020-06-29 21:48:22 501  20 runner  staff

# All files in current directory (with recursive files)
df <- listfiles(recursive = TRUE)
tail(df, 3)

#>           filename size isdir mode               mtime               ctime
#> 275 year_week.html    6 FALSE  644 2020-06-29 21:48:13 2020-06-29 21:48:13
#> 276    yt2mp3.html    6 FALSE  644 2020-06-29 21:48:13 2020-06-29 21:48:13
#> 277   zerovar.html    7 FALSE  644 2020-06-29 21:48:13 2020-06-29 21:48:13
#>                   atime uid gid  uname grname
#> 275 2020-06-29 21:48:13 501  20 runner  staff
#> 276 2020-06-29 21:48:13 501  20 runner  staff
#> 277 2020-06-29 21:48:13 501  20 runner  staff

# Check R files using regex
df <- listfiles(regex = "\\.R$")