Model Metrics and Performance Plots

This function generates plots of the metrics of a predictive model. This is an auxiliary function used in model_metrics() when the parameter plot is set to TRUE.

mplot_metrics(
  results,
  subtitle = NA,
  model_name = NA,
  save = FALSE,
  subdir = NA,
  file_name = "viz_metrics.png"
)

Arguments

results
Object. Results object from h2o_automl function

subtitle
Character. Subtitle to show in plot

model_name
Character. Model's name

save
Boolean. Save output plot into working directory

subdir
Character. Sub directory on which you wish to save the plot

file_name
Character. File name as you wish to save the plot

Value

Plot with results performance.

See also