Abbreviate numbers

This function converts a numeric vector's values into their abbreviated character equivalent, i.e. 100,000,000 into 100M.

num_abbr(x, n = 3)

Arguments

x
Numeric vector

n
Integer. Single numeric value, specifying number of significant figures to show. Range 1 to 6.

Value

A vector of character values that contain converted values

Examples

num_abbr(rnorm(100) * 1e6)

#>   [1] "-1.4M"  "255K"   "-2.44M" "-5.57K" "622K"   "1.15M"  "-1.82M" "-247K" 
#>   [9] "-244K"  "-283K"  "-554K"  "629K"   "2.07M"  "-1.63M" "512K"   "-1.86M"
#>  [17] "-522K"  "-52.6K" "543K"   "-914K"  "468K"   "363K"   "-1.3M"  "738K"  
#>  [25] "1.89M"  "-97.4K" "-936K"  "-16K"   "-827K"  "-1.51M" "935K"   "176K"  
#>  [33] "244K"   "1.62M"  "112K"   "-134K"  "-1.91M" "-279K"  "-313K"  "1.07M" 
#>  [41] "70K"    "-639K"  "-50K"   "-251K"  "445K"   "2.76M"  "46.5K"  "578K"  
#>  [49] "118K"   "-1.91M" "862K"   "-243K"  "-206K"  "19.2K"  "29.6K"  "550K"  
#>  [57] "-2.27M" "2.68M"  "-361K"  "213K"   "1.07M"  "-665K"  "1.11M"  "-246K" 
#>  [65] "-1.18M" "-976K"  "1.07M"  "132K"   "489K"   "-1.7M"  "-1.47M" "284K"  
#>  [73] "1.34M"  "237K"   "1.32M"  "524K"   "607K"   "-110K"  "172K"   "-90.3K"
#>  [81] "1.92M"  "1.3M"   "749K"   "556K"   "-548K"  "1.11M"  "-2.61M" "-156K" 
#>  [89] "434K"   "-382K"  "424K"   "1.06M"  "1.05M"  "-38.1K" "486K"   "1.67M" 
#>  [97] "-354K"  "946K"   "1.32M"  "-297K" 

num_abbr(rnorm(100) * 1e6, n = 1)

#>   [1] "-400K" "-800K" "-1M"   "-800K" "-2M"   "-700K" "-600K" "-500K" "200K" 
#>  [10] "1M"    "-200K" "-1M"   "300K"  "-400K" "600K"  "2M"    "400K"  "-2M"  
#>  [19] "200K"  "1M"    "2M"    "400K"  "1M"    "400K"  "100K"  "20K"   "600K" 
#>  [28] "-300K" "-500K" "200K"  "-1M"   "800K"  "80K"   "-1M"   "-900K" "500K" 
#>  [37] "-900K" "-300K" "900K"  "70K"   "-200K" "-800K" "2M"    "800K"  "1M"   
#>  [46] "1M"    "-1M"   "500K"  "-2M"   "2M"    "400K"  "100K"  "-1M"   "-600K"
#>  [55] "1M"    "700K"  "40K"   "-400K" "800K"  "800K"  "-2M"   "900K"  "-300K"
#>  [64] "300K"  "-2M"   "-800K" "-1M"   "400K"  "-300K" "500K"  "-100K" "-1M"  
#>  [73] "1M"    "1M"    "800K"  "1M"    "-500K" "-1M"   "-300K" "2M"    "-1M"  
#>  [82] "-1M"   "-2M"   "2M"    "600K"  "2M"    "-200K" "3M"    "1M"    "1M"   
#>  [91] "1M"    "-400K" "-800K" "-600K" "2M"    "100K"  "400K"  "1M"    "1M"   
#> [100] "600K"