One Hot Encoding for a Vector with Comma Separated Values

This function lets the user do one hot encoding on a variable with comma separated values

ohe_commas(df, variables, sep = ",", noval = "NoVal")

Arguments

df
Dataframe. May contain one or more columns with comma separated values which will be separated as one hot encoding

variables
Character. Which variables should split into new columns?

sep
Character. Which regular expression separates the elements?

noval
Character. No value text

See also

Examples

df <- data.frame(id = c(1:5),
                 x = c("AA, D", "AA,B", "B,  D", "A,D,B", NA),
                 z = c("AA+BB+AA", "AA", "BB,  AA", NA, "BB+AA"))
ohe_commas(df, "x")

#> # A tibble: 5 x 8
#>      id x     z        x_AA  x_D   x_B   x_A   x_NoVal
#>   <int> <chr> <chr>    <lgl> <lgl> <lgl> <lgl> <lgl>  
#> 1     1 AA, D AA+BB+AA TRUE  TRUE  FALSE FALSE FALSE  
#> 2     2 AA,B  AA       TRUE  FALSE TRUE  FALSE FALSE  
#> 3     3 B,  D BB,  AA  FALSE TRUE  TRUE  FALSE FALSE  
#> 4     4 A,D,B NA       FALSE TRUE  TRUE  TRUE  FALSE  
#> 5     5 NA    BB+AA    FALSE FALSE FALSE FALSE TRUE   

ohe_commas(df, c("x", "z"), sep = "\\+|,")

#> # A tibble: 5 x 11
#>      id x     z        x_AA  x_D   x_B   x_A   x_NoVal z_AA  z_BB  z_NoVal
#>   <int> <chr> <chr>    <lgl> <lgl> <lgl> <lgl> <lgl>   <lgl> <lgl> <lgl>  
#> 1     1 AA, D AA+BB+AA TRUE  TRUE  FALSE FALSE FALSE   TRUE  TRUE  FALSE  
#> 2     2 AA,B  AA       TRUE  FALSE TRUE  FALSE FALSE   TRUE  FALSE FALSE  
#> 3     3 B,  D BB,  AA  FALSE TRUE  TRUE  FALSE FALSE   TRUE  TRUE  FALSE  
#> 4     4 A,D,B NA       FALSE TRUE  TRUE  TRUE  FALSE   FALSE FALSE TRUE   
#> 5     5 NA    BB+AA    FALSE FALSE FALSE FALSE TRUE    TRUE  TRUE  FALSE