One Hot Smart Encoding (Dummy Variables)

This function lets the user automatically transform a dataframe with categorical columns into numerical by one hot encoding technic.

ohse(
  df,
  redundant = FALSE,
  drops = TRUE,
  ignore = NA,
  dates = FALSE,
  holidays = FALSE,
  country = "Colombia",
  currency_pair = NA,
  trim = 0,
  limit = 10,
  variance = 0.9,
  other_label = "OTHER",
  sep = "_",
  quiet = FALSE
)

Arguments

df
Dataframe

redundant
Boolean. Should we keep redundat columns? i.e. If the column only has two different values, should we keep both new columns?

drops
Boolean. Drop automatically some useless features?

ignore
Vector or character. Which column should be ignored?

dates
Boolean. Do you want the function to create more features out of the date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

trim
Integer. Trim names until the nth character

limit
Integer. Limit one hot encoding to the n most frequent values of each column. Set to NA to ignore argument.

variance
Numeric. Drop columns with more than n variance. Range: 0-1. For example: if a variable contains 91 unique different values out of 100 observations, this column will be suppressed if value is set to 0.9

other_label
Character. With which text do you wish to replace the filtered values with?

sep
Character. Separator's string

quiet
Boolean. Quiet all messages and summaries?

See also

Examples

# Example with Dataset 1
data(dft)
dft <- dft[,c(2,3,5,9,11)]

ohse(dft, limit = 3) %>% head(3)

#> >>> One Hot Encoding applied to 3 variables: 'Pclass', 'Embarked', 'Survived'

#> # A tibble: 3 x 8
#>   Survived_TRUE   Age  Fare Pclass_1 Pclass_2 Embarked_C Embarked_OTHER
#>           <dbl> <dbl> <dbl>    <dbl>    <dbl>      <dbl>          <dbl>
#> 1             0    22  7.25        0        0          0              0
#> 2             1    38 71.3         1        0          1              0
#> 3             1    26  7.92        0        0          0              0
#> # … with 1 more variable: Embarked_Q <dbl>

ohse(dft, limit = 3, redundant = TRUE) %>% head(3)

#> >>> One Hot Encoding applied to 3 variables: 'Survived', 'Pclass', 'Embarked'

#> # A tibble: 3 x 11
#>     Age  Fare Survived_FALSE Survived_TRUE Pclass_1 Pclass_2 Pclass_3 Embarked_C
#>   <dbl> <dbl>          <dbl>         <dbl>    <dbl>    <dbl>    <dbl>      <dbl>
#> 1    22  7.25              1             0        0        0        1          0
#> 2    38 71.3               0             1        1        0        0          1
#> 3    26  7.92              0             1        0        0        1          0
#> # … with 3 more variables: Embarked_OTHER <dbl>, Embarked_Q <dbl>,
#> #   Embarked_S <dbl>

# Getting rid of columns with no (or too much) variance
dft$no_variance1 <- 0
dft$no_variance2 <- c("A", rep("B", nrow(dft) - 1))
dft$no_variance3 <- as.character(rnorm(nrow(dft)))
dft$no_variance4 <- c(rep("A", 20), rnorm(nrow(dft) - 20))
ohse(dft, limit = 3) %>% head(3)

#> >>> One Hot Encoding applied to 4 variables: 'Pclass', 'Embarked', 'Survived', 'no_variance2'

#> >>> Automatically dropped 3 columns with 0% or >=90% variance: 'no_variance1', 'no_variance3', 'no_variance4'

#> # A tibble: 3 x 9
#>   Survived_TRUE   Age  Fare no_variance2_B Pclass_1 Pclass_2 Embarked_C
#>           <dbl> <dbl> <dbl>          <dbl>    <dbl>    <dbl>      <dbl>
#> 1             0    22  7.25              0        0        0          0
#> 2             1    38 71.3               1        1        0          1
#> 3             1    26  7.92              1        0        0          0
#> # … with 2 more variables: Embarked_OTHER <dbl>, Embarked_Q <dbl>

ohse(dft, limit = 3, var = 1) %>% head(3)

#> >>> One Hot Encoding applied to 5 variables: 'Pclass', 'Embarked', 'no_variance4', 'Survived', 'no_variance2'

#> >>> Automatically dropped 2 columns with 0% or >=100% variance: 'no_variance1', 'no_variance3'

#> # A tibble: 3 x 12
#>   Survived_TRUE   Age  Fare no_variance2_B Pclass_1 Pclass_2 Embarked_C
#>           <dbl> <dbl> <dbl>          <dbl>    <dbl>    <dbl>      <dbl>
#> 1             0    22  7.25              0        0        0          0
#> 2             1    38 71.3               1        1        0          1
#> 3             1    26  7.92              1        0        0          0
#> # … with 5 more variables: Embarked_OTHER <dbl>, Embarked_Q <dbl>,
#> #   no_variance4_.0.000452797290706798 <dbl>,
#> #   no_variance4_.0.00350143698256707 <dbl>, no_variance4_A <dbl>

# Examples with Dataset 2
data(dfl)
dfl$novar <- "novar"
ohse(dfl, dates = TRUE) %>% head(3)

#> >>> Processing 1 date/time columns: 'opp_date'

#> >>> One Hot Encoding applied to 5 variables: 'kind', 'opp_date_weekday', 'type', 'issued', 'opp_date_weekend'

#> >>> Automatically dropped 1 columns with 0% or >=90% variance: 'novar'

#> # A tibble: 3 x 24
#>   opp_id opp_date            type_TRUE values promo issued_TRUE opp_date_year
#>    <int> <dttm>                  <dbl>  <int> <dbl>       <dbl>         <dbl>
#> 1 992349 2019-02-06 12:22:48         1 480100     0           1          2019
#> 2 999436 2019-02-11 20:50:40         0 421750 50000           1          2019
#> 3 991823 2019-02-05 21:47:31         1 480100 12100           0          2019
#> # … with 17 more variables: opp_date_month <dbl>, opp_date_day <int>,
#> #   opp_date_week <dbl>, opp_date_weekend_TRUE <dbl>, opp_date_year_day <int>,
#> #   opp_date_hour <int>, opp_date_minute <int>, opp_date_minutes <int>,
#> #   opp_date_second <dbl>, kind_NAs <dbl>, kind_red <dbl>,
#> #   opp_date_weekday_Fri <dbl>, opp_date_weekday_Mon <dbl>,
#> #   opp_date_weekday_Sat <dbl>, opp_date_weekday_Sun <dbl>,
#> #   opp_date_weekday_Thu <dbl>, opp_date_weekday_Tue <dbl>

ohse(dfl, dates = TRUE, ignore = "issued") %>% head(3)

#> >>> Processing 1 date/time columns: 'opp_date'

#> >>> Omitting transformations for 'issued'

#> >>> One Hot Encoding applied to 4 variables: 'kind', 'opp_date_weekday', 'type', 'opp_date_weekend'

#> >>> Automatically dropped 1 columns with 0% or >=90% variance: 'novar'

#> # A tibble: 3 x 23
#>   opp_id opp_date            values promo type_TRUE opp_date_year opp_date_month
#>    <int> <dttm>               <int> <dbl>     <dbl>         <dbl>          <dbl>
#> 1 992349 2019-02-06 12:22:48 480100     0         1          2019              2
#> 2 999436 2019-02-11 20:50:40 421750 50000         0          2019              2
#> 3 991823 2019-02-05 21:47:31 480100 12100         1          2019              2
#> # … with 16 more variables: opp_date_day <int>, opp_date_week <dbl>,
#> #   opp_date_weekend_TRUE <dbl>, opp_date_year_day <int>, opp_date_hour <int>,
#> #   opp_date_minute <int>, opp_date_minutes <int>, opp_date_second <dbl>,
#> #   kind_NAs <dbl>, kind_red <dbl>, opp_date_weekday_Fri <dbl>,
#> #   opp_date_weekday_Mon <dbl>, opp_date_weekday_Sat <dbl>,
#> #   opp_date_weekday_Sun <dbl>, opp_date_weekday_Thu <dbl>,
#> #   opp_date_weekday_Tue <dbl>

if (FALSE) {
# Include holidays
ohse(dfl, dates = FALSE, holidays = TRUE, country = "Venezuela") %>% head()
}