Outliers: Z-score method

Z-score, also called a standard score, of an observation is a distance from the population center measured in number of normalization units. The default choice for center is sample mean and for normalization unit is standard deviation. Values are considered outliers based on z-score if its absolute value of default z-score is higher then the threshold (popular choice is 3).

outlier_zscore(x, thresh = 3, mad = FALSE)

Arguments

x
Numeric. Distribution

thresh
Numeric. Z-Score threshold for n standard deviations.

mad
Boolean. Use median absolute deviation instead?

See also