Plot timeline as Gantt Plot

This function plots groups of observartions with timelines in a Gantt Plot way. Only works if start and end are date format values.

plot_timeline(
  event,
  start,
  end = start + 1,
  label = NA,
  group = NA,
  title = "Curriculum Vitae Timeline",
  subtitle = "Bernardo Lares",
  size = 7,
  colour = "orange",
  save = FALSE,
  subdir = NA
)

Arguments

event
Vector. Event, role, label, or row.

start
Vector. Start date.

end
Vector. End date. Only one day be default if not defined

label
Vector. Place, institution, or label.

group
Vector. Academic, Work, Extracurricular...

title
Character. Title for the plot

subtitle
Character. Subtitle for the plot

size
Numeric. Bars' width

colour
Character. Colour when not using type

save
Boolean. Save the output plot in our working directory

subdir
Character. Into which subdirectory do you wish to save the plot to?