Scrabble: Word Scores

Get score for any word or list of words. You may set manually depending on the rules and languages you are playing with. Check the examples for Spanish and English values when I played Words With Friends.

scrabble_score(words, scores)

Arguments

words
Character vector. Words to score

scores
Dataframe. Must contain two columns: "tiles" with every letter of the alphabet and "scores" for each letter's score.

Value

data.frame with word, scores, and length values for each word.

See also

Examples

# \donttest{
# For Spanish words (default)
es_scores <- scrabble_points("es")

#> >>> Auto-loaded points for 'es'

# Custom scores for each letter
cu_scores <- data.frame(
  tiles = tolower(LETTERS),
  scores = c(1, 1, 1, 1, 1, 1, 1, 5, 1, 1, 5, 2, 4, 2, 1, 4, 10, 1, 1, 1, 2, 5, 4, 8, 3, 10)
)

# Score values for each set of rules
words <- c("Bernardo", "Whiskey", "R is great")
scrabble_score(words, es_scores)

#>         word scores length
#> 1    Whiskey     22      7
#> 2 R is great      9     10
#> 3   Bernardo      8      8

scrabble_score(words, cu_scores)

#>         word scores length
#> 1    Whiskey     16      7
#> 2   Bernardo      8      8
#> 3 R is great      7     10

# }