Send Slack Message (Webhook)

This function send a Slack message using its Webhooks.

slackSend(text, title = "", pretext = "", hook = NA, creds = NA)

Arguments

text, title, pretext
Character. Content on you Slack message.

hook
Character. Web hook URL. Ths value will be overwritten by creds if correctly used.

creds
Character. Credential's dir (see get_creds()). Set hook URL into the "slack" list in your YML file. Will use first value.

Details

For more help, you can follow the Sending messages using Incoming Webhooks original documentarion.

See also

Examples

if (FALSE) {
slackSend(text = "This is a message", title = "TEST", pretext = Sys.info()["user"])
}