Spread list column into new columns

Spread an existing list column into new columns on a data.frame. Note that every element on every observation must have a name for the function to do its work. Original column will be automatically suppressed but you can set the replace argument to avoid it.

spread_list(df, col, str = NULL, replace = TRUE)

Arguments

df
Dataframe

col
Variable name.

str
Character. Start column names with. If set to NULL, original name of column will be used.

replace
Boolean. Replace original values (delete column)

Examples

df <- dplyr::starwars
# Un-named list columns
spread_list(df, films, replace = FALSE) %>%
  select(name, starts_with("films")) %>%
  head(8)

#> # A tibble: 8 x 9
#>   name        films   `films_A New Hop… `films_Attack of th… `films_Return of t…
#>   <chr>       <list>  <lgl>             <lgl>                <lgl>              
#> 1 Luke Skywa… <chr [… TRUE              FALSE                TRUE               
#> 2 C-3PO       <chr [… TRUE              TRUE                 TRUE               
#> 3 R2-D2       <chr [… TRUE              TRUE                 TRUE               
#> 4 Darth Vader <chr [… TRUE              FALSE                TRUE               
#> 5 Leia Organa <chr [… TRUE              FALSE                TRUE               
#> 6 Owen Lars   <chr [… TRUE              TRUE                 FALSE              
#> 7 Beru White… <chr [… TRUE              TRUE                 FALSE              
#> 8 R5-D4       <chr [… TRUE              FALSE                FALSE              
#> # … with 4 more variables: films_Revenge of the Sith <lgl>,
#> #   films_The Empire Strikes Back <lgl>, films_The Force Awakens <lgl>,
#> #   films_The Phantom Menace <lgl>

# Named (and un-nammed) list columns
df <- tibble(id = 1:3, platform = list(
  list("fb" = 1, "ig" = 2), 
  list("fb" = 3), 
  list()))

#> Error in tibble(id = 1:3, platform = list(list(fb = 1, ig = 2), list(fb = 3),     list())): could not find function "tibble"

spread_list(df, platform)

#> Warning: The variable provided is not a list variable

#> # A tibble: 87 x 14
#>    name    height  mass hair_color  skin_color eye_color birth_year sex   gender
#>    <chr>    <int> <dbl> <chr>       <chr>      <chr>          <dbl> <chr> <chr> 
#>  1 Luke S…    172    77 blond       fair       blue            19   male  mascu…
#>  2 C-3PO      167    75 NA          gold       yellow         112   none  mascu…
#>  3 R2-D2       96    32 NA          white, bl… red             33   none  mascu…
#>  4 Darth …    202   136 none        white      yellow          41.9 male  mascu…
#>  5 Leia O…    150    49 brown       light      brown           19   fema… femin…
#>  6 Owen L…    178   120 brown, grey light      blue            52   male  mascu…
#>  7 Beru W…    165    75 brown       light      blue            47   fema… femin…
#>  8 R5-D4       97    32 NA          white, red red             NA   none  mascu…
#>  9 Biggs …    183    84 black       light      brown           24   male  mascu…
#> 10 Obi-Wa…    182    77 auburn, wh… fair       blue-gray       57   male  mascu…
#> # … with 77 more rows, and 5 more variables: homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>