Progressive Status Bar (Loading)

This function lets the user view a progressbar for a 'for' loop.

statusbar(
  run = 1,
  max.run = 100,
  label = run,
  msg = "DONE",
  type = "equal",
  start_time = NA
)

Arguments

run
Iterator. for loop or an integer with the current loop number. Start with 1 preferibly

max.run
Number. Maximum number of loops

label
String. With additionaly information to be printed at the end of the line. The default is run.

msg
Character. Finish message

type
Character. Loading type style: equal, domino

start_time
POSIXct. Start time to consider. If NA, then when first iteration starts will be set as start time. Useful for when first iteration is showed as done but started a few seconds/minutes ago.

See also

Examples

for (i in 1:5) {
  statusbar(i, 5)
  Sys.sleep(0.5)
}

#>  00:00:00 [=========                                ] 20% | 1                                         00:00:00 [=================                        ] 40% | 2                                         00:00:01 [=========================                ] 60% | 3                                         00:00:01 [=================================        ] 80% | 4                                         00:00:02 [=========================================] 100% | DONE