Sum Calculations and Plot

This function lets the user group, sum, calculate percentages and cumulatives. It also plots results if needed. Tidyverse friendly.

summer(df, ..., fun = sum, which = NA, top = NA, plot = FALSE)

Arguments

df
Data.frame

...
Variables. Variables you wish to process. Order matters. If no variables are passed, the whole data.frame will be considered

fun
Function (sum by default)

which
Character vector. Select variables to sum. If not set, all numeric variables will be returned

top
Integer. Filter and plot the n largest n values by group

plot
Boolean. Do you want to see a plot? Three variables tops.

See also