lares Theme for ggplot2

Based on hrbrthemes' theme_ipsum and customized for lares' use.

theme_lares2(
  font = "Arial Narrow",
  size = 12.5,
  main_colour = "darkorange3",
  hard_colour = "black",
  soft_colour = "grey30",
  bg_colour = "white",
  panel_colour = "transparent",
  legend = "right",
  grid = TRUE,
  mg = 9,
  pal = 0,
  which = "fct"
)

Arguments

font, size
Character and numeric. Base font family and base size for texts

main_colour, hard_colour, soft_colour, bg_colour, panel_colour
Character. Main colours for your theme

legend
Character. Legend position: top, right, bottom, left

grid
Character (TRUE, FALSE, or a combination of X, x, Y, y)

mg
Numeric. External margins

pal
Integer. 1 for fill and colour palette, 2 for only colour palette, 3 for only fill palette, 4 for personal labels-colour palette. 0 or else for nothing.

which
Character. When pal = 3, select which colours should be added with the custom colours pelette: fill, colour, text (fct) - first letters

Why Arial Narrow?

First and foremost, Arial Narrow is generally installed by default or readily available on any modern system, so it's "free"-ish; plus, it is a condensed font with solid default kerning pairs and geometric numbers.

See also