Stopwatch to measure R Timings

Start a stopwatch.

Stop a stopwatch.

Reset all tic and toc values in your environment.

tic(id = 1, quiet = TRUE)

toc(id = 1, msg = "Elapsed time:", units = TRUE, signif = 3, quiet = FALSE)

resettictoc(which = "both")

Arguments

id
Define ID if multiple tic() & toc() are being used

quiet
Boolean. Quiet messages?

msg
Character. Custom message shown

units
Boolean. Do you want nice format for the time units? If not, seconds elapsed as numerical values

signif
Integer. Significant digits

which
Character. Select: both, tic, toc

Value

toc returns an (invisible) list containing the timestamps tic and toc, time in seconds and the message msg.

See also

Examples

# Basic use (global stopwatch)
tic()
Sys.sleep(0.1)
toc()

#> Elapsed time: 0.139s

# Multiple tic tocs
tic(id = "two", quiet = FALSE)

#> Tic `id = two` start time: 2020-07-01 18:07:56

Sys.sleep(0.2)
toc(id = "two")

#> Elapsed time: 0.207s

# Global is still working (id = 1)
toc(msg = "The function finished its work in")

#> The function finished its work in 0.355s