Trim MP3 Audio File

This function trims MP3 files given a start and/or end numeric timestamp. Requires ffmpeg installed in your machine.

trim_mp3(
  file,
  start_time = 0,
  end_time = NA,
  overwrite = FALSE,
  ext = "mp3",
  quiet = FALSE
)

Arguments

file
Character. File name to trim.

start_time
Numeric. Start and end time to trim the audio output in seconds.

end_time
Numeric. Start and end time to trim the audio output in seconds.

overwrite
Boolean. Overwrite original file?

ext
Character. File extension/type.

quiet
Boolean. Keep quiet? If not, print messages.

See also