Google Sheets Writing (API v4)

Write data into Google Sheets knowing the file's title. You may write a single value into a cell or a data.frame into a cell range.

writeGS(
  data,
  title,
  sheet = "Hoja 1",
  range = "A1",
  reformat = FALSE,
  append = FALSE,
  json = NULL,
  email = NULL,
  api_key = NULL,
  server = FALSE,
  ...
)

Arguments

data
Object (value, vector, dataframe, list)

title
Character. Title of Google Drive file. Uses regular expressions so you may fetch with patterns instead of names.

sheet
Character. Working sheet to import

range
Character. A cell range to read from

reformat
Boolean. Reformat the affected cells?

append
Boolean.

json
Character. JSON filename with service auth

email
Character. If you have multiple pre-authorized accounts in your machine, you may non-interactively select which one you wish to use by email and/or api_key.

api_key
Character. If you have multiple pre-authorized accounts in your machine, you may non-interactively select which one you wish to use by email and/or api_key.

server
Boolean. Force interacting auth process?

...
Further read_sheet parameters

See also