Confussion Matrix

This function calculates a Confussion Matrix using crosstab for 2 or more categories. You can either set the score and threshold or the labels you wish to cross with.

conf_mat(tag, score, thresh = 0.5, sense = ">=", plot = FALSE)

Arguments

tag
Vector. Real known label

score
Vector. Predicted value or model's result

thresh
Numeric. Value which splits the results for the confusion matrix when binary.

sense
Character. Inequation sense for threshhold: <, <=, >=, >

plot
Boolean. Plot result?

See also

Examples

# Binary example
binary <- data.frame(
  tag = c("N","N","N","N","N","Y","Y","Y","N","Y"),
  score = c(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0))
conf_mat(tag = binary$tag, score = binary$score)

#> # A tibble: 2 x 3
#>   `Real x Pred`     N     Y
#>   <fct>         <int> <int>
#> 1 N                 2     4
#> 2 Y                 4     0

conf_mat(tag = binary$tag, score = binary$score, sense = ">")

#> # A tibble: 2 x 3
#>   `Real x Pred`     N     Y
#>   <fct>         <int> <int>
#> 1 N                 1     5
#> 2 Y                 4     0

conf_mat(tag = binary$tag, score = binary$score, thresh = 0.2)

#> # A tibble: 2 x 3
#>   `Real x Pred`     N     Y
#>   <fct>         <int> <int>
#> 1 N                 5     1
#> 2 Y                 4     0