Correlation table

This function correlates a whole dataframe, running one hot smart encoding (ohse) to transform non-numerical features. Note that it will automatically suppress columns with less than 3 non missing values and warn the user.

corr(
  df,
  method = "pearson",
  pvalue = FALSE,
  dec = 6,
  ignore = NA,
  dummy = TRUE,
  logs = FALSE,
  limit = 10,
  top = NA,
  ...
)

Arguments

df
Dataframe. It doesn't matter if it's got non-numerical columns: they will be filtered!

method
Character. Any of: c("pearson", "kendall", "spearman")

pvalue
Boolean. Returns a list, with correlations and statistical significance (p-value) for each value

dec
Integer. Number of decimals to round correlations and p-values

ignore
Vector or character. Which column should be ignored?

dummy
Boolean. Should One Hot (Smart) Encoding (ohse) be applied to categorical columns?

logs
Boolean. Calculate log(x)+1 for numerical columns?

limit
Integer. Limit one hot encoding to the n most frequent values of each column. Set to NA to ignore argument.

top
Integer. Select top N most relevant variables? Filtered and sorted by mean of each variable's correlations

...
Additional parameters to pass to ohse

Value

data.frame. Squared dimensions (nxn) to match every correlation between every df data.frame column/variable. Notice that when using ohse() you may get more dimensions.

See also

Examples

data(dft) # Titanic dataset
df <- dft[,2:5]

corr(df)

#>               Survived_TRUE  Sex_male  Pclass_1  Pclass_2
#> Survived_TRUE      1.000000 -0.543351  0.285904  0.093349
#> Sex_male          -0.543351  1.000000 -0.098013 -0.064746
#> Pclass_1           0.285904 -0.098013  1.000000 -0.288585
#> Pclass_2           0.093349 -0.064746 -0.288585  1.000000

# Ignore specific column
corr(df, ignore = "Pclass")

#>               Survived_TRUE  Sex_male
#> Survived_TRUE      1.000000 -0.543351
#> Sex_male          -0.543351  1.000000

# Keep redundant combinations
corr(df, redundant = TRUE)

#>                Survived_FALSE Survived_TRUE  Pclass_1  Pclass_2  Pclass_3
#> Survived_FALSE       1.000000     -1.000000 -0.285904 -0.093349  0.322308
#> Survived_TRUE       -1.000000      1.000000  0.285904  0.093349 -0.322308
#> Pclass_1            -0.285904      0.285904  1.000000 -0.288585 -0.626738
#> Pclass_2            -0.093349      0.093349 -0.288585  1.000000 -0.565210
#> Pclass_3             0.322308     -0.322308 -0.626738 -0.565210  1.000000
#> Sex_female          -0.543351      0.543351  0.098013  0.064746 -0.137143
#> Sex_male             0.543351     -0.543351 -0.098013 -0.064746  0.137143
#>                Sex_female  Sex_male
#> Survived_FALSE  -0.543351  0.543351
#> Survived_TRUE    0.543351 -0.543351
#> Pclass_1         0.098013 -0.098013
#> Pclass_2         0.064746 -0.064746
#> Pclass_3        -0.137143  0.137143
#> Sex_female       1.000000 -1.000000
#> Sex_male        -1.000000  1.000000

#' # Calculate p-values as well
corr(df, pvalue = TRUE)

#> $cor
#>               Survived_TRUE  Sex_male  Pclass_1  Pclass_2
#> Survived_TRUE      1.000000 -0.543351  0.285904  0.093349
#> Sex_male          -0.543351  1.000000 -0.098013 -0.064746
#> Pclass_1           0.285904 -0.098013  1.000000 -0.288585
#> Pclass_2           0.093349 -0.064746 -0.288585  1.000000
#> 
#> $pvalue
#>               Survived_TRUE     Sex_male     Pclass_1     Pclass_2
#> Survived_TRUE  0.000000e+00 1.406066e-69 3.190582e-18 5.293655e-03
#> Sex_male       1.406066e-69 0.000000e+00 3.404901e-03 5.336354e-02
#> Pclass_1       3.190582e-18 3.404901e-03 0.000000e+00 1.498959e-18
#> Pclass_2       5.293655e-03 5.336354e-02 1.498959e-18 0.000000e+00
#> 

# Test when no more than 2 non-missing values
df$trash <- c(1, rep(NA, nrow(df)-1))
# and another method...
corr(df, method = "spearman")

#> Warning: Dropped columns with less than 3 non-missing values: 'trash'

#>               Survived_TRUE  Sex_male  Pclass_1  Pclass_2
#> Survived_TRUE      1.000000 -0.543351  0.285904  0.093349
#> Sex_male          -0.543351  1.000000 -0.098013 -0.064746
#> Pclass_1           0.285904 -0.098013  1.000000 -0.288585
#> Pclass_2           0.093349 -0.064746 -0.288585  1.000000