Correlation table

This function correlates a whole dataframe, running one hot smart encoding (ohse) to transform non-numerical features. Note that it will automatically suppress columns with less than 3 non missing values and warn the user.

corr(
  df,
  method = "pearson",
  pvalue = FALSE,
  dec = 6,
  ignore = NA,
  dummy = TRUE,
  limit = 10,
  dates = FALSE,
  redundant = FALSE,
  logs = FALSE,
  top = NA
)

Arguments

df
Dataframe. It doesn't matter if it's got non-numerical columns: they will be filtered!

method
Character. Any of: c("pearson", "kendall", "spearman")

pvalue
Boolean. Returns a list, with correlations and statistical significance (p-value) for each value

dec
Integer. Number of decimals to round correlations and p-values

ignore
Character vector. Which columns do you wish to exclude?

dummy
Boolean. Should One Hot Encoding be applied to categorical columns?

limit
Integer. Limit one hot encoding to the n most frequent values of each column. Set to NA to ignore argument.

dates
Boolean. Do you want the function to create more features out of the date/time columns?

redundant
Boolean. Should we keep redundat columns? i.e. It the column only has two different values, should we keep both new columns?

logs
Boolean. Automatically calculate log(values) for numerical variables (not binaries)

top
Integer. Select top N most relevant variables? Filtered and sorted by mean of each variable's correlations

See also

Examples

data(dft) # Titanic dataset
df <- dft[,2:5]

corr(df)

#>               Survived_TRUE  Sex_male       Age  Pclass_1  Pclass_2
#> Survived_TRUE      1.000000 -0.543351 -0.077221  0.285904  0.093349
#> Sex_male          -0.543351  1.000000  0.093254 -0.098013 -0.064746
#> Age               -0.077221  0.093254  1.000000  0.348941  0.006954
#> Pclass_1           0.285904 -0.098013  0.348941  1.000000 -0.288585
#> Pclass_2           0.093349 -0.064746  0.006954 -0.288585  1.000000

# Ignore specific column
corr(df, ignore = "Pclass")

#>               Survived_TRUE  Sex_male       Age
#> Survived_TRUE      1.000000 -0.543351 -0.077221
#> Sex_male          -0.543351  1.000000  0.093254
#> Age               -0.077221  0.093254  1.000000

# Keep redundant combinations
corr(df, redundant = TRUE)

#>                      Age Survived_FALSE Survived_TRUE  Pclass_1  Pclass_2
#> Age             1.000000       0.077221     -0.077221  0.348941  0.006954
#> Survived_FALSE  0.077221       1.000000     -1.000000 -0.285904 -0.093349
#> Survived_TRUE  -0.077221      -1.000000      1.000000  0.285904  0.093349
#> Pclass_1        0.348941      -0.285904      0.285904  1.000000 -0.288585
#> Pclass_2        0.006954      -0.093349      0.093349 -0.288585  1.000000
#> Pclass_3       -0.312271       0.322308     -0.322308 -0.626738 -0.565210
#> Sex_female     -0.093254      -0.543351      0.543351  0.098013  0.064746
#> Sex_male        0.093254       0.543351     -0.543351 -0.098013 -0.064746
#>                 Pclass_3 Sex_female  Sex_male
#> Age            -0.312271  -0.093254  0.093254
#> Survived_FALSE  0.322308  -0.543351  0.543351
#> Survived_TRUE  -0.322308   0.543351 -0.543351
#> Pclass_1       -0.626738   0.098013 -0.098013
#> Pclass_2       -0.565210   0.064746 -0.064746
#> Pclass_3        1.000000  -0.137143  0.137143
#> Sex_female     -0.137143   1.000000 -1.000000
#> Sex_male        0.137143  -1.000000  1.000000

#' # Calculate p-values as well
corr(df, pvalue = TRUE)

#> $cor
#>               Survived_TRUE  Sex_male       Age  Pclass_1  Pclass_2
#> Survived_TRUE      1.000000 -0.543351 -0.077221  0.285904  0.093349
#> Sex_male          -0.543351  1.000000  0.093254 -0.098013 -0.064746
#> Age               -0.077221  0.093254  1.000000  0.348941  0.006954
#> Pclass_1           0.285904 -0.098013  0.348941  1.000000 -0.288585
#> Pclass_2           0.093349 -0.064746  0.006954 -0.288585  1.000000
#> 
#> $pvalue
#>               Survived_TRUE Sex_male      Age Pclass_1 Pclass_2
#> Survived_TRUE      0.000000 0.000000 0.039125 0.000000 0.005294
#> Sex_male           0.000000 0.000000 0.012671 0.003405 0.053364
#> Age                0.039125 0.012671 0.000000 0.000000 0.852842
#> Pclass_1           0.000000 0.003405 0.000000 0.000000 0.000000
#> Pclass_2           0.005294 0.053364 0.852842 0.000000 0.000000
#> 

# Test when no more than 2 non-missing values
df$trash <- c(1, rep(NA, nrow(df)-1))
# and another method...
corr(df, method = "spearman")

#> Warning: Dropped columns with less than 3 non-missing values: 'trash'

#>               Survived_TRUE  Sex_male       Age  Pclass_1  Pclass_2
#> Survived_TRUE      1.000000 -0.543351 -0.052565  0.285904  0.093349
#> Sex_male          -0.543351  1.000000  0.083330 -0.098013 -0.064746
#> Age               -0.052565  0.083330  1.000000  0.333881  0.031291
#> Pclass_1           0.285904 -0.098013  0.333881  1.000000 -0.288585
#> Pclass_2           0.093349 -0.064746  0.031291 -0.288585  1.000000