One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-12-02       2020          12         2         49           Wed
#> 2  2020-06-20       2020           6        20         25           Sat
#> 3  2020-12-18       2020          12        18         51           Fri
#> 4  2021-04-08       2021           4         8         14           Thu
#> 5  2020-10-02       2020          10         2         40           Fri
#> 6  2021-04-01       2021           4         1         13           Thu
#> 7  2021-01-21       2021           1        21          3           Thu
#> 8  2021-02-05       2021           2         5          6           Fri
#> 9  2021-06-06       2021           6         6         23           Sun
#> 10 2020-11-06       2020          11         6         45           Fri
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            336 2021-03-29 00:33:43       2021           3
#> 2           TRUE            171 2021-02-23 06:26:04       2021           2
#> 3          FALSE            352 2021-04-23 20:31:10       2021           4
#> 4          FALSE             97 2021-04-17 14:38:22       2021           4
#> 5          FALSE            275 2021-03-13 07:22:15       2021           3
#> 6          FALSE             90 2021-05-12 01:09:14       2021           5
#> 7          FALSE             20 2021-02-14 05:22:29       2021           2
#> 8          FALSE             35 2021-04-04 10:57:02       2021           4
#> 9           TRUE            156 2021-05-18 18:24:59       2021           5
#> 10         FALSE            310 2021-06-07 08:28:44       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         29         13           Mon         FALSE             87          0
#> 2         23          8           Tue         FALSE             53          6
#> 3         23         17           Fri         FALSE            112         20
#> 4         17         16           Sat          TRUE            106         14
#> 5         13         11           Sat          TRUE             71          7
#> 6         12         19           Wed         FALSE            131          1
#> 7         14          7           Sun          TRUE             44          5
#> 8          4         14           Sun          TRUE             93         10
#> 9         18         20           Tue         FALSE            137         18
#> 10         7         23           Mon         FALSE            157          8
#>    times_minute times_minutes times_second
#> 1            33            33           43
#> 2            26           386            4
#> 3            31          1231           10
#> 4            38           878           22
#> 5            22           442           15
#> 6             9            69           14
#> 7            22           322           29
#> 8            57           657            2
#> 9            24          1104           59
#> 10           28           508           44

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 5  2020     6  11   24     Thu   FALSE      162
#> 9  2020     6  15   24     Mon   FALSE      166
#> 14 2020     6  20   25     Sat    TRUE      171
#> 15 2020     6  21   25     Sun    TRUE      172
#> 16 2020     6  22   25     Mon   FALSE      173
#> 23 2020     6  29   26     Mon   FALSE      180
#>                            holidayname      holidaytype holiday_national
#> 5                       Corpus Christi       Observance            FALSE
#> 9                       Corpus Christi National holiday             TRUE
#> 14                       June Solstice           Season            FALSE
#> 15                        Father's Day       Observance            FALSE
#> 16                        Sacred Heart National holiday             TRUE
#> 23 Feast of Saint Peter and Saint Paul National holiday             TRUE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 5                TRUE        FALSE           FALSE          FALSE
#> 9               FALSE        FALSE           FALSE          FALSE
#> 14              FALSE        FALSE           FALSE           TRUE
#> 15               TRUE        FALSE           FALSE          FALSE
#> 16              FALSE        FALSE           FALSE          FALSE
#> 23              FALSE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 5           FALSE
#> 9           FALSE
#> 14          FALSE
#> 15          FALSE
#> 16          FALSE
#> 23          FALSE

# }