One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-12-04       2020          12         4         49           Fri
#> 2  2020-06-22       2020           6        22         25           Mon
#> 3  2020-12-20       2020          12        20         51           Sun
#> 4  2021-04-10       2021           4        10         15           Sat
#> 5  2020-10-04       2020          10         4         40           Sun
#> 6  2021-04-03       2021           4         3         14           Sat
#> 7  2021-01-23       2021           1        23          4           Sat
#> 8  2021-02-07       2021           2         7          6           Sun
#> 9  2021-06-08       2021           6         8         23           Tue
#> 10 2020-11-08       2020          11         8         45           Sun
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            338 2021-03-30 11:33:27       2021           3
#> 2          FALSE            173 2021-02-24 17:25:48       2021           2
#> 3           TRUE            354 2021-04-25 07:30:54       2021           4
#> 4           TRUE             99 2021-04-19 01:38:06       2021           4
#> 5           TRUE            277 2021-03-14 18:21:59       2021           3
#> 6           TRUE             92 2021-05-13 12:08:58       2021           5
#> 7           TRUE             22 2021-02-15 16:22:13       2021           2
#> 8           TRUE             37 2021-04-05 21:56:46       2021           4
#> 9          FALSE            158 2021-05-20 05:24:43       2021           5
#> 10          TRUE            312 2021-06-08 19:28:28       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         30         13           Tue         FALSE             88         11
#> 2         24          8           Wed         FALSE             54         17
#> 3         25         17           Sun          TRUE            114          7
#> 4         19         16           Mon         FALSE            108          1
#> 5         14         11           Sun          TRUE             72         18
#> 6         13         19           Thu         FALSE            132         12
#> 7         15          7           Mon         FALSE             45         16
#> 8          5         14           Mon         FALSE             94         21
#> 9         20         20           Thu         FALSE            139          5
#> 10         8         23           Tue         FALSE            158         19
#>    times_minute times_minutes times_second
#> 1            33           693           27
#> 2            25          1045           48
#> 3            30           450           54
#> 4            38            98            6
#> 5            21          1101           59
#> 6             8           728           58
#> 7            22           982           13
#> 8            56          1316           46
#> 9            24           324           43
#> 10           28          1168           28

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 3  2020     6  11   24     Thu   FALSE      162
#> 7  2020     6  15   24     Mon   FALSE      166
#> 12 2020     6  20   25     Sat    TRUE      171
#> 13 2020     6  21   25     Sun    TRUE      172
#> 14 2020     6  22   25     Mon   FALSE      173
#> 21 2020     6  29   26     Mon   FALSE      180
#>                            holidayname      holidaytype holiday_national
#> 3                       Corpus Christi       Observance            FALSE
#> 7                       Corpus Christi National holiday             TRUE
#> 12                       June Solstice           Season            FALSE
#> 13                        Father's Day       Observance            FALSE
#> 14                        Sacred Heart National holiday             TRUE
#> 21 Feast of Saint Peter and Saint Paul National holiday             TRUE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 3                TRUE        FALSE           FALSE          FALSE
#> 7               FALSE        FALSE           FALSE          FALSE
#> 12              FALSE        FALSE           FALSE           TRUE
#> 13               TRUE        FALSE           FALSE          FALSE
#> 14              FALSE        FALSE           FALSE          FALSE
#> 21              FALSE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 3           FALSE
#> 7           FALSE
#> 12          FALSE
#> 13          FALSE
#> 14          FALSE
#> 21          FALSE

# }