One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-09-22       2020           9        22         38           Tue
#> 2  2020-07-21       2020           7        21         29           Tue
#> 3  2020-07-22       2020           7        22         30           Wed
#> 4  2021-01-03       2021           1         3          1           Sun
#> 5  2020-11-05       2020          11         5         45           Thu
#> 6  2021-04-21       2021           4        21         16           Wed
#> 7  2021-04-05       2021           4         5         14           Mon
#> 8  2020-09-06       2020           9         6         36           Sun
#> 9  2021-04-12       2021           4        12         15           Mon
#> 10 2021-03-16       2021           3        16         11           Tue
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            265 2021-03-04 05:59:57       2021           3
#> 2          FALSE            202 2021-02-04 10:57:43       2021           2
#> 3          FALSE            203 2021-01-23 07:16:10       2021           1
#> 4           TRUE              2 2021-01-07 17:23:07       2021           1
#> 5          FALSE            309 2021-02-06 02:00:44       2021           2
#> 6          FALSE            110 2021-03-06 03:28:22       2021           3
#> 7          FALSE             94 2021-01-20 01:43:30       2021           1
#> 8           TRUE            249 2021-03-22 02:58:28       2021           3
#> 9          FALSE            101 2021-02-15 06:13:45       2021           2
#> 10         FALSE             74 2021-04-15 11:18:43       2021           4
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          4          9           Thu         FALSE             62          5
#> 2          4          5           Thu         FALSE             34         10
#> 3         23          4           Sat          TRUE             22          7
#> 4          7          1           Thu         FALSE              6         17
#> 5          6          6           Sat          TRUE             36          2
#> 6          6         10           Sat          TRUE             64          3
#> 7         20          3           Wed         FALSE             19          1
#> 8         22         12           Mon         FALSE             80          2
#> 9         15          7           Mon         FALSE             45          6
#> 10        15         15           Thu         FALSE            104         11
#>    times_minute times_minutes times_second
#> 1            59           359           57
#> 2            57           657           43
#> 3            16           436           10
#> 4            23          1043            7
#> 5             0           120           44
#> 6            28           208           22
#> 7            43           103           30
#> 8            58           178           28
#> 9            13           373           45
#> 10           18           678           43

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}