One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50)
)

# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-12-24       2020          12        24         52           Thu
#> 2  2020-07-12       2020           7        12         28           Sun
#> 3  2021-01-09       2021           1         9          2           Sat
#> 4  2021-04-30       2021           4        30         18           Fri
#> 5  2020-10-24       2020          10        24         43           Sat
#> 6  2021-04-23       2021           4        23         17           Fri
#> 7  2021-02-12       2021           2        12          7           Fri
#> 8  2021-02-27       2021           2        27          9           Sat
#> 9  2021-06-28       2021           6        28         26           Mon
#> 10 2020-11-28       2020          11        28         48           Sat
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            358 2021-04-19 12:46:18       2021           4
#> 2           TRUE            193 2021-03-16 18:38:39       2021           3
#> 3           TRUE              8 2021-05-15 08:43:45       2021           5
#> 4          FALSE            119 2021-05-09 02:50:57       2021           5
#> 5           TRUE            297 2021-04-03 19:34:50       2021           4
#> 6          FALSE            112 2021-06-02 13:21:49       2021           6
#> 7          FALSE             42 2021-03-07 17:35:04       2021           3
#> 8           TRUE             57 2021-04-25 23:09:37       2021           4
#> 9          FALSE            178 2021-06-09 06:37:34       2021           6
#> 10          TRUE            332 2021-06-28 20:41:19       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         19         16           Mon         FALSE            108         12
#> 2         16         11           Tue         FALSE             74         18
#> 3         15         20           Sat          TRUE            134          8
#> 4          9         19           Sun          TRUE            128          2
#> 5          3         14           Sat          TRUE             92         19
#> 6          2         22           Wed         FALSE            152         13
#> 7          7         10           Sun          TRUE             65         17
#> 8         25         17           Sun          TRUE            114         23
#> 9          9         23           Wed         FALSE            159          6
#> 10        28         26           Mon         FALSE            178         20
#>    times_minute times_minutes times_second
#> 1            46           766           18
#> 2            38          1118           39
#> 3            43           523           45
#> 4            50           170           57
#> 5            34          1174           50
#> 6            21           801           49
#> 7            35          1055            4
#> 8             9          1389           37
#> 9            37           397           34
#> 10           41          1241           19

# Holidays given a date range and country
# \donttest{
hol <- date_feats(
  seq(Sys.Date() - 365, Sys.Date(), by = 1),
  keep_originals = TRUE,
  holidays = TRUE,
  country = "Colombia"
)

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

head(hol[!is.na(hol$holidayname), ])

#>    year month day week weekday weekend year_day
#> 1  2020     6  29   26     Mon   FALSE      180
#> 22 2020     7  20   29     Mon   FALSE      201
#> 40 2020     8   7   32     Fri   FALSE      219
#> 48 2020     8  15   33     Sat    TRUE      227
#> 50 2020     8  17   33     Mon   FALSE      229
#> 83 2020     9  19   38     Sat    TRUE      262
#>                            holidayname      holidaytype holiday_national
#> 1  Feast of Saint Peter and Saint Paul National holiday             TRUE
#> 22                    Independence Day National holiday             TRUE
#> 40                Battle of Boyacá Day National holiday             TRUE
#> 48                  Assumption of Mary National holiday             TRUE
#> 50      Day off for Assumption of Mary National holiday             TRUE
#> 83                     Valentine's Day       Observance            FALSE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 1               FALSE        FALSE           FALSE          FALSE
#> 22              FALSE        FALSE           FALSE          FALSE
#> 40              FALSE        FALSE           FALSE          FALSE
#> 48              FALSE        FALSE           FALSE          FALSE
#> 50              FALSE        FALSE           FALSE          FALSE
#> 83               TRUE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 1           FALSE
#> 22          FALSE
#> 40          FALSE
#> 48          FALSE
#> 50          FALSE
#> 83          FALSE

# }