One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-09-25       2020           9        25         39           Fri
#> 2  2020-07-24       2020           7        24         30           Fri
#> 3  2020-07-25       2020           7        25         30           Sat
#> 4  2021-01-06       2021           1         6          1           Wed
#> 5  2020-11-08       2020          11         8         45           Sun
#> 6  2021-04-24       2021           4        24         17           Sat
#> 7  2021-04-08       2021           4         8         14           Thu
#> 8  2020-09-09       2020           9         9         37           Wed
#> 9  2021-04-15       2021           4        15         15           Thu
#> 10 2021-03-19       2021           3        19         12           Fri
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            268 2021-03-07 06:16:38       2021           3
#> 2          FALSE            205 2021-02-07 11:14:24       2021           2
#> 3           TRUE            206 2021-01-26 07:32:51       2021           1
#> 4          FALSE              5 2021-01-10 17:39:48       2021           1
#> 5           TRUE            312 2021-02-09 02:17:25       2021           2
#> 6           TRUE            113 2021-03-09 03:45:03       2021           3
#> 7          FALSE             97 2021-01-23 02:00:11       2021           1
#> 8          FALSE            252 2021-03-25 03:15:09       2021           3
#> 9          FALSE            104 2021-02-18 06:30:26       2021           2
#> 10         FALSE             77 2021-04-18 11:35:24       2021           4
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          7         10           Sun          TRUE             65          6
#> 2          7          6           Sun          TRUE             37         11
#> 3         26          4           Tue         FALSE             25          7
#> 4         10          2           Sun          TRUE              9         17
#> 5          9          6           Tue         FALSE             39          2
#> 6          9         10           Tue         FALSE             67          3
#> 7         23          4           Sat          TRUE             22          2
#> 8         25         12           Thu         FALSE             83          3
#> 9         18          7           Thu         FALSE             48          6
#> 10        18         16           Sun          TRUE            107         11
#>    times_minute times_minutes times_second
#> 1            16           376           38
#> 2            14           674           24
#> 3            32           452           51
#> 4            39          1059           48
#> 5            17           137           25
#> 6            45           225            3
#> 7             0           120           11
#> 8            15           195            9
#> 9            30           390           26
#> 10           35           695           24

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}