One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-06-20       2020           6        20         25           Sat
#> 2  2020-04-18       2020           4        18         16           Sat
#> 3  2020-04-19       2020           4        19         16           Sun
#> 4  2020-10-01       2020          10         1         40           Thu
#> 5  2020-08-03       2020           8         3         31           Mon
#> 6  2021-01-17       2021           1        17          3           Sun
#> 7  2021-01-01       2021           1         1          1           Fri
#> 8  2020-06-04       2020           6         4         23           Thu
#> 9  2021-01-08       2021           1         8          2           Fri
#> 10 2020-12-12       2020          12        12         50           Sat
#>    dates_weekend dates_year_day               times times_year times_month
#> 1           TRUE            171 2020-11-30 03:15:59       2020          11
#> 2           TRUE            108 2020-11-02 08:13:45       2020          11
#> 3           TRUE            109 2020-10-21 04:32:12       2020          10
#> 4          FALSE            274 2020-10-05 14:39:09       2020          10
#> 5          FALSE            215 2020-11-03 23:16:46       2020          11
#> 6           TRUE             16 2020-12-02 00:44:24       2020          12
#> 7          FALSE              0 2020-10-17 22:59:32       2020          10
#> 8          FALSE            155 2020-12-18 00:14:30       2020          12
#> 9          FALSE              7 2020-11-13 03:29:47       2020          11
#> 10          TRUE            346 2021-01-11 08:34:45       2021           1
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         30         48           Mon         FALSE            334          3
#> 2          2         44           Mon         FALSE            306          8
#> 3         21         43           Wed         FALSE            294          4
#> 4          5         40           Mon         FALSE            278         14
#> 5          3         44           Tue         FALSE            307         23
#> 6          2         49           Wed         FALSE            336          0
#> 7         17         42           Sat          TRUE            290         22
#> 8         18         51           Fri         FALSE            352          0
#> 9         13         46           Fri         FALSE            317          3
#> 10        11          2           Mon         FALSE             10          8
#>    times_minute times_minutes times_second
#> 1            15           195           59
#> 2            13           493           45
#> 3            32           272           12
#> 4            39           879            9
#> 5            16          1396           46
#> 6            44            44           24
#> 7            59          1379           32
#> 8            14            14           30
#> 9            29           209           47
#> 10           34           514           45

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}