One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-03-24       2020           3        24         12           Tue
#> 2  2020-01-21       2020           1        21          3           Tue
#> 3  2020-01-22       2020           1        22          4           Wed
#> 4  2020-07-05       2020           7         5         27           Sun
#> 5  2020-05-07       2020           5         7         19           Thu
#> 6  2020-10-21       2020          10        21         43           Wed
#> 7  2020-10-05       2020          10         5         40           Mon
#> 8  2020-03-08       2020           3         8         10           Sun
#> 9  2020-10-12       2020          10        12         41           Mon
#> 10 2020-09-15       2020           9        15         37           Tue
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE             83 2020-09-03 06:15:16       2020           9
#> 2          FALSE             20 2020-08-06 11:13:02       2020           8
#> 3          FALSE             21 2020-07-25 07:31:29       2020           7
#> 4           TRUE            186 2020-07-09 17:38:26       2020           7
#> 5          FALSE            127 2020-08-08 02:16:03       2020           8
#> 6          FALSE            294 2020-09-05 03:43:41       2020           9
#> 7          FALSE            278 2020-07-22 01:58:49       2020           7
#> 8           TRUE             67 2020-09-21 03:13:47       2020           9
#> 9          FALSE            285 2020-08-17 06:29:04       2020           8
#> 10         FALSE            258 2020-10-15 11:34:02       2020          10
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          3         36           Thu         FALSE            246          6
#> 2          6         32           Thu         FALSE            218         11
#> 3         25         30           Sat          TRUE            206          7
#> 4          9         28           Thu         FALSE            190         17
#> 5          8         32           Sat          TRUE            220          2
#> 6          5         36           Sat          TRUE            248          3
#> 7         22         30           Wed         FALSE            203          1
#> 8         21         38           Mon         FALSE            264          3
#> 9         17         33           Mon         FALSE            229          6
#> 10        15         42           Thu         FALSE            288         11
#>    times_minute times_minutes times_second
#> 1            15           375           16
#> 2            13           673            2
#> 3            31           451           29
#> 4            38          1058           26
#> 5            16           136            3
#> 6            43           223           41
#> 7            58           118           49
#> 8            13           193           47
#> 9            29           389            4
#> 10           34           694            2

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}