One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-07-14       2020           7        14         28           Tue
#> 2  2020-05-12       2020           5        12         19           Tue
#> 3  2020-05-13       2020           5        13         20           Wed
#> 4  2020-10-25       2020          10        25         43           Sun
#> 5  2020-08-27       2020           8        27         35           Thu
#> 6  2021-02-10       2021           2        10          6           Wed
#> 7  2021-01-25       2021           1        25          4           Mon
#> 8  2020-06-28       2020           6        28         26           Sun
#> 9  2021-02-01       2021           2         1          5           Mon
#> 10 2021-01-05       2021           1         5          1           Tue
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            195 2020-12-24 06:19:29       2020          12
#> 2          FALSE            132 2020-11-26 11:17:15       2020          11
#> 3          FALSE            133 2020-11-14 07:35:42       2020          11
#> 4           TRUE            298 2020-10-29 17:42:39       2020          10
#> 5          FALSE            239 2020-11-28 02:20:16       2020          11
#> 6          FALSE             40 2020-12-26 03:47:54       2020          12
#> 7          FALSE             24 2020-11-11 02:03:02       2020          11
#> 8           TRUE            179 2021-01-11 03:18:00       2021           1
#> 9          FALSE             31 2020-12-07 06:33:17       2020          12
#> 10         FALSE              4 2021-02-04 11:38:15       2021           2
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         24         52           Thu         FALSE            358          6
#> 2         26         48           Thu         FALSE            330         11
#> 3         14         46           Sat          TRUE            318          7
#> 4         29         44           Thu         FALSE            302         17
#> 5         28         48           Sat          TRUE            332          2
#> 6         26         52           Sat          TRUE            360          3
#> 7         11         46           Wed         FALSE            315          2
#> 8         11          2           Mon         FALSE             10          3
#> 9          7         49           Mon         FALSE            341          6
#> 10         4          5           Thu         FALSE             34         11
#>    times_minute times_minutes times_second
#> 1            19           379           29
#> 2            17           677           15
#> 3            35           455           42
#> 4            42          1062           39
#> 5            20           140           16
#> 6            47           227           54
#> 7             3           123            2
#> 8            18           198            0
#> 9            33           393           17
#> 10           38           698           15

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}