One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-12-18       2020          12        18         51           Fri
#> 2  2020-07-06       2020           7         6         27           Mon
#> 3  2021-01-03       2021           1         3          1           Sun
#> 4  2021-04-24       2021           4        24         17           Sat
#> 5  2020-10-18       2020          10        18         42           Sun
#> 6  2021-04-17       2021           4        17         16           Sat
#> 7  2021-02-06       2021           2         6          6           Sat
#> 8  2021-02-21       2021           2        21          8           Sun
#> 9  2021-06-22       2021           6        22         25           Tue
#> 10 2020-11-22       2020          11        22         47           Sun
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            352 2021-04-14 10:00:02       2021           4
#> 2          FALSE            187 2021-03-11 15:52:23       2021           3
#> 3           TRUE              2 2021-05-10 05:57:29       2021           5
#> 4           TRUE            113 2021-05-04 00:04:41       2021           5
#> 5           TRUE            291 2021-03-29 16:48:34       2021           3
#> 6           TRUE            106 2021-05-28 10:35:33       2021           5
#> 7           TRUE             36 2021-03-02 14:48:48       2021           3
#> 8           TRUE             51 2021-04-20 20:23:21       2021           4
#> 9          FALSE            172 2021-06-04 03:51:18       2021           6
#> 10          TRUE            326 2021-06-23 17:55:03       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         14         15           Wed         FALSE            103         10
#> 2         11         10           Thu         FALSE             69         15
#> 3         10         19           Mon         FALSE            129          5
#> 4          4         18           Tue         FALSE            123          0
#> 5         29         13           Mon         FALSE             87         16
#> 6         28         22           Fri         FALSE            147         10
#> 7          2          9           Tue         FALSE             60         14
#> 8         20         16           Tue         FALSE            109         20
#> 9          4         23           Fri         FALSE            154          3
#> 10        23         25           Wed         FALSE            173         17
#>    times_minute times_minutes times_second
#> 1             0           600            2
#> 2            52           952           23
#> 3            57           357           29
#> 4             4             4           41
#> 5            48          1008           34
#> 6            35           635           33
#> 7            48           888           48
#> 8            23          1223           21
#> 9            51           231           18
#> 10           55          1075            3

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 7  2020     6  29   26     Mon   FALSE      180
#> 28 2020     7  20   29     Mon   FALSE      201
#> 46 2020     8   7   32     Fri   FALSE      219
#> 54 2020     8  15   33     Sat    TRUE      227
#> 56 2020     8  17   33     Mon   FALSE      229
#> 89 2020     9  19   38     Sat    TRUE      262
#>                            holidayname      holidaytype holiday_national
#> 7  Feast of Saint Peter and Saint Paul National holiday             TRUE
#> 28                    Independence Day National holiday             TRUE
#> 46                Battle of Boyacá Day National holiday             TRUE
#> 54                  Assumption of Mary National holiday             TRUE
#> 56      Day off for Assumption of Mary National holiday             TRUE
#> 89                     Valentine's Day       Observance            FALSE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 7               FALSE        FALSE           FALSE          FALSE
#> 28              FALSE        FALSE           FALSE          FALSE
#> 46              FALSE        FALSE           FALSE          FALSE
#> 54              FALSE        FALSE           FALSE          FALSE
#> 56              FALSE        FALSE           FALSE          FALSE
#> 89               TRUE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 7           FALSE
#> 28          FALSE
#> 46          FALSE
#> 54          FALSE
#> 56          FALSE
#> 89          FALSE

# }