One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-03-26       2020           3        26         13           Thu
#> 2  2020-01-23       2020           1        23          4           Thu
#> 3  2020-01-24       2020           1        24          4           Fri
#> 4  2020-07-07       2020           7         7         27           Tue
#> 5  2020-05-09       2020           5         9         19           Sat
#> 6  2020-10-23       2020          10        23         43           Fri
#> 7  2020-10-07       2020          10         7         41           Wed
#> 8  2020-03-10       2020           3        10         10           Tue
#> 9  2020-10-14       2020          10        14         42           Wed
#> 10 2020-09-17       2020           9        17         38           Thu
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE             85 2020-09-05 01:29:35       2020           9
#> 2          FALSE             22 2020-08-08 06:27:21       2020           8
#> 3          FALSE             23 2020-07-27 02:45:48       2020           7
#> 4          FALSE            188 2020-07-11 12:52:45       2020           7
#> 5           TRUE            129 2020-08-09 21:30:22       2020           8
#> 6          FALSE            296 2020-09-06 22:58:00       2020           9
#> 7          FALSE            280 2020-07-23 21:13:08       2020           7
#> 8          FALSE             69 2020-09-22 22:28:06       2020           9
#> 9          FALSE            287 2020-08-19 01:43:23       2020           8
#> 10         FALSE            260 2020-10-17 06:48:21       2020          10
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          5         36           Sat          TRUE            248          1
#> 2          8         32           Sat          TRUE            220          6
#> 3         27         30           Mon         FALSE            208          2
#> 4         11         28           Sat          TRUE            192         12
#> 5          9         32           Sun          TRUE            221         21
#> 6          6         36           Sun          TRUE            249         22
#> 7         23         30           Thu         FALSE            204         21
#> 8         22         38           Tue         FALSE            265         22
#> 9         19         34           Wed         FALSE            231          1
#> 10        17         42           Sat          TRUE            290          6
#>    times_minute times_minutes times_second
#> 1            29            89           35
#> 2            27           387           21
#> 3            45           165           48
#> 4            52           772           45
#> 5            30          1290           22
#> 6            58          1378            0
#> 7            13          1273            8
#> 8            28          1348            6
#> 9            43           103           23
#> 10           48           408           21

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}