One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-11-26       2020          11        26         48           Thu
#> 2  2020-06-14       2020           6        14         24           Sun
#> 3  2020-12-12       2020          12        12         50           Sat
#> 4  2021-04-02       2021           4         2         14           Fri
#> 5  2020-09-26       2020           9        26         39           Sat
#> 6  2021-03-26       2021           3        26         13           Fri
#> 7  2021-01-15       2021           1        15          3           Fri
#> 8  2021-01-30       2021           1        30          5           Sat
#> 9  2021-05-31       2021           5        31         22           Mon
#> 10 2020-10-31       2020          10        31         44           Sat
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            330 2021-03-23 06:25:09       2021           3
#> 2           TRUE            165 2021-02-17 12:17:30       2021           2
#> 3           TRUE            346 2021-04-18 02:22:36       2021           4
#> 4          FALSE             91 2021-04-11 20:29:48       2021           4
#> 5           TRUE            269 2021-03-07 13:13:41       2021           3
#> 6          FALSE             84 2021-05-06 07:00:40       2021           5
#> 7          FALSE             14 2021-02-08 11:13:55       2021           2
#> 8           TRUE             29 2021-03-29 16:48:28       2021           3
#> 9          FALSE            150 2021-05-13 00:16:25       2021           5
#> 10          TRUE            304 2021-06-01 14:20:10       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         23         12           Tue         FALSE             81          6
#> 2         17          7           Wed         FALSE             47         12
#> 3         18         16           Sun          TRUE            107          2
#> 4         11         15           Sun          TRUE            100         20
#> 5          7         10           Sun          TRUE             65         13
#> 6          6         18           Thu         FALSE            125          7
#> 7          8          6           Mon         FALSE             38         11
#> 8         29         13           Mon         FALSE             87         16
#> 9         13         19           Thu         FALSE            132          0
#> 10         1         22           Tue         FALSE            151         14
#>    times_minute times_minutes times_second
#> 1            25           385            9
#> 2            17           737           30
#> 3            22           142           36
#> 4            29          1229           48
#> 5            13           793           41
#> 6             0           420           40
#> 7            13           673           55
#> 8            48          1008           28
#> 9            16            16           25
#> 10           20           860           10

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 11 2020     6  11   24     Thu   FALSE      162
#> 15 2020     6  15   24     Mon   FALSE      166
#> 20 2020     6  20   25     Sat    TRUE      171
#> 21 2020     6  21   25     Sun    TRUE      172
#> 22 2020     6  22   25     Mon   FALSE      173
#> 29 2020     6  29   26     Mon   FALSE      180
#>                            holidayname      holidaytype holiday_national
#> 11                      Corpus Christi       Observance            FALSE
#> 15                      Corpus Christi National holiday             TRUE
#> 20                       June Solstice           Season            FALSE
#> 21                        Father's Day       Observance            FALSE
#> 22                        Sacred Heart National holiday             TRUE
#> 29 Feast of Saint Peter and Saint Paul National holiday             TRUE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 11               TRUE        FALSE           FALSE          FALSE
#> 15              FALSE        FALSE           FALSE          FALSE
#> 20              FALSE        FALSE           FALSE           TRUE
#> 21               TRUE        FALSE           FALSE          FALSE
#> 22              FALSE        FALSE           FALSE          FALSE
#> 29              FALSE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 11          FALSE
#> 15          FALSE
#> 20          FALSE
#> 21          FALSE
#> 22          FALSE
#> 29          FALSE

# }