One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-05-30       2020           5        30         22           Sat
#> 2  2020-03-28       2020           3        28         13           Sat
#> 3  2020-03-29       2020           3        29         13           Sun
#> 4  2020-09-10       2020           9        10         37           Thu
#> 5  2020-07-13       2020           7        13         28           Mon
#> 6  2020-12-27       2020          12        27         52           Sun
#> 7  2020-12-11       2020          12        11         50           Fri
#> 8  2020-05-14       2020           5        14         20           Thu
#> 9  2020-12-18       2020          12        18         51           Fri
#> 10 2020-11-21       2020          11        21         47           Sat
#>    dates_weekend dates_year_day               times times_year times_month
#> 1           TRUE            150 2020-11-09 04:08:28       2020          11
#> 2           TRUE             87 2020-10-12 09:06:14       2020          10
#> 3           TRUE             88 2020-09-30 05:24:41       2020           9
#> 4          FALSE            253 2020-09-14 15:31:38       2020           9
#> 5          FALSE            194 2020-10-14 00:09:15       2020          10
#> 6           TRUE            361 2020-11-11 01:36:53       2020          11
#> 7          FALSE            345 2020-09-26 23:52:01       2020           9
#> 8          FALSE            134 2020-11-27 01:06:59       2020          11
#> 9          FALSE            352 2020-10-23 04:22:16       2020          10
#> 10          TRUE            325 2020-12-21 09:27:14       2020          12
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          9         45           Mon         FALSE            313          4
#> 2         12         41           Mon         FALSE            285          9
#> 3         30         40           Wed         FALSE            273          5
#> 4         14         37           Mon         FALSE            257         15
#> 5         14         42           Wed         FALSE            287          0
#> 6         11         46           Wed         FALSE            315          1
#> 7         26         39           Sat          TRUE            269         23
#> 8         27         48           Fri         FALSE            331          1
#> 9         23         43           Fri         FALSE            296          4
#> 10        21         51           Mon         FALSE            355          9
#>    times_minute times_minutes times_second
#> 1             8           248           28
#> 2             6           546           14
#> 3            24           324           41
#> 4            31           931           38
#> 5             9             9           15
#> 6            36            96           53
#> 7            52          1432            1
#> 8             6            66           59
#> 9            22           262           16
#> 10           27           567           14

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}