One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-12-17       2020          12        17         51           Thu
#> 2  2020-07-05       2020           7         5         27           Sun
#> 3  2021-01-02       2021           1         2          1           Sat
#> 4  2021-04-23       2021           4        23         17           Fri
#> 5  2020-10-17       2020          10        17         42           Sat
#> 6  2021-04-16       2021           4        16         16           Fri
#> 7  2021-02-05       2021           2         5          6           Fri
#> 8  2021-02-20       2021           2        20          8           Sat
#> 9  2021-06-21       2021           6        21         25           Mon
#> 10 2020-11-21       2020          11        21         47           Sat
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            351 2021-04-13 05:29:13       2021           4
#> 2           TRUE            186 2021-03-10 11:21:34       2021           3
#> 3           TRUE              1 2021-05-09 01:26:40       2021           5
#> 4          FALSE            112 2021-05-02 19:33:52       2021           5
#> 5           TRUE            290 2021-03-28 12:17:45       2021           3
#> 6          FALSE            105 2021-05-27 06:04:44       2021           5
#> 7          FALSE             35 2021-03-01 10:17:59       2021           3
#> 8           TRUE             50 2021-04-19 15:52:32       2021           4
#> 9          FALSE            171 2021-06-02 23:20:29       2021           6
#> 10          TRUE            325 2021-06-22 13:24:14       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         13         15           Tue         FALSE            102          5
#> 2         10         10           Wed         FALSE             68         11
#> 3          9         19           Sun          TRUE            128          1
#> 4          2         18           Sun          TRUE            121         19
#> 5         28         13           Sun          TRUE             86         12
#> 6         27         21           Thu         FALSE            146          6
#> 7          1          9           Mon         FALSE             59         10
#> 8         19         16           Mon         FALSE            108         15
#> 9          2         22           Wed         FALSE            152         23
#> 10        22         25           Tue         FALSE            172         13
#>    times_minute times_minutes times_second
#> 1            29           329           13
#> 2            21           681           34
#> 3            26            86           40
#> 4            33          1173           52
#> 5            17           737           45
#> 6             4           364           44
#> 7            17           617           59
#> 8            52           952           32
#> 9            20          1400           29
#> 10           24           804           14

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 1  2020     6  22   25     Mon   FALSE      173
#> 8  2020     6  29   26     Mon   FALSE      180
#> 29 2020     7  20   29     Mon   FALSE      201
#> 47 2020     8   7   32     Fri   FALSE      219
#> 55 2020     8  15   33     Sat    TRUE      227
#> 57 2020     8  17   33     Mon   FALSE      229
#>                            holidayname      holidaytype holiday_national
#> 1                         Sacred Heart National holiday             TRUE
#> 8  Feast of Saint Peter and Saint Paul National holiday             TRUE
#> 29                    Independence Day National holiday             TRUE
#> 47                Battle of Boyacá Day National holiday             TRUE
#> 55                  Assumption of Mary National holiday             TRUE
#> 57      Day off for Assumption of Mary National holiday             TRUE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 1               FALSE        FALSE           FALSE          FALSE
#> 8               FALSE        FALSE           FALSE          FALSE
#> 29              FALSE        FALSE           FALSE          FALSE
#> 47              FALSE        FALSE           FALSE          FALSE
#> 55              FALSE        FALSE           FALSE          FALSE
#> 57              FALSE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 1           FALSE
#> 8           FALSE
#> 29          FALSE
#> 47          FALSE
#> 55          FALSE
#> 57          FALSE

# }