One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

data(dfl) # Dummy dataset
# Imput as a vector or dataframe
df <- date_feats(dfl, keep_originals = TRUE)

#> >>> Processing 1 date/time columns: 'opp_date'

# Same as running date_feats(dfl$opp_date)
head(df, 10)

#>               opp_date opp_date_year opp_date_month opp_date_day opp_date_week
#> 1  2019-02-06 12:22:48          2019              2            6             6
#> 2  2019-02-11 20:50:40          2019              2           11             6
#> 3  2019-02-05 21:47:31          2019              2            5             6
#> 4  2019-02-11 16:40:53          2019              2           11             6
#> 5  2019-01-29 17:08:57          2019              1           29             5
#> 6  2019-01-30 01:19:15          2019              1           30             5
#> 7  2019-01-30 18:28:46          2019              1           30             5
#> 8  2019-01-22 12:27:28          2019              1           22             4
#> 9  2019-02-09 17:48:51          2019              2            9             6
#> 10 2019-01-16 20:39:38          2019              1           16             3
#>    opp_date_weekday opp_date_weekend opp_date_year_day opp_date_hour
#> 1               Wed            FALSE                36            12
#> 2               Mon            FALSE                41            20
#> 3               Tue            FALSE                35            21
#> 4               Mon            FALSE                41            16
#> 5               Tue            FALSE                28            17
#> 6               Wed            FALSE                29             1
#> 7               Wed            FALSE                29            18
#> 8               Tue            FALSE                21            12
#> 9               Sat             TRUE                39            17
#> 10              Wed            FALSE                15            20
#>    opp_date_minute opp_date_minutes opp_date_second
#> 1               22              742              48
#> 2               50             1250              40
#> 3               47             1307              31
#> 4               40             1000              53
#> 5                8             1028              57
#> 6               19               79              15
#> 7               28             1108              46
#> 8               27              747              28
#> 9               48             1068              51
#> 10              39             1239              38