One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-12-14       2020          12        14         50           Mon
#> 2  2020-07-02       2020           7         2         27           Thu
#> 3  2020-12-30       2020          12        30         53           Wed
#> 4  2021-04-20       2021           4        20         16           Tue
#> 5  2020-10-14       2020          10        14         42           Wed
#> 6  2021-04-13       2021           4        13         15           Tue
#> 7  2021-02-02       2021           2         2          5           Tue
#> 8  2021-02-17       2021           2        17          7           Wed
#> 9  2021-06-18       2021           6        18         25           Fri
#> 10 2020-11-18       2020          11        18         47           Wed
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            348 2021-04-09 16:07:49       2021           4
#> 2          FALSE            183 2021-03-06 22:00:10       2021           3
#> 3          FALSE            364 2021-05-05 12:05:16       2021           5
#> 4          FALSE            109 2021-04-29 06:12:28       2021           4
#> 5          FALSE            287 2021-03-24 22:56:21       2021           3
#> 6          FALSE            102 2021-05-23 16:43:20       2021           5
#> 7          FALSE             32 2021-02-25 20:56:35       2021           2
#> 8          FALSE             47 2021-04-16 02:31:08       2021           4
#> 9          FALSE            168 2021-05-30 09:59:05       2021           5
#> 10         FALSE            322 2021-06-19 00:02:50       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          9         15           Fri         FALSE             98         16
#> 2          6         10           Sat          TRUE             64         22
#> 3          5         18           Wed         FALSE            124         12
#> 4         29         17           Thu         FALSE            118          6
#> 5         24         12           Wed         FALSE             82         22
#> 6         23         21           Sun          TRUE            142         16
#> 7         25          8           Thu         FALSE             55         20
#> 8         16         16           Fri         FALSE            105          2
#> 9         30         22           Sun          TRUE            149          9
#> 10        19         25           Sat          TRUE            169          0
#>    times_minute times_minutes times_second
#> 1             7           967           49
#> 2             0          1320           10
#> 3             5           725           16
#> 4            12           372           28
#> 5            56          1376           21
#> 6            43          1003           20
#> 7            56          1256           35
#> 8            31           151            8
#> 9            59           599            5
#> 10            2             2           50

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 2  2020     6  20   25     Sat    TRUE      171
#> 3  2020     6  21   25     Sun    TRUE      172
#> 4  2020     6  22   25     Mon   FALSE      173
#> 11 2020     6  29   26     Mon   FALSE      180
#> 32 2020     7  20   29     Mon   FALSE      201
#> 50 2020     8   7   32     Fri   FALSE      219
#>                            holidayname      holidaytype holiday_national
#> 2                        June Solstice           Season            FALSE
#> 3                         Father's Day       Observance            FALSE
#> 4                         Sacred Heart National holiday             TRUE
#> 11 Feast of Saint Peter and Saint Paul National holiday             TRUE
#> 32                    Independence Day National holiday             TRUE
#> 50                Battle of Boyacá Day National holiday             TRUE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 2               FALSE        FALSE           FALSE           TRUE
#> 3                TRUE        FALSE           FALSE          FALSE
#> 4               FALSE        FALSE           FALSE          FALSE
#> 11              FALSE        FALSE           FALSE          FALSE
#> 32              FALSE        FALSE           FALSE          FALSE
#> 50              FALSE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 2           FALSE
#> 3           FALSE
#> 4           FALSE
#> 11          FALSE
#> 32          FALSE
#> 50          FALSE

# }