One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-10-24       2020          10        24         43           Sat
#> 2  2020-08-22       2020           8        22         34           Sat
#> 3  2020-08-23       2020           8        23         34           Sun
#> 4  2021-02-04       2021           2         4          5           Thu
#> 5  2020-12-07       2020          12         7         49           Mon
#> 6  2021-05-23       2021           5        23         21           Sun
#> 7  2021-05-07       2021           5         7         19           Fri
#> 8  2020-10-08       2020          10         8         41           Thu
#> 9  2021-05-14       2021           5        14         20           Fri
#> 10 2021-04-17       2021           4        17         16           Sat
#>    dates_weekend dates_year_day               times times_year times_month
#> 1           TRUE            297 2021-04-05 05:40:22       2021           4
#> 2           TRUE            234 2021-03-08 10:38:08       2021           3
#> 3           TRUE            235 2021-02-24 06:56:35       2021           2
#> 4          FALSE             34 2021-02-08 17:03:32       2021           2
#> 5          FALSE            341 2021-03-10 01:41:09       2021           3
#> 6           TRUE            142 2021-04-07 03:08:47       2021           4
#> 7          FALSE            126 2021-02-21 01:23:55       2021           2
#> 8          FALSE            281 2021-04-23 02:38:53       2021           4
#> 9          FALSE            133 2021-03-19 05:54:10       2021           3
#> 10          TRUE            106 2021-05-17 10:59:08       2021           5
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          5         14           Mon         FALSE             94          5
#> 2          8         10           Mon         FALSE             66         10
#> 3         24          8           Wed         FALSE             54          6
#> 4          8          6           Mon         FALSE             38         17
#> 5         10         10           Wed         FALSE             68          1
#> 6          7         14           Wed         FALSE             96          3
#> 7         21          8           Sun          TRUE             51          1
#> 8         23         17           Fri         FALSE            112          2
#> 9         19         12           Fri         FALSE             77          5
#> 10        17         20           Mon         FALSE            136         10
#>    times_minute times_minutes times_second
#> 1            40           340           22
#> 2            38           638            8
#> 3            56           416           35
#> 4             3          1023           32
#> 5            41           101            9
#> 6             8           188           47
#> 7            23            83           55
#> 8            38           158           53
#> 9            54           354           10
#> 10           59           659            8

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}