One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-03-10       2020           3        10         10           Tue
#> 2  2020-01-07       2020           1         7          1           Tue
#> 3  2020-01-08       2020           1         8          2           Wed
#> 4  2020-06-21       2020           6        21         25           Sun
#> 5  2020-04-23       2020           4        23         17           Thu
#> 6  2020-10-07       2020          10         7         41           Wed
#> 7  2020-09-21       2020           9        21         38           Mon
#> 8  2020-02-23       2020           2        23          8           Sun
#> 9  2020-09-28       2020           9        28         39           Mon
#> 10 2020-09-01       2020           9         1         35           Tue
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE             69 2020-08-20 09:12:59       2020           8
#> 2          FALSE              6 2020-07-23 14:10:45       2020           7
#> 3          FALSE              7 2020-07-11 10:29:12       2020           7
#> 4           TRUE            172 2020-06-25 20:36:09       2020           6
#> 5          FALSE            113 2020-07-25 05:13:46       2020           7
#> 6          FALSE            280 2020-08-22 06:41:24       2020           8
#> 7          FALSE            264 2020-07-08 04:56:32       2020           7
#> 8           TRUE             53 2020-09-07 06:11:30       2020           9
#> 9          FALSE            271 2020-08-03 09:26:47       2020           8
#> 10         FALSE            244 2020-10-01 14:31:45       2020          10
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         20         34           Thu         FALSE            232          9
#> 2         23         30           Thu         FALSE            204         14
#> 3         11         28           Sat          TRUE            192         10
#> 4         25         26           Thu         FALSE            176         20
#> 5         25         30           Sat          TRUE            206          5
#> 6         22         34           Sat          TRUE            234          6
#> 7          8         28           Wed         FALSE            189          4
#> 8          7         36           Mon         FALSE            250          6
#> 9          3         31           Mon         FALSE            215          9
#> 10         1         40           Thu         FALSE            274         14
#>    times_minute times_minutes times_second
#> 1            12           552           59
#> 2            10           850           45
#> 3            29           629           12
#> 4            36          1236            9
#> 5            13           313           46
#> 6            41           401           24
#> 7            56           296           32
#> 8            11           371           30
#> 9            26           566           47
#> 10           31           871           45

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}