One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

Usage

date_feats(
  dates,
  drop = FALSE,
  only = NA,
  append = FALSE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates

Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

drop

Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only

Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

append

Boolean. Append results to existing data.frame? If FALSE, only calculated values will be returned.

holidays

Boolean. Include holidays as new columns?

country

Character or vector. For which countries should the holidays be included?

currency_pair

Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet

Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50)
)

# Input as a vector or dataframe
date_feats(df, drop = TRUE) %>% head(10)
#> >>> Processing 2 date/time columns: 'dates', 'times'
#> Warning: argument 'pattern' has length > 1 and only the first element will be used
#> # A tibble: 10 × 14
#>    dates_year dates_mo…¹ dates…² dates…³ dates…⁴ dates…⁵ dates…⁶ times…⁷ times…⁸
#>         <dbl>      <dbl>   <int>   <dbl> <chr>   <lgl>     <int>   <dbl>   <dbl>
#>  1       2022          8      19      33 Fri     FALSE       230    2022      12
#>  2       2022          3       7      10 Mon     FALSE        65    2022      11
#>  3       2022          9       4      36 Sun     TRUE        246    2023       1
#>  4       2022         12      24      52 Sat     TRUE        357    2023       1
#>  5       2022          6      19      25 Sun     TRUE        169    2022      11
#>  6       2022         12      17      51 Sat     TRUE        350    2023       1
#>  7       2022         10       8      41 Sat     TRUE        280    2022      11
#>  8       2022         10      23      43 Sun     TRUE        295    2022      12
#>  9       2023          2      21       8 Tue     FALSE        51    2023       2
#> 10       2022          7      24      30 Sun     TRUE        204    2023       2
#> # … with 5 more variables: times_day <int>, times_week <dbl>,
#> #   times_weekday <chr>, times_weekend <lgl>, times_year_day <int>, and
#> #   abbreviated variable names ¹​dates_month, ²​dates_day, ³​dates_week,
#> #   ⁴​dates_weekday, ⁵​dates_weekend, ⁶​dates_year_day, ⁷​times_year, ⁸​times_month

# Holidays given a date range and country
# \donttest{
hol <- date_feats(
  seq(Sys.Date() - 365, Sys.Date(), by = 1),
  holidays = TRUE,
  country = "Colombia"
)
#> >>> Processing 1 date/time columns: 'df'
#> >>> Extracting Colombia's holidays for 2022
#> >>> Extracting Colombia's holidays for 2023
#> Warning: There was 1 warning in `mutate()`.
#> ℹ In argument: `values_date_holiday_county = (structure(function (..., .x =
#>   ..1, .y = ..2, . = ..1) ...`.
#> Caused by warning in `[<-.factor`:
#> ! invalid factor level, NA generated
head(hol[!is.na(hol$holidayname), ])
#> # A tibble: 6 × 16
#>   values month   day  week weekday weekend year_day holidayname  holid…¹ holid…²
#>    <dbl> <dbl> <int> <dbl> <chr>   <lgl>      <int> <chr>        <chr>   <lgl>  
#> 1   2022     3     8    10 Tue     FALSE         66 Women's Day  Observ… FALSE  
#> 2   2022     3    19    12 Sat     TRUE          77 Saint Josep… Observ… FALSE  
#> 3   2022     3    20    12 Sun     TRUE          78 March Equin… Season  FALSE  
#> 4   2022     3    21    12 Mon     FALSE         79 Saint Josep… Nation… TRUE   
#> 5   2022     4    10    15 Sun     TRUE          99 Palm Sunday  Observ… FALSE  
#> 6   2022     4    14    15 Thu     FALSE        103 Maundy Thur… Nation… TRUE   
#> # … with 6 more variables: holiday_observance <lgl>, holiday_bank <lgl>,
#> #   holiday_nonwork <lgl>, holiday_season <lgl>, holiday_hother <lgl>,
#> #   holiday_county <fct>, and abbreviated variable names ¹​holidaytype,
#> #   ²​holiday_national
# }