One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-11-29       2020          11        29         48           Sun
#> 2  2020-06-17       2020           6        17         25           Wed
#> 3  2020-12-15       2020          12        15         50           Tue
#> 4  2021-04-05       2021           4         5         14           Mon
#> 5  2020-09-29       2020           9        29         39           Tue
#> 6  2021-03-29       2021           3        29         13           Mon
#> 7  2021-01-18       2021           1        18          3           Mon
#> 8  2021-02-02       2021           2         2          5           Tue
#> 9  2021-06-03       2021           6         3         22           Thu
#> 10 2020-11-03       2020          11         3         44           Tue
#>    dates_weekend dates_year_day               times times_year times_month
#> 1           TRUE            333 2021-03-26 04:23:11       2021           3
#> 2          FALSE            168 2021-02-20 10:15:32       2021           2
#> 3          FALSE            349 2021-04-21 00:20:38       2021           4
#> 4          FALSE             94 2021-04-14 18:27:50       2021           4
#> 5          FALSE            272 2021-03-10 11:11:43       2021           3
#> 6          FALSE             87 2021-05-09 04:58:42       2021           5
#> 7          FALSE             17 2021-02-11 09:11:57       2021           2
#> 8          FALSE             32 2021-04-01 14:46:30       2021           4
#> 9          FALSE            153 2021-05-15 22:14:27       2021           5
#> 10         FALSE            307 2021-06-04 12:18:12       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         26         13           Fri         FALSE             84          4
#> 2         20          8           Sat          TRUE             50         10
#> 3         21         16           Wed         FALSE            110          0
#> 4         14         15           Wed         FALSE            103         18
#> 5         10         10           Wed         FALSE             68         11
#> 6          9         19           Sun          TRUE            128          4
#> 7         11          6           Thu         FALSE             41          9
#> 8          1         13           Thu         FALSE             90         14
#> 9         15         20           Sat          TRUE            134         22
#> 10         4         23           Fri         FALSE            154         12
#>    times_minute times_minutes times_second
#> 1            23           263           11
#> 2            15           615           32
#> 3            20            20           38
#> 4            27          1107           50
#> 5            11           671           43
#> 6            58           298           42
#> 7            11           551           57
#> 8            46           886           30
#> 9            14          1334           27
#> 10           18           738           12

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 8  2020     6  11   24     Thu   FALSE      162
#> 12 2020     6  15   24     Mon   FALSE      166
#> 17 2020     6  20   25     Sat    TRUE      171
#> 18 2020     6  21   25     Sun    TRUE      172
#> 19 2020     6  22   25     Mon   FALSE      173
#> 26 2020     6  29   26     Mon   FALSE      180
#>                            holidayname      holidaytype holiday_national
#> 8                       Corpus Christi       Observance            FALSE
#> 12                      Corpus Christi National holiday             TRUE
#> 17                       June Solstice           Season            FALSE
#> 18                        Father's Day       Observance            FALSE
#> 19                        Sacred Heart National holiday             TRUE
#> 26 Feast of Saint Peter and Saint Paul National holiday             TRUE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 8                TRUE        FALSE           FALSE          FALSE
#> 12              FALSE        FALSE           FALSE          FALSE
#> 17              FALSE        FALSE           FALSE           TRUE
#> 18               TRUE        FALSE           FALSE          FALSE
#> 19              FALSE        FALSE           FALSE          FALSE
#> 26              FALSE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 8           FALSE
#> 12          FALSE
#> 17          FALSE
#> 18          FALSE
#> 19          FALSE
#> 26          FALSE

# }