One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-09-29       2020           9        29         39           Tue
#> 2  2020-07-28       2020           7        28         30           Tue
#> 3  2020-07-29       2020           7        29         31           Wed
#> 4  2021-01-10       2021           1        10          2           Sun
#> 5  2020-11-12       2020          11        12         46           Thu
#> 6  2021-04-28       2021           4        28         17           Wed
#> 7  2021-04-12       2021           4        12         15           Mon
#> 8  2020-09-13       2020           9        13         37           Sun
#> 9  2021-04-19       2021           4        19         16           Mon
#> 10 2021-03-23       2021           3        23         12           Tue
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            272 2021-03-11 04:13:33       2021           3
#> 2          FALSE            209 2021-02-11 09:11:19       2021           2
#> 3          FALSE            210 2021-01-30 05:29:46       2021           1
#> 4           TRUE              9 2021-01-14 15:36:43       2021           1
#> 5          FALSE            316 2021-02-13 00:14:20       2021           2
#> 6          FALSE            117 2021-03-13 01:41:58       2021           3
#> 7          FALSE            101 2021-01-26 23:57:06       2021           1
#> 8           TRUE            256 2021-03-29 01:12:04       2021           3
#> 9          FALSE            108 2021-02-22 04:27:21       2021           2
#> 10         FALSE             81 2021-04-22 09:32:19       2021           4
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         11         10           Thu         FALSE             69          4
#> 2         11          6           Thu         FALSE             41          9
#> 3         30          5           Sat          TRUE             29          5
#> 4         14          2           Thu         FALSE             13         15
#> 5         13          7           Sat          TRUE             43          0
#> 6         13         11           Sat          TRUE             71          1
#> 7         26          4           Tue         FALSE             25         23
#> 8         29         13           Mon         FALSE             87          1
#> 9         22          8           Mon         FALSE             52          4
#> 10        22         16           Thu         FALSE            111          9
#>    times_minute times_minutes times_second
#> 1            13           253           33
#> 2            11           551           19
#> 3            29           329           46
#> 4            36           936           43
#> 5            14            14           20
#> 6            41           101           58
#> 7            57          1437            6
#> 8            12            72            4
#> 9            27           267           21
#> 10           32           572           19

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}