One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  drop = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

drop
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50)
)

# Input as a vector or dataframe
date_feats(df, drop = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> # A tibble: 10 × 22
#>    dates_year dates_month dates_day dates_week dates_weekday dates_weekend
#>         <dbl>       <dbl>     <int>      <dbl> <chr>         <lgl>        
#>  1       2021           1        22          4 Fri           FALSE        
#>  2       2020           8        10         32 Mon           FALSE        
#>  3       2021           2         7          6 Sun           TRUE         
#>  4       2021           5        29         22 Sat           TRUE         
#>  5       2020          11        22         47 Sun           TRUE         
#>  6       2021           5        22         21 Sat           TRUE         
#>  7       2021           3        13         11 Sat           TRUE         
#>  8       2021           3        28         13 Sun           TRUE         
#>  9       2021           7        27         30 Tue           FALSE        
#> 10       2020          12        27         52 Sun           TRUE         
#> # … with 16 more variables: dates_year_day <int>, dates_hour <int>,
#> #   dates_minute <int>, dates_minutes <int>, dates_second <dbl>,
#> #   times_year <dbl>, times_month <dbl>, times_day <int>, times_week <dbl>,
#> #   times_weekday <chr>, times_weekend <lgl>, times_year_day <int>,
#> #   times_hour <int>, times_minute <int>, times_minutes <int>,
#> #   times_second <dbl>

# Holidays given a date range and country
# \donttest{
hol <- date_feats(
  seq(Sys.Date() - 365, Sys.Date(), by = 1),
  holidays = TRUE,
  country = "Colombia"
)

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> >>> Extracting Colombia's holidays for 2021

#> Warning: invalid factor level, NA generated

head(hol[!is.na(hol$holidayname), ])

#> # A tibble: 6 × 20
#>    year month   day  week weekday weekend year_day  hour minute minutes second
#>   <dbl> <dbl> <int> <dbl> <chr>   <lgl>      <int> <int>  <int>   <int>  <dbl>
#> 1  2020     8     7    32 Fri     FALSE        219     0      0       0      0
#> 2  2020     8    15    33 Sat     TRUE         227     0      0       0      0
#> 3  2020     8    17    33 Mon     FALSE        229     0      0       0      0
#> 4  2020     9    19    38 Sat     TRUE         262     0      0       0      0
#> 5  2020     9    22    38 Tue     FALSE        265     0      0       0      0
#> 6  2020    10    12    41 Mon     FALSE        285     0      0       0      0
#> # … with 9 more variables: holidayname <chr>, holidaytype <chr>,
#> #   holiday_national <lgl>, holiday_observance <lgl>, holiday_bank <lgl>,
#> #   holiday_nonwork <lgl>, holiday_season <lgl>, holiday_hother <lgl>,
#> #   holiday_county <fct>

# }