One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-10-18       2020          10        18         42           Sun
#> 2  2020-08-16       2020           8        16         33           Sun
#> 3  2020-08-17       2020           8        17         33           Mon
#> 4  2021-01-29       2021           1        29          5           Fri
#> 5  2020-12-01       2020          12         1         48           Tue
#> 6  2021-05-17       2021           5        17         20           Mon
#> 7  2021-05-01       2021           5         1         18           Sat
#> 8  2020-10-02       2020          10         2         40           Fri
#> 9  2021-05-08       2021           5         8         19           Sat
#> 10 2021-04-11       2021           4        11         15           Sun
#>    dates_weekend dates_year_day               times times_year times_month
#> 1           TRUE            291 2021-03-30 07:57:13       2021           3
#> 2           TRUE            228 2021-03-02 12:54:59       2021           3
#> 3          FALSE            229 2021-02-18 09:13:26       2021           2
#> 4          FALSE             28 2021-02-02 19:20:23       2021           2
#> 5          FALSE            335 2021-03-04 03:58:00       2021           3
#> 6          FALSE            136 2021-04-01 05:25:38       2021           4
#> 7           TRUE            120 2021-02-15 03:40:46       2021           2
#> 8          FALSE            275 2021-04-17 04:55:44       2021           4
#> 9           TRUE            127 2021-03-13 08:11:01       2021           3
#> 10          TRUE            100 2021-05-11 13:15:59       2021           5
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         30         13           Tue         FALSE             88          7
#> 2          2          9           Tue         FALSE             60         12
#> 3         18          7           Thu         FALSE             48          9
#> 4          2          5           Tue         FALSE             32         19
#> 5          4          9           Thu         FALSE             62          3
#> 6          1         13           Thu         FALSE             90          5
#> 7         15          7           Mon         FALSE             45          3
#> 8         17         16           Sat          TRUE            106          4
#> 9         13         11           Sat          TRUE             71          8
#> 10        11         19           Tue         FALSE            130         13
#>    times_minute times_minutes times_second
#> 1            57           477           13
#> 2            54           774           59
#> 3            13           553           26
#> 4            20          1160           23
#> 5            58           238            0
#> 6            25           325           38
#> 7            40           220           46
#> 8            55           295           44
#> 9            11           491            1
#> 10           15           795           59

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}