One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

Value

data.frame with additional features calculated out of time or date vectors.

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-11-27       2020          11        27         48           Fri
#> 2  2020-06-15       2020           6        15         24           Mon
#> 3  2020-12-13       2020          12        13         50           Sun
#> 4  2021-04-03       2021           4         3         14           Sat
#> 5  2020-09-27       2020           9        27         39           Sun
#> 6  2021-03-27       2021           3        27         13           Sat
#> 7  2021-01-16       2021           1        16          3           Sat
#> 8  2021-01-31       2021           1        31          5           Sun
#> 9  2021-06-01       2021           6         1         22           Tue
#> 10 2020-11-01       2020          11         1         44           Sun
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            331 2021-03-24 03:45:48       2021           3
#> 2          FALSE            166 2021-02-18 09:38:09       2021           2
#> 3           TRUE            347 2021-04-18 23:43:15       2021           4
#> 4           TRUE             92 2021-04-12 17:50:27       2021           4
#> 5           TRUE            270 2021-03-08 10:34:20       2021           3
#> 6           TRUE             85 2021-05-07 04:21:19       2021           5
#> 7           TRUE             15 2021-02-09 08:34:34       2021           2
#> 8           TRUE             30 2021-03-30 14:09:07       2021           3
#> 9          FALSE            151 2021-05-13 21:37:04       2021           5
#> 10          TRUE            305 2021-06-02 11:40:49       2021           6
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1         24         12           Wed         FALSE             82          3
#> 2         18          7           Thu         FALSE             48          9
#> 3         18         16           Sun          TRUE            107         23
#> 4         12         15           Mon         FALSE            101         17
#> 5          8         10           Mon         FALSE             66         10
#> 6          7         19           Fri         FALSE            126          4
#> 7          9          6           Tue         FALSE             39          8
#> 8         30         13           Tue         FALSE             88         14
#> 9         13         19           Thu         FALSE            132         21
#> 10         2         22           Wed         FALSE            152         11
#>    times_minute times_minutes times_second
#> 1            45           225           48
#> 2            38           578            9
#> 3            43          1423           15
#> 4            50          1070           27
#> 5            34           634           20
#> 6            21           261           19
#> 7            34           514           34
#> 8             9           849            7
#> 9            37          1297            4
#> 10           40           700           49

# Holidays given a date range and country
# \donttest{
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")

#> >>> Processing 1 date/time columns: 'df'

#> >>> Extracting Colombia's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Colombia's holidays for 2021

#> Warning: NAs introduced by coercion

#> Warning: All formats failed to parse. No formats found.

  head(hol[!is.na(hol$holidayname),])

#>    year month day week weekday weekend year_day
#> 10 2020     6  11   24     Thu   FALSE      162
#> 14 2020     6  15   24     Mon   FALSE      166
#> 19 2020     6  20   25     Sat    TRUE      171
#> 20 2020     6  21   25     Sun    TRUE      172
#> 21 2020     6  22   25     Mon   FALSE      173
#> 28 2020     6  29   26     Mon   FALSE      180
#>                            holidayname      holidaytype holiday_national
#> 10                      Corpus Christi       Observance            FALSE
#> 14                      Corpus Christi National holiday             TRUE
#> 19                       June Solstice           Season            FALSE
#> 20                        Father's Day       Observance            FALSE
#> 21                        Sacred Heart National holiday             TRUE
#> 28 Feast of Saint Peter and Saint Paul National holiday             TRUE
#>    holiday_observance holiday_bank holiday_nonwork holiday_season
#> 10               TRUE        FALSE           FALSE          FALSE
#> 14              FALSE        FALSE           FALSE          FALSE
#> 19              FALSE        FALSE           FALSE           TRUE
#> 20               TRUE        FALSE           FALSE          FALSE
#> 21              FALSE        FALSE           FALSE          FALSE
#> 28              FALSE        FALSE           FALSE          FALSE
#>    holiday_hother
#> 10          FALSE
#> 14          FALSE
#> 19          FALSE
#> 20          FALSE
#> 21          FALSE
#> 28          FALSE

# }