One Hot Encoding for Date/Time Variables (Dummy Variables)

This function lets the user automatically create new columns out of a dataframe or vector with date/time variables.

date_feats(
  dates,
  keep_originals = FALSE,
  only = NA,
  features = TRUE,
  holidays = FALSE,
  country = "Venezuela",
  currency_pair = NA,
  quiet = FALSE
)

Arguments

dates
Vector or dataframe. Non-date/time columns will be automatically ignored/extracted.

keep_originals
Boolean. Should the original date/time columns be kept in the results? Only valid when input is a dataframe.

only
Character or vector. Which columns do you wish to process? If non are explicitly defined, all will be processed

features
Create features out of date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

quiet
Boolean. Quiet all messages?

See also

Examples

df <- data.frame(
  dates = sample(seq(Sys.Date() - 365, Sys.Date(), by = 1), 50),
  times = sample(seq(Sys.time() - 1e7, Sys.time(), by = 1), 50))
  
# Input as a vector or dataframe
date_feats(df, keep_originals = TRUE) %>% head(10)

#> >>> Processing 2 date/time columns: 'dates', 'times'

#> Warning: All formats failed to parse. No formats found.

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#> Warning: argument 'replacement' has length > 1 and only the first element will be used

#>         dates dates_year dates_month dates_day dates_week dates_weekday
#> 1  2020-05-29       2020           5        29         22           Fri
#> 2  2020-03-27       2020           3        27         13           Fri
#> 3  2020-03-28       2020           3        28         13           Sat
#> 4  2020-09-09       2020           9         9         37           Wed
#> 5  2020-07-12       2020           7        12         28           Sun
#> 6  2020-12-26       2020          12        26         52           Sat
#> 7  2020-12-10       2020          12        10         50           Thu
#> 8  2020-05-13       2020           5        13         20           Wed
#> 9  2020-12-17       2020          12        17         51           Thu
#> 10 2020-11-20       2020          11        20         47           Fri
#>    dates_weekend dates_year_day               times times_year times_month
#> 1          FALSE            149 2020-11-08 09:32:22       2020          11
#> 2          FALSE             86 2020-10-11 14:30:08       2020          10
#> 3           TRUE             87 2020-09-29 10:48:35       2020           9
#> 4          FALSE            252 2020-09-13 20:55:32       2020           9
#> 5           TRUE            193 2020-10-13 05:33:09       2020          10
#> 6           TRUE            360 2020-11-10 07:00:47       2020          11
#> 7          FALSE            344 2020-09-26 05:15:55       2020           9
#> 8          FALSE            133 2020-11-26 06:30:53       2020          11
#> 9          FALSE            351 2020-10-22 09:46:10       2020          10
#> 10         FALSE            324 2020-12-20 14:51:08       2020          12
#>    times_day times_week times_weekday times_weekend times_year_day times_hour
#> 1          8         45           Sun          TRUE            312          9
#> 2         11         41           Sun          TRUE            284         14
#> 3         29         39           Tue         FALSE            272         10
#> 4         13         37           Sun          TRUE            256         20
#> 5         13         41           Tue         FALSE            286          5
#> 6         10         45           Tue         FALSE            314          7
#> 7         26         39           Sat          TRUE            269          5
#> 8         26         48           Thu         FALSE            330          6
#> 9         22         43           Thu         FALSE            295          9
#> 10        20         51           Sun          TRUE            354         14
#>    times_minute times_minutes times_second
#> 1            32           572           22
#> 2            30           870            8
#> 3            48           648           35
#> 4            55          1255           32
#> 5            33           333            9
#> 6             0           420           47
#> 7            15           315           55
#> 8            30           390           53
#> 9            46           586           10
#> 10           51           891            8

# Holidays
if (FALSE) {
  hol <- date_feats(
    seq(Sys.Date() - 365, Sys.Date(), by = 1), 
    keep_originals = TRUE, 
    holidays = TRUE, 
    country = "Colombia")
  head(hol[!is.na(hol$holidayname),])
}