Export h2o_automl's Results

Export RDS, TXT, POJO, MOJO and all results from h2o_automl().

export_results(
  results,
  thresh = 10,
  which = c("txt", "csv", "rds", "binary", "mojo", "plots", "dev", "production"),
  note = NA,
  subdir = NA,
  save = TRUE
)

Arguments

results
h2o_automl or h2o model

thresh
Integer. Threshold for selecting binary or regression models: this number is the threshold of unique values we should have in 'tag' (more than: regression; less than: classification)

which
Character vector. Select which file format to export: Possible values: txt, csv, rds, binary, mojo, plots. You might also use dev (txt, csv, rds) or production (binary, mojo) or simply don't use parameter to export everything

note
Character. Add a note to the txt file. Useful when lots of models are trained and saved to remember which one is which one

subdir
Character. In which directory do you wish to save the results?

save
Boolean. Do you wish to save/export results?

See also