Automated H2O's AutoML

This function lets the user create a robust and fast model, using H2O's AutoML function. The result is a list with the best model, its parameters, datasets, performance metrics, variables importance, and plots. Read more about the h2o_automl() pipeline here.

h2o_automl(
  df,
  y = "tag",
  ignore = NULL,
  train_test = NA,
  split = 0.7,
  weight = NULL,
  target = "auto",
  balance = FALSE,
  impute = FALSE,
  no_outliers = TRUE,
  unique_train = TRUE,
  center = FALSE,
  scale = FALSE,
  thresh = 10,
  seed = 0,
  nfolds = 5,
  max_models = 3,
  max_time = 10 * 60,
  start_clean = FALSE,
  exclude_algos = c("StackedEnsemble", "DeepLearning"),
  include_algos = NULL,
  plots = TRUE,
  alarm = TRUE,
  quiet = FALSE,
  print = TRUE,
  save = FALSE,
  subdir = NA,
  project = "AutoML Results",
  ...
)

# S3 method for h2o_automl
plot(x, ...)

# S3 method for h2o_automl
print(x, importance = TRUE, ...)

Arguments

df

Dataframe. Dataframe containing all your data, including the independent variable labeled as 'tag'. If you want to define which variable should be used instead, use the y parameter.

y

Variable or Character. Name of the independent variable.

ignore

Character vector. Force columns for the model to ignore

train_test

Character. If needed, df's column name with 'test' and 'train' values to split

split

Numeric. Value between 0 and 1 to split as train/test datasets. Value is for training set. Set value to 1 to train with all available data and test with same data (cross-validation will still be used when training). If train_test is set, value will be overwritten with its real split rate.

weight

Column with observation weights. Giving some observation a weight of zero is equivalent to excluding it from the dataset; giving an observation a relative weight of 2 is equivalent to repeating that row twice. Negative weights are not allowed.

target

Value. Which is your target positive value? If set to 'auto', the target with largest mean(score) will be selected. Change the value to overwrite. Only used when binary categorical model.

balance

Boolean. Auto-balance train dataset with under-sampling?

impute

Boolean. Fill NA values with MICE?

no_outliers

Boolean/Numeric. Remove y's outliers from the dataset? Will remove those values that are farther than n standard deviations from the independent variable's mean (Z-score). Set to TRUE for default (3) or numeric to set a different multiplier.

unique_train

Boolean. Keep only unique row observations for training data?

center, scale

Boolean. Using the base function scale, do you wish to center and/or scale all numerical values?

thresh

Integer. Threshold for selecting binary or regression models: this number is the threshold of unique values we should have in 'tag' (more than: regression; less than: classification)

seed

Integer. Set a seed for reproducibility. AutoML can only guarantee reproducibility if max_models is used because max_time is resource limited.

nfolds

Number of folds for k-fold cross-validation. Must be >= 2; defaults to 5. Use 0 to disable cross-validation; this will also disable Stacked Ensemble (thus decreasing the overall model performance).

max_models, max_time

Numeric. Max number of models and seconds you wish for the function to iterate. Note that max_models guarantees reproducibility and max_time not (because it depends entirely on your machine's computational characteristics)

start_clean

Boolean. Erase everything in the current h2o instance before we start to train models? You may want to keep other models or not. To group results into a custom common AutoML project, you may use project_name argument.

exclude_algos, include_algos

Vector of character strings. Algorithms to skip or include during the model-building phase. Set NULL to ignore. When both are defined, only include_algos will be valid.

plots

Boolean. Create plots objects?

alarm

Boolean. Ping (sound) when done. Requires beepr.

quiet

Boolean. Quiet all messages, warnings, recommendations?

print

Boolean. Print summary when process ends?

save

Boolean. Do you wish to save/export results into your working directory?

subdir

Character. In which directory do you wish to save the results? Working directory as default.

project

Character. Your project's name

...

Additional parameters on h2o::h2o.automl

x

h2o_automl object

importance

Boolean. Print important variables?

Value

List. Trained model, predicted scores and datasets used, performance metrics, parameters, importance data.frame, seed, and plots when plots=TRUE.

List of algorithms

-> Read more here

DRF

Distributed Random Forest, including Random Forest (RF) and Extremely-Randomized Trees (XRT)

GLM

Generalized Linear Model

XGBoost

eXtreme Grading Boosting

GBM

Gradient Boosting Machine

DeepLearning

Fully-connected multi-layer artificial neural network

StackedEnsemble

Stacked Ensemble

Methods

print

Use print method to print models stats and summary

plot

Use plot method to plot results using mplot_full()

See also

Examples

# \donttest{
# CRAN
data(dft) # Titanic dataset
dft <- subset(dft, select = -c(Ticket, PassengerId, Cabin))

# Classification: Binomial - 2 Classes
r <- h2o_automl(dft, y = Survived, max_models = 1, impute = FALSE, target = "TRUE", alarm = FALSE)
#> 2021-12-15 19:42:39 | Started process...
#> - INDEPENDENT VARIABLE: Survived
#> - MODEL TYPE: Classification
#> # A tibble: 2 × 5
#>   tag       n     p order  pcum
#>   <lgl> <int> <dbl> <int> <dbl>
#> 1 FALSE   549  61.6     1  61.6
#> 2 TRUE    342  38.4     2 100  
#> - MISSINGS: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.
#> - CATEGORICALS: There are 3 non-numerical features. Consider using ohse() or equivalent prior to encode categorical variables.
#> >>> Splitting data: train = 0.7 & test = 0.3
#> train_size  test_size 
#>        623        268 
#> - REPEATED: There were 57 repeated rows which are being suppressed from the train dataset
#> - ALGORITHMS: excluded 'StackedEnsemble', 'DeepLearning'
#> - CACHE: Previous models are not being erased. You may use 'start_clean' [clear] or 'project_name' [join]
#> - UI: You may check results using H2O Flow's interactive platform: http://localhost:54321/flow/index.html
#> >>> Iterating until 1 models or 600 seconds...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> - EUREKA: Succesfully generated 1 models
#>                             model_id       auc   logloss     aucpr
#> 1 XGBoost_1_AutoML_1_20211215_194245 0.8445152 0.4667281 0.8209007
#>   mean_per_class_error     rmse       mse
#> 1            0.2133057 0.385946 0.1489543
#> SELECTED MODEL: XGBoost_1_AutoML_1_20211215_194245
#> - NOTE: The following variables were the least important: SibSp, Embarked.S, Parch, Pclass.2, Embarked.C
#> >>> Running predictions for Survived...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Warning: Test/Validation dataset column 'Embarked' has levels not trained on: [""]
#> Target value: TRUE
#> >>> Generating plots...
#> Model (1/1): XGBoost_1_AutoML_1_20211215_194245
#> Independent Variable: Survived
#> Type: Classification (2 classes)
#> Algorithm: XGBOOST
#> Split: 70% training data (of 891 observations)
#> Seed: 0
#> 
#> Test metrics:
#>    AUC = 0.87173
#>    ACC = 0.14925
#>    PRC = 0.125
#>    TPR = 0.28571
#>    TNR = 0.086957
#> 
#> Most important variables:
#>    Sex.female (25.8%)
#>    Age (18.4%)
#>    Fare (16.2%)
#>    Sex.male (16.1%)
#>    Pclass.3 (13.2%)
#> Process duration: 16.9s

# Let's see all the stuff we have inside:
lapply(r, names)
#> $model
#> NULL
#> 
#> $y
#> NULL
#> 
#> $scores_test
#> [1] "tag"    "scores"
#> 
#> $metrics
#> [1] "dictionary"       "confusion_matrix" "gain_lift"        "metrics"         
#> [5] "cv_metrics"       "max_metrics"     
#> 
#> $parameters
#>  [1] "model_id"                          "training_frame"                   
#>  [3] "nfolds"                            "keep_cross_validation_models"     
#>  [5] "keep_cross_validation_predictions" "fold_assignment"                  
#>  [7] "stopping_tolerance"                "seed"                             
#>  [9] "distribution"                      "categorical_encoding"             
#> [11] "ntrees"                            "max_depth"                        
#> [13] "min_rows"                          "min_child_weight"                 
#> [15] "sample_rate"                       "subsample"                        
#> [17] "col_sample_rate"                   "colsample_bylevel"                
#> [19] "col_sample_rate_per_tree"          "colsample_bytree"                 
#> [21] "score_tree_interval"               "tree_method"                      
#> [23] "dmatrix_type"                      "backend"                          
#> [25] "x"                                 "y"                                
#> 
#> $importance
#> [1] "variable"            "relative_importance" "scaled_importance"  
#> [4] "importance"         
#> 
#> $datasets
#> [1] "global" "test"  
#> 
#> $scoring_history
#> [1] "timestamp"                     "duration"                     
#> [3] "number_of_trees"               "training_rmse"                
#> [5] "training_logloss"              "training_auc"                 
#> [7] "training_pr_auc"               "training_lift"                
#> [9] "training_classification_error"
#> 
#> $categoricals
#> [1] "Pclass"     "Sex"        "Embarked"   "train_test" "predict"   
#> 
#> $type
#> NULL
#> 
#> $split
#> NULL
#> 
#> $threshold
#> NULL
#> 
#> $model_name
#> NULL
#> 
#> $algorithm
#> NULL
#> 
#> $leaderboard
#> [1] "model_id"             "auc"                  "logloss"             
#> [4] "aucpr"                "mean_per_class_error" "rmse"                
#> [7] "mse"                 
#> 
#> $project
#> NULL
#> 
#> $seed
#> NULL
#> 
#> $h2o
#> NULL
#> 
#> $plots
#> [1] "dashboard"  "metrics"    "importance"
#> 

# Classification: Multi-Categorical - 3 Classes
r <- h2o_automl(dft, Pclass, ignore = c("Fare", "Cabin"), max_time = 30, plots = FALSE)
#> 2021-12-15 19:42:56 | Started process...
#> - INDEPENDENT VARIABLE: Pclass
#> - MODEL TYPE: Classification
#> # A tibble: 3 × 5
#>   tag       n     p order  pcum
#>   <fct> <int> <dbl> <int> <dbl>
#> 1 n_3     491  55.1     1  55.1
#> 2 n_1     216  24.2     2  79.4
#> 3 n_2     184  20.6     3 100  
#> - MISSINGS: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.
#> - CATEGORICALS: There are 3 non-numerical features. Consider using ohse() or equivalent prior to encode categorical variables.
#> >>> Splitting data: train = 0.7 & test = 0.3
#> train_size  test_size 
#>        623        268 
#> - REPEATED: There were 65 repeated rows which are being suppressed from the train dataset
#> - ALGORITHMS: excluded 'StackedEnsemble', 'DeepLearning'
#> - CACHE: Previous models are not being erased. You may use 'start_clean' [clear] or 'project_name' [join]
#> - UI: You may check results using H2O Flow's interactive platform: http://localhost:54321/flow/index.html
#> >>> Iterating until 3 models or 30 seconds...
#> - EUREKA: Succesfully generated 3 models
#>                             model_id mean_per_class_error   logloss      rmse
#> 1     GLM_1_AutoML_2_20211215_194257            0.4742325 0.8170807 0.5409388
#> 2 XGBoost_1_AutoML_2_20211215_194257            0.4946035 0.8255072 0.5392879
#> 3     GBM_1_AutoML_2_20211215_194257            0.5037168 0.8620436 0.5616772
#>         mse
#> 1 0.2926147
#> 2 0.2908315
#> 3 0.3154812
#> SELECTED MODEL: GLM_1_AutoML_2_20211215_194257
#> - NOTE: The following variables were the least important: Sex.male, Sex.female, Parch
#> >>> Running predictions for Pclass...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Model (1/3): GLM_1_AutoML_2_20211215_194257
#> Independent Variable: Pclass
#> Type: Classification (3 classes)
#> Algorithm: GLM
#> Split: 70% training data (of 891 observations)
#> Seed: 0
#> 
#> Test metrics:
#>    AUC = 0.76337
#>    ACC = 0.64179
#> 
#> Most important variables:
#>    Embarked.Q (25.3%)
#>    Embarked.C (13.5%)
#>    Embarked.S (13.3%)
#>    Age (11.9%)
#>    Survived.FALSE (10.6%)
#> Process duration: 15.6s

# Regression: Continuous Values
r <- h2o_automl(dft, y = "Fare", ignore = c("Pclass"), exclude_algos = NULL, quiet = TRUE)
print(r)
#> Model (1/7): StackedEnsemble_BestOfFamily_1_AutoML_3_20211215_194312
#> Independent Variable: Fare
#> Type: Regression
#> Algorithm: STACKEDENSEMBLE
#> Split: 70% training data (of 871 observations)
#> Seed: 0
#> 
#> Test metrics:
#>    rmse = 20.17
#>    mae = 14.079
#>    mape = 0.068862
#>    mse = 406.82
#>    rsq = 0.367
#>    rsqa = 0.3645
#> 
#> 

# WITH PRE-DEFINED TRAIN/TEST DATAFRAMES
splits <- msplit(dft, size = 0.8)
#> train_size  test_size 
#>        712        179 
splits$train$split <- "train"
splits$test$split <- "test"
df <- rbind(splits$train, splits$test)
r <- h2o_automl(df, "Survived", max_models = 1, train_test = "split")
#> 2021-12-15 19:43:30 | Started process...
#> - INDEPENDENT VARIABLE: Survived
#> - MODEL TYPE: Classification
#> # A tibble: 2 × 5
#>   tag       n     p order  pcum
#>   <lgl> <int> <dbl> <int> <dbl>
#> 1 FALSE   549  61.6     1  61.6
#> 2 TRUE    342  38.4     2 100  
#> - MISSINGS: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.
#> - CATEGORICALS: There are 4 non-numerical features. Consider using ohse() or equivalent prior to encode categorical variables.
#> 
#>  test train 
#>   179   712 
#> - REPEATED: There were 77 repeated rows which are being suppressed from the train dataset
#> - ALGORITHMS: excluded 'StackedEnsemble', 'DeepLearning'
#> - CACHE: Previous models are not being erased. You may use 'start_clean' [clear] or 'project_name' [join]
#> - UI: You may check results using H2O Flow's interactive platform: http://localhost:54321/flow/index.html
#> >>> Iterating until 1 models or 600 seconds...
#> - EUREKA: Succesfully generated 1 models
#>                             model_id       auc   logloss     aucpr
#> 1 XGBoost_1_AutoML_4_20211215_194330 0.8450812 0.4500589 0.8188965
#>   mean_per_class_error      rmse       mse
#> 1            0.2150513 0.3784227 0.1432037
#> SELECTED MODEL: XGBoost_1_AutoML_4_20211215_194330
#> - NOTE: The following variables were the least important: Embarked.S, SibSp, Embarked.C, Pclass.2
#> >>> Running predictions for Survived...
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Target value: FALSE
#> >>> Generating plots...
#> Model (1/1): XGBoost_1_AutoML_4_20211215_194330
#> Independent Variable: Survived
#> Type: Classification (2 classes)
#> Algorithm: XGBOOST
#> Split: 80% training data (of 891 observations)
#> Seed: 0
#> 
#> Test metrics:
#>    AUC = 0.87174
#>    ACC = 0.84358
#>    PRC = 0.89091
#>    TPR = 0.69014
#>    TNR = 0.94444
#> 
#> Most important variables:
#>    Sex.female (32.8%)
#>    Fare (21.8%)
#>    Age (11.7%)
#>    Pclass.3 (11.6%)
#>    Sex.male (9.2%)
#> Process duration: 7.56s
# }