Automated H2O's AutoML

This function lets the user create a robust and fast model, using H2O's AutoML function. The result is a list with the best model, its parameters, datasets, performance metrics, variables importance, and plots. Read more about the h2o_automl() pipeline here.

h2o_automl(
  df,
  y = "tag",
  ignore = NULL,
  train_test = NA,
  split = 0.7,
  weight = NULL,
  target = "auto",
  balance = FALSE,
  impute = FALSE,
  no_outliers = TRUE,
  unique_train = TRUE,
  center = FALSE,
  scale = FALSE,
  thresh = 10,
  seed = 0,
  nfolds = 5,
  max_models = 3,
  max_time = 10 * 60,
  start_clean = FALSE,
  exclude_algos = c("StackedEnsemble", "DeepLearning"),
  include_algos = NULL,
  plots = TRUE,
  alarm = TRUE,
  quiet = FALSE,
  print = TRUE,
  save = FALSE,
  subdir = NA,
  project = "ML Project",
  ...
)

# S3 method for h2o_automl
plot(x, ...)

# S3 method for h2o_automl
print(x, importance = TRUE, ...)

Arguments

df
Dataframe. Dataframe containing all your data, including the independent variable labeled as 'tag'. If you want to define which variable should be used instead, use the y parameter.

y
Variable or Character. Name of the independent variable.

ignore
Character vector. Force columns for the model to ignore

train_test
Character. If needed, df's column name with 'test' and 'train' values to split

split
Numeric. Value between 0 and 1 to split as train/test datasets. Value is for training set. Set value to 1 to train with all available data and test with same data (cross-validation will still be used when training). If train_test is set, value will be overwritten with its real split rate.

weight
Column with observation weights. Giving some observation a weight of zero is equivalent to excluding it from the dataset; giving an observation a relative weight of 2 is equivalent to repeating that row twice. Negative weights are not allowed.

target
Value. Which is your target positive value? If set to 'auto', the target with largest mean(score) will be selected. Change the value to overwrite. Only used when binary categorical model.

balance
Boolean. Auto-balance train dataset with under-sampling?

impute
Boolean. Fill NA values with MICE?

no_outliers
Boolean/Numeric. Remove y's outliers from the dataset? Will remove those values that are farther than n standard deviations from the independent variable's mean (Z-score). Set to TRUE for default (3) or numeric to set a different multiplier.

unique_train
Boolean. Keep only unique row observations for training data?

center, scale
Boolean. Using the base function scale, do you wish to center and/or scale all numerical values?

thresh
Integer. Threshold for selecting binary or regression models: this number is the threshold of unique values we should have in 'tag' (more than: regression; less than: classification)

seed
Integer. Set a seed for reproducibility. AutoML can only guarantee reproducibility if max_models is used because max_time is resource limited.

nfolds
Number of folds for k-fold cross-validation. Defaults to 5. Use 0 to disable cross-validation; this will also disable Stacked Ensemble (thus decreasing the overall model performance).

max_models, max_time
Numeric. Max number of models and seconds you wish for the function to iterate. Note that max_models guarantees reproducibility and max_time not (because it depends entirely on your machine's computational characteristics)

start_clean
Boolean. Erase everything in the current h2o instance before we start to train models? You may want to keep other models or not. To group results into a custom common AutoML project, you may use project_name argument.

exclude_algos, include_algos
Vector of character strings. Algorithms to skip or include during the model-building phase. Set NULL to ignore. When both are defined, only include_algos will be valid.

plots
Boolean. Create plots objects?

alarm
Boolean. Ping (sound) when done. Requires beepr.

quiet
Boolean. Quiet all messages, warnings, recommendations?

print
Boolean. Print summary when process ends?

save
Boolean. Do you wish to save/export results into your working directory?

subdir
Character. In which directory do you wish to save the results? Working directory as default.

project
Character. Your project's name

...
Additional parameters on h2o::h2o.automl

x
h2o_automl object

importance
Boolean. Print important variables?

Value

List. Trained model, predicted scores and datasets used, performance metrics, parameters, importance data.frame, seed, and plots when plots=TRUE.

List of algorithms

-> Read more here

DRF

Distributed Random Forest, including Random Forest (RF) and Extremely-Randomized Trees (XRT)

GLM

Generalized Linear Model

XGBoost

eXtreme Grading Boosting

GBM

Gradient Boosting Machine

DeepLearning

Fully-connected multi-layer artificial neural network

StackedEnsemble

Stacked Ensemble

Methods

print

Use print method to print models stats and summary

plot

Use plot method to plot results using mplot_full()

See also

Examples

 # CRAN
data(dft) # Titanic dataset
dft <- subset(dft, select = -c(Ticket, PassengerId, Cabin))

# Classification: Binomial - 2 Classes
r <- h2o_automl(dft, y = Survived, max_models = 1, impute = FALSE, target = "TRUE")

#> 2021-06-25 14:45:44 | Started process...

#> - INDEPENDENT VARIABLE: Survived

#> - MODEL TYPE: Classification

#> # A tibble: 2 x 5
#>   tag       n     p order  pcum
#>   <lgl> <int> <dbl> <int> <dbl>
#> 1 FALSE   549  61.6     1  61.6
#> 2 TRUE    342  38.4     2 100  

#> - MISSINGS: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - CATEGORICALS: There are 3 non-numerical features. Consider using ohse() or equivalent prior to encode categorical variables.

#> >>> Splitting data: train = 0.7 & test = 0.3

#> train_size  test_size 
#>        623        268 

#> - REPEATED: There were 50 repeated rows which are being suppressed from the train dataset

#> - ALGORITHMS: excluded 'StackedEnsemble', 'DeepLearning'

#> - CACHE: Previous models are not being erased. You may use 'start_clean' [clear] or 'project_name' [join]

#> - UI: You may check results using H2O Flow's interactive platform: http://localhost:54321/flow/index.html

#> >>> Iterating until 1 models or 600 seconds...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> - EUREKA: Succesfully generated 1 models

#>                           model_id      auc   logloss    aucpr
#> 1 XGBoost_1_AutoML_20210625_144554 0.839969 0.4703277 0.818098
#>   mean_per_class_error      rmse       mse
#> 1            0.2139996 0.3854216 0.1485498

#> SELECTED MODEL: XGBoost_1_AutoML_20210625_144554

#> - NOTE: The following variables were the least important: Pclass.2, Embarked.S, SibSp, Parch, Embarked.C

#> >>> Running predictions for Survived...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Target value: TRUE

#> >>> Generating plots...

#> Model (1/1): XGBoost_1_AutoML_20210625_144554
#> Independent Variable: Survived
#> Type: Classification (2 classes)
#> Algorithm: XGBOOST
#> Split: 70% training data (of 891 observations)
#> Seed: 0
#> 
#> Test metrics: 
#>    AUC = 0.86372
#>    ACC = 0.16418
#>    PRC = 0.14286
#>    TPR = 0.30337
#>    TNR = 0.094972
#> 
#> Most important variables:
#>    Sex.female (31.7%)
#>    Fare (23.5%)
#>    Age (21.3%)
#>    Pclass.3 (10.3%)
#>    Sex.male (4.0%)

#> Process duration: 21.7s

# Let's see all the stuff we have inside:
lapply(r, names)

#> $model
#> NULL
#> 
#> $y
#> NULL
#> 
#> $scores_test
#> [1] "tag"    "scores"
#> 
#> $metrics
#> [1] "dictionary"       "confusion_matrix" "gain_lift"        "metrics"         
#> [5] "cv_metrics"       "max_metrics"     
#> 
#> $parameters
#>  [1] "model_id"                          "training_frame"                   
#>  [3] "nfolds"                            "keep_cross_validation_models"     
#>  [5] "keep_cross_validation_predictions" "fold_assignment"                  
#>  [7] "stopping_tolerance"                "seed"                             
#>  [9] "distribution"                      "categorical_encoding"             
#> [11] "ntrees"                            "max_depth"                        
#> [13] "min_rows"                          "min_child_weight"                 
#> [15] "sample_rate"                       "subsample"                        
#> [17] "col_sample_rate"                   "colsample_bylevel"                
#> [19] "col_sample_rate_per_tree"          "colsample_bytree"                 
#> [21] "score_tree_interval"               "tree_method"                      
#> [23] "dmatrix_type"                      "backend"                          
#> [25] "x"                                 "y"                                
#> 
#> $importance
#> [1] "variable"            "relative_importance" "scaled_importance"  
#> [4] "importance"         
#> 
#> $datasets
#> [1] "global" "test"  
#> 
#> $scoring_history
#> [1] "timestamp"                     "duration"                     
#> [3] "number_of_trees"               "training_rmse"                
#> [5] "training_logloss"              "training_auc"                 
#> [7] "training_pr_auc"               "training_lift"                
#> [9] "training_classification_error"
#> 
#> $categoricals
#> [1] "Pclass"     "Sex"        "Embarked"   "train_test" "predict"   
#> 
#> $type
#> NULL
#> 
#> $split
#> NULL
#> 
#> $threshold
#> NULL
#> 
#> $model_name
#> NULL
#> 
#> $algorithm
#> NULL
#> 
#> $leaderboard
#> [1] "model_id"             "auc"                  "logloss"             
#> [4] "aucpr"                "mean_per_class_error" "rmse"                
#> [7] "mse"                 
#> 
#> $project
#> NULL
#> 
#> $seed
#> NULL
#> 
#> $h2o
#> NULL
#> 
#> $plots
#> [1] "dashboard"  "metrics"    "importance"
#> 

# Classification: Multi-Categorical - 3 Classes
r <- h2o_automl(dft, Pclass, ignore = c("Fare", "Cabin"), max_time = 30, plots = FALSE)

#> 2021-06-25 14:46:06 | Started process...

#> - INDEPENDENT VARIABLE: Pclass

#> - MODEL TYPE: Classification

#> # A tibble: 3 x 5
#>   tag       n     p order  pcum
#>   <fct> <int> <dbl> <int> <dbl>
#> 1 n_3     491  55.1     1  55.1
#> 2 n_1     216  24.2     2  79.4
#> 3 n_2     184  20.6     3 100  

#> - MISSINGS: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - CATEGORICALS: There are 3 non-numerical features. Consider using ohse() or equivalent prior to encode categorical variables.

#> >>> Splitting data: train = 0.7 & test = 0.3

#> train_size  test_size 
#>        623        268 

#> - REPEATED: There were 65 repeated rows which are being suppressed from the train dataset

#> - ALGORITHMS: excluded 'StackedEnsemble', 'DeepLearning'

#> - CACHE: Previous models are not being erased. You may use 'start_clean' [clear] or 'project_name' [join]

#> - UI: You may check results using H2O Flow's interactive platform: http://localhost:54321/flow/index.html

#> >>> Iterating until 3 models or 30 seconds...

#> - EUREKA: Succesfully generated 3 models

#>                           model_id mean_per_class_error   logloss      rmse
#> 1 XGBoost_2_AutoML_20210625_144606            0.4848238 0.8222259 0.5384101
#> 2 XGBoost_1_AutoML_20210625_144606            0.4861030 0.8451478 0.5422224
#> 3 XGBoost_3_AutoML_20210625_144606            0.4904451 0.8522329 0.5440560
#>         mse auc aucpr
#> 1 0.2898854 NaN   NaN
#> 2 0.2940051 NaN   NaN
#> 3 0.2959969 NaN   NaN

#> SELECTED MODEL: XGBoost_2_AutoML_20210625_144606

#> - NOTE: The following variables were the least important: Embarked.Q, Sex.male

#> >>> Running predictions for Pclass...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#> Model (1/3): XGBoost_2_AutoML_20210625_144606
#> Independent Variable: Pclass
#> Type: Classification (3 classes)
#> Algorithm: XGBOOST
#> Split: 70% training data (of 891 observations)
#> Seed: 0
#> 
#> Test metrics: 
#>    AUC = 0.7842
#>    ACC = 0.66791
#> 
#> Most important variables:
#>    Age (51.6%)
#>    Survived.FALSE (12.3%)
#>    Embarked.C (7.6%)
#>    Sex.female (5.5%)
#>    Survived.TRUE (5.2%)

#> Process duration: 11.8s

# Regression: Continuous Values
r <- h2o_automl(dft, y = "Fare", ignore = c("Pclass"), exclude_algos = NULL, quiet = TRUE)
print(r)

#> Model (1/4): StackedEnsemble_AllModels_AutoML_20210625_144618
#> Independent Variable: Fare
#> Type: Regression
#> Algorithm: STACKEDENSEMBLE
#> Split: 70% training data (of 871 observations)
#> Seed: 0
#> 
#> Test metrics: 
#>    rmse = 20.309
#>    mae = 14.244
#>    mape = 0.07304
#>    mse = 412.45
#>    rsq = 0.3169
#>    rsqa = 0.3143
#> 
#> 

# WITH PRE-DEFINED TRAIN/TEST DATAFRAMES
splits <- msplit(dft, size = 0.8)

#> train_size  test_size 
#>        712        179 

splits$train$split <- "train"
splits$test$split <- "test"
df <- rbind(splits$train, splits$test)
r <- h2o_automl(df, "Survived", max_models = 1, train_test = "split")

#> 2021-06-25 14:46:31 | Started process...

#> - INDEPENDENT VARIABLE: Survived

#> - MODEL TYPE: Classification

#> # A tibble: 2 x 5
#>   tag       n     p order  pcum
#>   <lgl> <int> <dbl> <int> <dbl>
#> 1 FALSE   549  61.6     1  61.6
#> 2 TRUE    342  38.4     2 100  

#> - MISSINGS: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - CATEGORICALS: There are 4 non-numerical features. Consider using ohse() or equivalent prior to encode categorical variables.

#> 
#>  test train 
#>   179   712 

#> - REPEATED: There were 77 repeated rows which are being suppressed from the train dataset

#> - ALGORITHMS: excluded 'StackedEnsemble', 'DeepLearning'

#> - CACHE: Previous models are not being erased. You may use 'start_clean' [clear] or 'project_name' [join]

#> - UI: You may check results using H2O Flow's interactive platform: http://localhost:54321/flow/index.html

#> >>> Iterating until 1 models or 600 seconds...

#> - EUREKA: Succesfully generated 1 models

#>                           model_id       auc   logloss     aucpr
#> 1 XGBoost_1_AutoML_20210625_144631 0.8595485 0.4359836 0.8300753
#>   mean_per_class_error      rmse       mse
#> 1            0.1990026 0.3695265 0.1365498

#> SELECTED MODEL: XGBoost_1_AutoML_20210625_144631

#> - NOTE: The following variables were the least important: Pclass.2, Embarked.S, Embarked.Q

#> >>> Running predictions for Survived...

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Target value: FALSE

#> >>> Generating plots...

#> Model (1/1): XGBoost_1_AutoML_20210625_144631
#> Independent Variable: Survived
#> Type: Classification (2 classes)
#> Algorithm: XGBOOST
#> Split: 80% training data (of 891 observations)
#> Seed: 0
#> 
#> Test metrics: 
#>    AUC = 0.8838
#>    ACC = 0.86034
#>    PRC = 0.89655
#>    TPR = 0.73239
#>    TNR = 0.94444
#> 
#> Most important variables:
#>    Sex.female (26.5%)
#>    Fare (23.7%)
#>    Age (18.2%)
#>    Pclass.3 (7.8%)
#>    Sex.male (7.3%)

#> Process duration: 8.21s