Automated H2O's AutoML Results

This is an auxiliary function to calculate predictions and results when using the h2o_automl() function.

h2o_results(
  h2o_object,
  test,
  train,
  y = "tag",
  which = 1,
  model_type,
  target = "auto",
  split = 0.7,
  ignore = c(),
  quiet = FALSE,
  project = "ML Project",
  seed = 0,
  leaderboard = list(),
  plots = TRUE,
  ...
)

Arguments

h2o_object
H2O Leaderboard (H2OFrame/H2OAutoML) or Model (h2o)

test, train
Dataframe. Must have the same columns

y
Variable or Character. Name of the independent variable.

which
Integer. Which model to select from leaderboard

model_type
Character. Select "Classification" or "Regression"

target
Value. Which is your target positive value? If set to 'auto', the target with largest mean(score) will be selected. Change the value to overwrite. Only used when binary categorical model.

split
Numeric. Value between 0 and 1 to split as train/test datasets. Value is for training set. Set value to 1 to train will all available data and test with same data (cross-validation will still be used when training)

ignore
Character vector. Columns too ignore

quiet
Boolean. Quiet all messages, warnings, recommendations?

project
Character. Your project's name

seed
Integer. Set a seed for reproducibility. AutoML can only guarantee reproducibility if max_models is used because max_time is resource limited.

leaderboard
H2O's Leaderboard. Passed when using h2o_selectmodel as it contains plain model and no leader board.

plots
Boolean. Create plots objects?

...
Additional parameters on h2o::h2o.automl