Holidays in your Country

This function lets the user automatically scrap holiday dates from any country and year within +- 5 years. Thanks to timeanddate.com!

Usage

holidays(countries = "Venezuela", years = year(Sys.Date()))

Arguments

countries

Character or vector. For which country(ies) should the holidays be imported?

years

Character or vector. For which year(s) do you wish to import holiday dates?

Value

data.frame with holidays data for given countries and years.

See also

Examples

# \donttest{
holidays(countries = "Argentina")
#> >>> Extracting Argentina's holidays for 2023
#> # A tibble: 40 × 10
#>    holiday    holid…¹ holid…² natio…³ obser…⁴ bank  nonwork season hother county
#>    <date>     <chr>   <chr>   <lgl>   <lgl>   <lgl> <lgl>   <lgl>  <lgl>  <fct> 
#>  1 2023-02-20 Carniv… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE  Argen…
#>  2 2023-02-21 Carniv… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE  Argen…
#>  3 2023-03-20 March … Season  FALSE   FALSE   FALSE FALSE   TRUE   FALSE  Argen…
#>  4 2023-03-24 Memori… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE  Argen…
#>  5 2023-04-02 Day of… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE  Argen…
#>  6 2023-04-05 Passov… Observ… FALSE   TRUE    FALSE FALSE   FALSE  FALSE  Argen…
#>  7 2023-04-06 Maundy… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE  Argen…
#>  8 2023-04-06 First … Option… FALSE   FALSE   FALSE FALSE   FALSE  TRUE   Argen…
#>  9 2023-04-07 Good F… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE  Argen…
#> 10 2023-04-07 Second… Option… FALSE   FALSE   FALSE FALSE   FALSE  TRUE   Argen…
#> # … with 30 more rows, and abbreviated variable names ¹​holidayname,
#> #   ²​holidaytype, ³​national, ⁴​observance
holidays(countries = c("Argentina", "Venezuela"), years = c(2019, 2020))
#> >>> Extracting Argentina's holidays for 2019
#> >>> Extracting Argentina's holidays for 2020
#> >>> Extracting Venezuela's holidays for 2019
#> >>> Extracting Venezuela's holidays for 2020
#> # A tibble: 158 × 11
#>    holiday    holidayname    holid…¹ natio…² obser…³ bank  nonwork season hother
#>    <date>     <chr>          <chr>   <lgl>   <lgl>   <lgl> <lgl>   <lgl>  <lgl> 
#>  1 2019-03-04 Carnival / Sh… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE 
#>  2 2019-03-05 Carnival / Sh… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE 
#>  3 2019-03-20 March Equinox  Season  FALSE   FALSE   FALSE FALSE   TRUE   FALSE 
#>  4 2019-03-24 Memorial Day   Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE 
#>  5 2019-04-02 Day of the Ve… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE 
#>  6 2019-04-18 Maundy Thursd… Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE 
#>  7 2019-04-19 Good Friday    Nation… TRUE    FALSE   FALSE FALSE   FALSE  FALSE 
#>  8 2019-04-19 Passover Eve   Observ… FALSE   TRUE    FALSE FALSE   FALSE  FALSE 
#>  9 2019-04-20 First day of … Option… FALSE   FALSE   FALSE FALSE   FALSE  TRUE  
#> 10 2019-04-21 Easter Sunday  Observ… FALSE   TRUE    FALSE FALSE   FALSE  FALSE 
#> # … with 148 more rows, 2 more variables: country <fct>, county <fct>, and
#> #   abbreviated variable names ¹​holidaytype, ²​national, ³​observance
# }