Holidays in your Country

This function lets the user automatically scrap holiday dates from any country and year within +- 5 years. Thanks to timeanddate.com!

holidays(countries = "Colombia", years = year(Sys.Date()))

Arguments

countries
Character or vector. For which country(ies) should the holidays be imported?

years
Character or vector. For which year(s) do you wish to import holiday dates?

Value

data.frame with holidays data for given countries and years.

See also

Examples

# \donttest{
holidays(countries = "Argentina") 

#> >>> Extracting Argentina's holidays for 2021

#> Warning: NAs introduced by coercion

#> # A tibble: 41 x 9
#>    holiday    holidayname  holidaytype  national observance bank  nonwork season
#>    <date>     <chr>        <chr>        <lgl>    <lgl>      <lgl> <lgl>   <lgl> 
#>  1 2021-02-15 Carnival / … National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  2 2021-02-16 Carnival / … National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  3 2021-03-20 March Equin… Season       FALSE    FALSE      FALSE FALSE   TRUE  
#>  4 2021-03-24 Memorial Day National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  5 2021-03-27 Passover Eve Observance   FALSE    TRUE       FALSE FALSE   FALSE 
#>  6 2021-03-28 First day o… Optional ho… FALSE    FALSE      FALSE FALSE   FALSE 
#>  7 2021-03-29 Second Day … Optional ho… FALSE    FALSE      FALSE FALSE   FALSE 
#>  8 2021-04-01 Maundy Thur… National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  9 2021-04-02 Good Friday  National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#> 10 2021-04-02 Day of the … National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#> # … with 31 more rows, and 1 more variable: hother <lgl>

holidays(countries = c("Argentina", "Venezuela"), years = c(2019, 2020))

#> >>> Extracting Argentina's holidays for 2019

#> Warning: NAs introduced by coercion

#> >>> Extracting Argentina's holidays for 2020

#> Warning: NAs introduced by coercion

#> >>> Extracting Venezuela's holidays for 2019

#> Warning: NAs introduced by coercion

#> >>> Extracting Venezuela's holidays for 2020

#> Warning: NAs introduced by coercion

#> # A tibble: 158 x 10
#>    holiday    holidayname  holidaytype  national observance bank  nonwork season
#>    <date>     <chr>        <chr>        <lgl>    <lgl>      <lgl> <lgl>   <lgl> 
#>  1 2019-03-04 Carnival / … National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  2 2019-03-05 Carnival / … National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  3 2019-03-20 March Equin… Season       FALSE    FALSE      FALSE FALSE   TRUE  
#>  4 2019-03-24 Memorial Day National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  5 2019-04-02 Day of the … National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  6 2019-04-18 Maundy Thur… National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  7 2019-04-19 Good Friday  National ho… TRUE     FALSE      FALSE FALSE   FALSE 
#>  8 2019-04-19 Passover Eve Observance   FALSE    TRUE       FALSE FALSE   FALSE 
#>  9 2019-04-20 First day o… Optional ho… FALSE    FALSE      FALSE FALSE   FALSE 
#> 10 2019-04-21 Easter Sund… Observance,… FALSE    TRUE       FALSE FALSE   FALSE 
#> # … with 148 more rows, and 2 more variables: hother <lgl>, country <fct>

# }