Holidays in your Country

This function lets the user automatically scrap holiday dates from any country and year within +- 5 years. Thanks to timeanddate.com!

Usage

holidays(
  countries = "Venezuela",
  years = year(Sys.Date()),
  quiet = FALSE,
  include_regions = FALSE
)

Arguments

countries

Character or vector. For which country(ies) should the holidays be imported?

years

Character or vector. For which year(s) do you wish to import holiday dates?

quiet

Boolean. Keep quiet? If not, informative messages will be shown.

include_regions

Boolean. Default FALSE. If TRUE, for countries with internal subdivisions, it will provide details on which sub-state the found holidays apply.

Value

data.frame with holidays data for given countries and years.

See also

Examples

# \donttest{
holidays(countries = "Argentina")
#> >>> Extracting Argentina's holidays for 2025
#> Warning: All formats failed to parse. No formats found.
#> # A tibble: 0 × 10
#> # ℹ 10 variables: holiday <date>, holiday_name <chr>, holiday_type <chr>,
#> #   national <lgl>, observance <lgl>, bank <lgl>, nonwork <lgl>, season <lgl>,
#> #   hother <lgl>, county <fct>
year <- as.integer(format(Sys.Date(), format = "%Y"))
holidays(countries = c("Spain", "Venezuela"), years = year)
#> >>> Extracting Spain's holidays for 2025
#> Warning: All formats failed to parse. No formats found.
#> >>> Extracting Venezuela's holidays for 2025
#> Warning: All formats failed to parse. No formats found.
#> # A tibble: 0 × 11
#> # ℹ 11 variables: holiday <date>, holiday_name <chr>, holiday_type <chr>,
#> #   national <lgl>, observance <lgl>, bank <lgl>, nonwork <lgl>, season <lgl>,
#> #   hother <lgl>, country <fct>, county <fct>
holidays(countries = "Germany", include_regions = TRUE)
#> >>> Extracting Germany's holidays for 2025
#> Warning: All formats failed to parse. No formats found.
#> # A tibble: 0 × 11
#> # ℹ 11 variables: holiday <date>, holiday_name <chr>, holiday_type <chr>,
#> #   holiday_details <chr>, national <lgl>, observance <lgl>, bank <lgl>,
#> #   nonwork <lgl>, season <lgl>, hother <lgl>, county <fct>
# }