Scrap data based on IP address

This function lets the user scrap https://db-ip.com/ given IP address(es) to get their associated address type, ASN, ISP, organization, country, state or region, county, city, ZIP postal code, weather station, coordinates, Timezone, local time, languages, and currency.

ip_data(ip = myip(), quiet = FALSE)

Arguments

ip

Vector. Vector with all IP's we wish to search.

quiet

Boolean. Do not show the loading statusbar?

Value

data.frame. Each row is an unique ip address, and columns will bee created for all the additional information found.

See also

Examples

# \donttest{
ip_data("163.114.132.0")
#>              id addresstype                   asn          isp  organization
#> 1 163.114.132.0        IPv4 54115 - FACEBOOK-CORP Facebook Inc Facebook Corp
#>         country stateregion districtcounty    city zippostalcode
#> 1 United States  Washington           King Seattle         98127
#>       weatherstation       coordinates                    timezone localtime
#> 1 USWA0395 - Seattle 47.6062, -122.332 America/Los_Angeles (UTC-8)          
#>               languages     currency
#> 1 en-US, es-US, haw, fr Dollar (USD)
ip_data(ip = c(myip(), "201.244.197.199"), quiet = TRUE)
#>                id addresstype                                asn
#> 1    13.105.49.78        IPv4 8075 - MICROSOFT-CORP-MSN-AS-BLOCK
#> 2 201.244.197.199        IPv4             19429 - ETB - Colombia
#>                     isp connection       country stateregion districtcounty
#> 1 Microsoft Corporation    Hosting United States  Washington          Grant
#> 2        ETB - Colombia       <NA>      Colombia Bogota D.C.    Bogotá D.C.
#>                        city zippostalcode        weatherstation
#> 1                Moses Lake         98837 USWA0285 - Moses Lake
#> 2 Bogotá (Comuna Chapinero)          <NA>  COXX6242 - Chapinero
#>         coordinates                    timezone localtime             languages
#> 1 47.1301, -119.278 America/Los_Angeles (UTC-8)           en-US, es-US, haw, fr
#> 2 4.64225, -74.0622      America/Bogota (UTC-5)                           es-CO
#>       currency                                   hostname
#> 1 Dollar (USD)                                       <NA>
#> 2   Peso (COP) dynamic-201-244-197-199.dynamic.etb.net.co
# }