Scrap data based on IP address

This function lets the user scrap https://db-ip.com/ given IP address(es) to get their associated address type, ASN, ISP, organization, country, state or region, county, city, ZIP postal code, weather station, coordinates, Timezone, local time, languages, and currency.

ip_data(ip = myip(), quiet = FALSE)

Arguments

ip
Vector. Vector with all IP's we wish to search.

quiet
Boolean. Do not show the loading statusbar?

Value

data.frame. Each row is an unique ip address, and columns will bee created for all the additional information found.

See also

Examples

# \donttest{
ip_data("163.114.132.0")

#>              id addresstype                   asn          isp  organization
#> 1 163.114.132.0        IPv4 54115 - FACEBOOK-CORP Facebook Inc Facebook Corp
#>         country stateregion districtcounty      city zippostalcode
#> 1 United States  California    Santa Clara Sunnyvale         94089
#>         weatherstation       coordinates                    timezone localtime
#> 1 USCA1116 - Sunnyvale 37.3688, -122.036 America/Los_Angeles (UTC-7)          
#>               languages     currency
#> 1 en-US, es-US, haw, fr Dollar (USD)

ip_data(ip = c(myip(), "201.244.197.199"), quiet = TRUE)

#>                id addresstype                                asn
#> 1    13.105.49.11        IPv4 8075 - MICROSOFT-CORP-MSN-AS-BLOCK
#> 2 201.244.197.199        IPv4             19429 - ETB - Colombia
#>                     isp connection       country stateregion districtcounty
#> 1 Microsoft Corporation    Hosting United States  Washington          Grant
#> 2        ETB - Colombia       <NA>      Colombia Bogota D.C.    Bogotá D.C.
#>              city zippostalcode            weatherstation       coordinates
#> 1      Moses Lake         98837     USWA0285 - Moses Lake 47.1301, -119.278
#> 2 Barrio San Luis          <NA> COXX7742 - Vereda El Hato 4.66779, -74.0215
#>                      timezone localtime             languages     currency
#> 1 America/Los_Angeles (UTC-7)           en-US, es-US, haw, fr Dollar (USD)
#> 2      America/Bogota (UTC-5)                           es-CO   Peso (COP)
#>                                     hostname
#> 1                                       <NA>
#> 2 dynamic-201-244-197-199.dynamic.etb.net.co

# }