Scrap data based on IP address

This function lets the user scrap https://db-ip.com/ given IP address(es) to get their associated address type, ASN, ISP, organization, country, state or region, county, city, ZIP postal code, weather station, coordinates, Timezone, local time, languages, and currency.

Usage

ip_data(ip = myip(), quiet = FALSE)

Arguments

ip

Vector. Vector with all IP's we wish to search.

quiet

Boolean. Do not show the loading statusbar?

Value

data.frame. Each row is an unique ip address, and columns will bee created for all the additional information found.

See also

Examples

# \donttest{
ip_data("163.114.132.0")
#>              id addresstype                   asn          isp connection
#> 1 163.114.132.0        IPv4 54115 - FACEBOOK-CORP Facebook Inc    Hosting
#>    organization       country stateregion      city zippostalcode
#> 1 Facebook Corp United States  California San Mateo         94497
#>         weatherstation      coordinates                    timezone localtime
#> 1 USCA1005 - San Mateo 37.563, -122.326 America/Los_Angeles (UTC-7)          
#>               languages     currency
#> 1 en-US, es-US, haw, fr Dollar (USD)
ip_data(ip = c(myip(), "201.244.197.199"), quiet = TRUE)
#>                id addresstype             asn                   isp connection
#> 1     52.157.4.94        IPv4   8075 - AS8075 Microsoft Corporation    Hosting
#> 2 201.244.197.199        IPv4 19429 - AS19429        ETB - Colombia       <NA>
#>         country stateregion districtcounty     city zippostalcode
#> 1 United States  California    Santa Clara San Jose         95141
#> 2      Colombia Bogota D.C.    Bogotá D.C.   Bogotá          <NA>
#>            weatherstation       coordinates                    timezone
#> 1     USCA0993 - San Jose 37.3382, -121.886 America/Los_Angeles (UTC-7)
#> 2 COXX2421 - Los Martires 4.60971, -74.0817      America/Bogota (UTC-5)
#>   localtime             languages     currency
#> 1           en-US, es-US, haw, fr Dollar (USD)
#> 2                           es-CO   Peso (COP)
#>                                     hostname
#> 1                                       <NA>
#> 2 dynamic-201-244-197-199.dynamic.etb.net.co
# }