Find Most Relevant Features Using Lasso Regression

Use Lasso regression to identify the most relevant variables that can predict/identify another variable. You might want to compare with corr_var() results for further understanding. No need to standardize your data.

lasso_vars(df, variable, ignore = NA, nlambdas = 100, seed = 123, ...)

Arguments

df
Dataframe

variable
Variable.

ignore
Character vector. Variables to exclude from study

nlambdas
Integer. Number of lambdas to be used in a search

seed
Numeric

...
ohse parameters

See also

Examples

if (FALSE) {
data(dft)
m <- lasso_vars(dft, Survived, ignore = c("Ticket", "Cabin"))
}