List files in a directory

This function lets the user list all files on a given directory. It also lets filter files which contains a string.

Usage

listfiles(folder = getwd(), recursive = TRUE, regex = NA, images = FALSE)

Arguments

folder

Character. Directory which contains files

recursive

Boolean. Should the listing recurse into directories?

regex

Character. String to use for filtering files

images

Boolean. Bring only image files?

Value

data.frame with relevant data for each file on folder directory.

See also

Examples

# All files in current directory (without recursive files)
df <- listfiles(recursive = TRUE)
head(df, 3)
#>       filename size isdir mode               mtime               ctime
#> 1     ROC.html   19 FALSE  644 2022-02-15 18:24:38 2022-02-15 18:24:38
#> 2 Rplot001.png    0 FALSE  644 2022-02-15 18:27:21 2022-02-15 18:27:21
#> 3 Rplot002.png   10 FALSE  644 2022-02-15 18:25:41 2022-02-15 18:25:41
#>                 atime uid gid  uname grname
#> 1 2022-02-15 18:24:38 501  20 runner  staff
#> 2 2022-02-15 18:24:38 501  20 runner  staff
#> 3 2022-02-15 18:24:45 501  20 runner  staff

# All files in current directory (with recursive files)
df <- listfiles(recursive = TRUE)
tail(df, 3)
#>            filename size isdir mode               mtime               ctime
#> 137    li_auth.html    8 FALSE  644 2022-02-15 18:27:20 2022-02-15 18:27:20
#> 138 li_profile.html    8 FALSE  644 2022-02-15 18:27:20 2022-02-15 18:27:20
#> 139  list_cats.html   11 FALSE  644 2022-02-15 18:27:21 2022-02-15 18:27:21
#>                   atime uid gid  uname grname
#> 137 2022-02-15 18:27:20 501  20 runner  staff
#> 138 2022-02-15 18:27:20 501  20 runner  staff
#> 139 2022-02-15 18:27:21 501  20 runner  staff

# Check R files using regex
df <- listfiles(regex = "\\.R$")