List files in a directory

This function lets the user list all files on a given directory. It also lets filter files which contains a string.

listfiles(folder = getwd(), recursive = TRUE, regex = NA, images = FALSE)

Arguments

folder
Character. Directory which contains files

recursive
Boolean. Should the listing recurse into directories?

regex
Character. String to use for filtering files

images
Boolean. Bring only image files?

See also

Examples

# All files in current directory (without recursive files)
df <- listfiles(recursive = TRUE)
head(df, 3)

#>            filename size isdir mode               mtime               ctime
#> 1     autoline.html    8 FALSE  644 2020-06-08 22:02:53 2020-06-08 22:02:53
#> 2 balance_data.html   10 FALSE  644 2020-06-08 22:02:53 2020-06-08 22:02:53
#> 3    bindfiles.html    7 FALSE  644 2020-06-08 22:02:53 2020-06-08 22:02:53
#>                 atime uid gid  uname grname
#> 1 2020-06-08 22:03:08 501  20 runner  staff
#> 2 2020-06-08 22:03:09 501  20 runner  staff
#> 3 2020-06-08 22:03:09 501  20 runner  staff

# All files in current directory (with recursive files)
df <- listfiles(recursive = TRUE)
tail(df, 3)

#>           filename size isdir mode               mtime               ctime
#> 267 year_week.html    7 FALSE  644 2020-06-08 22:02:55 2020-06-08 22:02:55
#> 268    yt2mp3.html    6 FALSE  644 2020-06-08 22:02:55 2020-06-08 22:02:55
#> 269   zerovar.html    8 FALSE  644 2020-06-08 22:02:55 2020-06-08 22:02:55
#>                   atime uid gid  uname grname
#> 267 2020-06-08 22:02:55 501  20 runner  staff
#> 268 2020-06-08 22:02:55 501  20 runner  staff
#> 269 2020-06-08 22:02:55 501  20 runner  staff

# Check R files using regex
df <- listfiles(regex = "\\.R$")