Automate Data Preprocess for Modeling

Pre-process your data before training a model. This is the prior step on the h2o_automl() function's pipeline. Enabling for other use cases when wanting too use any other framework, library, or custom algorithm.

model_preprocess(
  df,
  y = "tag",
  ignore = c(),
  train_test = NA,
  split = 0.7,
  weight = NULL,
  target = "auto",
  balance = FALSE,
  impute = FALSE,
  no_outliers = TRUE,
  center = FALSE,
  scale = FALSE,
  thresh = 10,
  seed = 0,
  quiet = FALSE
)

Arguments

df
Dataframe. Dataframe containing all your data, including the independent variable labeled as 'tag'. If you want to define which variable should be used instead, use the y parameter.

y
Character. Column name for independent variable.

ignore
Character vector. Force columns for the model to ignore

train_test
Character. If needed, df's column name with 'test' and 'train' values to split

split
Numeric. Value between 0 and 1 to split as train/test datasets. Value is for training set. Set value to 1 to train will all available data and test with same data (cross-validation will still be used when training)

weight
Column with observation weights. Giving some observation a weight of zero is equivalent to excluding it from the dataset; giving an observation a relative weight of 2 is equivalent to repeating that row twice. Negative weights are not allowed.

target
Value. Which is your target positive value? If set to 'auto', the target with largest mean(score) will be selected. Change the value to overwrite. Only used when binary categorical model.

balance
Boolean. Auto-balance train dataset with under-sampling?

impute
Boolean. Fill NA values with MICE?

no_outliers
Boolean/Numeric. Remove y's outliers from the dataset? Will remove those values that are farther than n standard deviations from the independent variable's mean (Z-score). Set to TRUE for default (3) or numeric to set a different multiplier.

center
Boolean. Using the base function scale, do you wish to center and/or scale all numerical values?

scale
Boolean. Using the base function scale, do you wish to center and/or scale all numerical values?

thresh
Integer. Threshold for selecting binary or regression models: this number is the threshold of unique values we should have in 'tag' (more than: regression; less than: classification)

seed
Integer. Set a seed for reproducibility. AutoML can only guarantee reproducibility if max_models is used because max_time is resource limited.

quiet
Boolean. Quiet all messages, warnings, recommendations?

See also

Examples

data(dft) # Titanic dataset

model_preprocess(dft, "Survived", balance = TRUE)

#> - INDEPENDENT VARIABLE: Survived

#> - MODEL TYPE: Classifier

#> Error in if (is.vector) df <- data.frame(values = df): argument is not interpretable as logical

model_preprocess(dft, "Fare", split = 0.5, scale = TRUE)

#> - INDEPENDENT VARIABLE: Fare

#> - MODEL TYPE: Regression

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    0.00    7.91   14.45   32.20   31.00  512.33 

#> - MISSINGS: The following variables contain missing observations: Age (19.87%). Consider using the impute parameter.

#> - CATEGORICALS: There are 6 non-numerical features. Consider using ohse() or equivalent prior to encode categorical variables.

#> - TRANSFORMATIONS: All numerical features (4) were scaled

#> >>> Splitting datasets...

#> train_size  test_size 
#>        435        436 

model_preprocess(dft, "Pclass", ignore = c("Fare", "Cabin"))

#> - INDEPENDENT VARIABLE: Pclass

#> - MODEL TYPE: Classifier

#> Error in if (is.vector) df <- data.frame(values = df): argument is not interpretable as logical

model_preprocess(dft, "Pclass", quiet = TRUE)

#> - MODEL TYPE: Classifier