Split a dataframe for training and testing sets

This function splits automatically a dataframe into train and test datasets. You can define a seed to get the same results every time, but has a default value. You can prevent it from printing the split counter result.

msplit(df, size = 0.7, seed = 0, print = T)

Arguments

df
Dataframe to split

size
Numeric. Split rate value, between 0 and 1. If set to 1, the train and test set will be the same.

seed
Seed for random split

print
Print summary results

Value

A list with both datasets, summary, and split rate

See also