One Hot Smart Encoding (Dummy Variables)

This function lets the user automatically transform a dataframe with categorical columns into numerical by one hot encoding technic.

ohse(
  df,
  redundant = FALSE,
  drops = TRUE,
  ignore = NA,
  dates = FALSE,
  holidays = FALSE,
  country = "Colombia",
  currency_pair = NA,
  trim = 0,
  limit = 10,
  variance = 0.9,
  other_label = "OTHER",
  sep = "_",
  summary = TRUE
)

Arguments

df
Dataframe

redundant
Boolean. Should we keep redundat columns? i.e. If the column only has two different values, should we keep both new columns?

drops
Boolean. Drop automatically some useless features?

ignore
Vector or character. Which column should be ignored?

dates
Boolean. Do you want the function to create more features out of the date/time columns?

holidays
Boolean. Include holidays as new columns?

country
Character or vector. For which countries should the holidays be included?

currency_pair
Character. Which currency exchange do you wish to get the history from? i.e, USD/COP, EUR/USD...

trim
Integer. Trim names until the nth character

limit
Integer. Limit one hot encoding to the n most frequent values of each column

variance
Numeric. Drop columns with more than n variance. Range: 0-1. For example: if a variable contains 91 unique different values out of 100 observations, this column will be suppressed if value is set to 0.9

other_label
Character. With which text do you wish to replace the filtered values with?

sep
Character. Separator's string

summary
Boolean. Print a summary of the operations?

See also

Examples

# Example with Dataset 1
data(dft)
dft <- dft[,c(2,3,5,9,11)]

ohse(dft, limit = 3) %>% head(3)

#> >>> One Hot Encoding applied to 3 variables: 'Pclass', 'Embarked', 'Survived'

#>   Survived_TRUE Age    Fare Pclass_1 Pclass_2 Embarked_C Embarked_OTHER
#> 1             0  22  7.2500        0        0          0              0
#> 2             1  38 71.2833        1        0          1              0
#> 3             1  26  7.9250        0        0          0              0
#>   Embarked_Q
#> 1          0
#> 2          0
#> 3          0

ohse(dft, limit = 3, redundant = TRUE) %>% head(3)

#> >>> One Hot Encoding applied to 3 variables: 'Survived', 'Pclass', 'Embarked'

#>   Age    Fare Survived_FALSE Survived_TRUE Pclass_1 Pclass_2 Pclass_3
#> 1  22  7.2500              1             0        0        0        1
#> 2  38 71.2833              0             1        1        0        0
#> 3  26  7.9250              0             1        0        0        1
#>   Embarked_C Embarked_OTHER Embarked_Q Embarked_S
#> 1          0              0          0          1
#> 2          1              0          0          0
#> 3          0              0          0          1

# Getting rid of columns with no (or too much) variance
dft$no_variance1 <- 0
dft$no_variance2 <- c("A", rep("B", nrow(dft) - 1))
dft$no_variance3 <- as.character(rnorm(nrow(dft)))
dft$no_variance4 <- c(rep("A", 20), rnorm(nrow(dft) - 20))
ohse(dft, limit = 3) %>% head(3)

#> >>> One Hot Encoding applied to 4 variables: 'Pclass', 'Embarked', 'Survived', 'no_variance2'

#> >>> Automatically dropped 3 columns with 0% or >=90% variance: 'no_variance1', 'no_variance3', 'no_variance4'

#>   Survived_TRUE Age    Fare no_variance2_B Pclass_1 Pclass_2 Embarked_C
#> 1             0  22  7.2500              0        0        0          0
#> 2             1  38 71.2833              1        1        0          1
#> 3             1  26  7.9250              1        0        0          0
#>   Embarked_OTHER Embarked_Q
#> 1              0          0
#> 2              0          0
#> 3              0          0

ohse(dft, limit = 3, var = 1) %>% head(3)

#> >>> One Hot Encoding applied to 5 variables: 'Pclass', 'Embarked', 'no_variance4', 'Survived', 'no_variance2'

#> >>> Automatically dropped 2 columns with 0% or >=100% variance: 'no_variance1', 'no_variance3'

#>   Survived_TRUE Age    Fare no_variance2_B Pclass_1 Pclass_2 Embarked_C
#> 1             0  22  7.2500              0        0        0          0
#> 2             1  38 71.2833              1        1        0          1
#> 3             1  26  7.9250              1        0        0          0
#>   Embarked_OTHER Embarked_Q no_variance4_.0.000384648716023502
#> 1              0          0                                  0
#> 2              0          0                                  0
#> 3              0          0                                  0
#>   no_variance4_.0.00358030839978085 no_variance4_A
#> 1                                 0              1
#> 2                                 0              1
#> 3                                 0              1

# Examples with Dataset 2
data(dfl)
dfl$novar <- "novar"
ohse(dfl, dates = TRUE) %>% head(3)

#> >>> Processing 1 date/time columns: 'opp_date'

#> >>> One Hot Encoding applied to 5 variables: 'kind', 'opp_date_weekday', 'type', 'issued', 'opp_date_weekend'

#> >>> Automatically dropped 1 columns with 0% or >=90% variance: 'novar'

#>   opp_id            opp_date type_TRUE values promo issued_TRUE opp_date_year
#> 1 992349 2019-02-06 12:22:48         1 480100     0           1          2019
#> 2 999436 2019-02-11 20:50:40         0 421750 50000           1          2019
#> 3 991823 2019-02-05 21:47:31         1 480100 12100           0          2019
#>   opp_date_month opp_date_day opp_date_week opp_date_weekend_TRUE
#> 1              2            6             6                     0
#> 2              2           11             6                     0
#> 3              2            5             6                     0
#>   opp_date_year_day opp_date_hour opp_date_minute opp_date_minutes
#> 1                36            12              22              742
#> 2                41            20              50             1250
#> 3                35            21              47             1307
#>   opp_date_second kind_NAs kind_red opp_date_weekday_Fri opp_date_weekday_Mon
#> 1              48        0        1                    0                    0
#> 2              40        0        0                    0                    1
#> 3              31        0        0                    0                    0
#>   opp_date_weekday_Sat opp_date_weekday_Sun opp_date_weekday_Thu
#> 1                    0                    0                    0
#> 2                    0                    0                    0
#> 3                    0                    0                    0
#>   opp_date_weekday_Tue
#> 1                    0
#> 2                    0
#> 3                    1

ohse(dfl, dates = TRUE, ignore = "issued") %>% head(3)

#> >>> Processing 1 date/time columns: 'opp_date'

#> >>> Omitting transformations for 'issued'

#> >>> One Hot Encoding applied to 4 variables: 'kind', 'opp_date_weekday', 'type', 'opp_date_weekend'

#> >>> Automatically dropped 1 columns with 0% or >=90% variance: 'novar'

#>   opp_id            opp_date values promo type_TRUE opp_date_year
#> 1 992349 2019-02-06 12:22:48 480100     0         1          2019
#> 2 999436 2019-02-11 20:50:40 421750 50000         0          2019
#> 3 991823 2019-02-05 21:47:31 480100 12100         1          2019
#>   opp_date_month opp_date_day opp_date_week opp_date_weekend_TRUE
#> 1              2            6             6                     0
#> 2              2           11             6                     0
#> 3              2            5             6                     0
#>   opp_date_year_day opp_date_hour opp_date_minute opp_date_minutes
#> 1                36            12              22              742
#> 2                41            20              50             1250
#> 3                35            21              47             1307
#>   opp_date_second kind_NAs kind_red opp_date_weekday_Fri opp_date_weekday_Mon
#> 1              48        0        1                    0                    0
#> 2              40        0        0                    0                    1
#> 3              31        0        0                    0                    0
#>   opp_date_weekday_Sat opp_date_weekday_Sun opp_date_weekday_Thu
#> 1                    0                    0                    0
#> 2                    0                    0                    0
#> 3                    0                    0                    0
#>   opp_date_weekday_Tue
#> 1                    0
#> 2                    0
#> 3                    1

if (FALSE) {
# Include holidays
ohse(dfl, dates = FALSE, holidays = TRUE, country = "Venezuela") %>% head()
}