SHAP-based dependence plots for categorical/numerical features (PDP)

Having a h2o_shap object, plot a dependence plot for any categorical or numerical feature.

shap_var(x, var, keep_outliers = FALSE)

Arguments

x
h2o_shap object

var
Variable name

keep_outliers
Boolean. Outliers detected with z-score and 3sd may be suppress or kept in your plot. Keep them?

Value

ggplot2 objct with shap values plotted

See also

Examples

# \donttest{
# Train a h2o_automl model
model <- h2o_automl(dft, Survived, max_models = 1, target = TRUE,
                    ignore = c("Ticket", "Cabin", "PassengerId"),
                    quiet = TRUE)

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> Target value: TRUE

# Calculate SHAP values
SHAP_values <- h2o_shap(model)
# Equivalent to:
# SHAP_values <- h2o_shap(
#  model = model$model,
#  test = model$datasets$test,
#  scores = model$scores_test$scores)

# Check SHAP results
head(SHAP_values)

#>   Embarked.  Embarked.C Embarked.Q  Embarked.S Embarked.missing(NA)   Pclass.1
#> 1         0 -0.08374244          0  0.03704862                    0 -0.1711508
#> 2         0  0.24667242          0  0.08512114                    0 -0.2609911
#> 3         0 -0.06764929          0 -0.04369200                    0  0.6880375
#> 4         0 -0.09540343          0 -0.07565676                    0 -0.2751508
#> 5         0  0.20894507          0  0.18138103                    0 -0.1446183
#> 6         0 -0.13432354          0 -0.01077519                    0 -0.2933404
#>      Pclass.2   Pclass.3 Pclass.missing(NA) Sex.female   Sex.male
#> 1 -0.12365140 -0.2116218                  0 -0.8004984 -0.2046820
#> 2  0.43403959  0.3921137                  0  1.3796868  0.4372064
#> 3 -0.03333040  0.4352387                  0  1.7077609  0.6711848
#> 4 -0.14012104 -0.5015697                  0  1.4525059  0.3728108
#> 5 -0.08761095 -0.3058314                  0  1.4785577  0.5023771
#> 6  0.20727664  0.3371496                  0 -0.9443169 -0.1525176
#>   Sex.missing(NA)         Age      SibSp       Parch       Fare   BiasTerm
#> 1               0  0.10703400 0.03008852 -0.04445532 -0.1122510 -0.5544208
#> 2               0  0.15334396 0.01802009 -0.00323644  0.3280579 -0.5544208
#> 3               0 -0.96233398 0.01622745  0.09054694  0.6818806 -0.5544208
#> 4               0  0.29412436 0.16871874  0.01267175  0.3348548 -0.5544208
#> 5               0 -0.04966021 0.02171060  0.05654531 -0.4308090 -0.5544208
#> 6               0 -0.15258318 0.05731664 -0.07630640 -0.2363376 -0.5544208

# You must have \code{ggbeeswarm} library to use this auxiliary function:
if (FALSE) {
# Plot SHAP values (feature importance)
plot(SHAP_values)

# Plot some of the variables (categorical)
shap_var(SHAP_values, Pclass)

# Plot some of the variables (numerical)
shap_var(SHAP_values, Fare)
}
# }