Download Stocks Historical Data

This function lets the user download stocks historical data

stocks_hist(
  symbols = c("VTI", "TSLA"),
  from = Sys.Date() - 365,
  to = Sys.Date(),
  today = TRUE,
  tax = 30,
  parg = FALSE,
  cache = TRUE,
  quiet = FALSE
)

Arguments

symbols
Character Vector. List of symbols to download historical data

from, to
Date. Dates for range. If not set, 1 year will be downloaded. Do use more than 4 days or will be over-written.

today
Boolean. Do you wish to add today's live quote? This will happen only if to value is the same as today's date

tax
Numeric. How much [0-99] of your dividends are gone with taxes?

parg
Boolean. Personal argument. Used to personalize stuff, in this case, taxes changed from A to B in given date (hard-coded)

cache
Boolean. Use daily cache if available?

quiet
Boolean. Keep quiet? If not, informative messages will be printed.

Value

data.frame for each Date and Symbol closing quote value.

See also

Examples

# \donttest{
stocks_hist(symbols = c("VTI", "FB"), from = Sys.Date() - 7, quiet = TRUE)

#> # A tibble: 8 x 3
#>   Date       Symbol Value
#>   <date>     <chr>  <dbl>
#> 1 2021-06-04 VTI     219.
#> 2 2021-06-04 FB      330.
#> 3 2021-06-03 VTI     218.
#> 4 2021-06-03 FB      326.
#> 5 2021-06-02 VTI     218.
#> 6 2021-06-02 FB      329.
#> 7 2021-06-01 VTI     218.
#> 8 2021-06-01 FB      329.

# }