Download Stocks Historical and Current Values

This function lets the user download stocks live data.

This function lets the user download stocks historical data.

Usage

stocks_quote(symbols, ...)

stocks_hist(
  symbols = c("VTI", "META"),
  from = Sys.Date() - 365,
  to = Sys.Date(),
  today = TRUE,
  tax = 15,
  parg = FALSE,
  cache = TRUE,
  quiet = FALSE,
  ...
)

# S3 method for stocks_hist
plot(x, type = 1, ...)

Arguments

symbols

Character Vector. List of symbols to download historical data.

...

Additional parameters

from, to

Date. Dates for range. If not set, 1 year will be downloaded. Do use more than 4 days or will be over-written.

today

Boolean. Do you wish to add today's live quote? This will happen only if to value is the same as today's date

tax

Numeric. How much [0-99] of your dividends are gone with taxes?

parg

Boolean. Personal argument. Used to personalize stuff, in this case, taxes changed from A to B in given date (hard-coded)

cache

Boolean. Use daily cache if available?

quiet

Boolean. Keep quiet? If not, informative messages will be printed.

x

stocks_hist object

type

Integer. Select type of plot.

Value

data.frame with Symbol, Type of stock, Quote time, current value, Daily Change, Market, and Symbol Name.

See also

Examples

# \donttest{
# Multiple quotes at the same time
stocks_quote(c("VTI", "VOO", "TSLA"))
#> # A tibble: 3 × 9
#>   Symbol QuoteTime           Value DailyChange DailyChangeP  Open  High   Low
#>   <chr>  <dttm>              <dbl>       <dbl>        <dbl> <dbl> <dbl> <dbl>
#> 1 VTI    2024-01-16 14:20:34  235.       -1.92       -0.810  236.  237.  235.
#> 2 VOO    2024-01-16 14:20:49  435.       -3.07       -0.701  437.  438.  435.
#> 3 TSLA   2024-01-16 14:20:52  220.        1.10        0.502  215.  223.  212.
#> # ℹ 1 more variable: Volume <int>
# }
if (FALSE) {
df <- stocks_hist(symbols = c("VTI", "META", "FIW"), from = Sys.Date() - 180)
print(head(df))
plot(df)
}