Keyword/Topic identification using RAKE

RAKE is a basic algorithm which tries to identify keywords in text. Based on udpipe library, model models, and keywords_rake function.

Usage

topics_rake(text, file = "english-ewt-ud-2.4-190531.udpipe", lang = "english")

Arguments

text

Character vector

file

Character. Name of udpipe model previously downloaded for a specific language

lang

Character. If file does not exist, this language will be downloaded from udpipe's models.

Value

data.frame with topics for each text input.

See also