Recursive Partitioning and Regression Trees

Fit and plot a rpart model for exploratory purposes using rpart and rpart.plot libraries. Idea from explore library.

tree_var(
  df,
  target,
  max = 3,
  min = 20,
  cp = 0,
  size = 0.7,
  ohse = TRUE,
  plot = TRUE,
  ...
)

Arguments

df
Data frame

target
Variable

max
Integer. Maximal depth of the tree

min
Integer. The minimum number of observations that must exist in a node in order for a split to be attempted

cp
Numeric. Complexity parameter

size
Numeric. Textsize of plot

ohse
Boolean. Auto generate One Hot Smart Encoding?

plot
Boolean. Return a plot? If not, rpart object

...
rpart.plot custom parameters

See also